@@ -6,7 +6,7 @@ import 'package:picPics/components/circular_menu_item.dart';
 
 class CircularMenu extends StatefulWidget {
   /// use global key to control animation anywhere in the code
-  final GlobalKey<CircularMenuState> key;
+  final GlobalKey<CircularMenuState>? key;
 
   /// list of CircularMenuItem contains at least two items.
   final List<CircularMenuItem> items;
@@ -18,7 +18,7 @@ class CircularMenu extends StatefulWidget {
   final double radius;
 
   /// widget holds actual page content
-  final Widget backgroundWidget;
+  final Widget? backgroundWidget;
 
   /// animation duration
   final Duration animationDuration;
@@ -30,13 +30,13 @@ class CircularMenu extends StatefulWidget {
   final Curve reverseCurve;
 
   /// callback
-  final VoidCallback toggleButtonOnPressed;
-  final Color toggleButtonColor;
+  final VoidCallback? toggleButtonOnPressed;
+  final Color? toggleButtonColor;
   final double toggleButtonSize;
   final List<BoxShadow> toggleButtonBoxShadow;
   final double toggleButtonPadding;
   final double toggleButtonMargin;
-  final Color toggleButtonIconColor;
+  final Color? toggleButtonIconColor;
 
   final bool useInHorizontal;
   final bool isExpanded;
@@ -46,7 +46,7 @@ class CircularMenu extends StatefulWidget {
   /// equal or greater than zero.
   /// [items] must not be null and it must contains two elements at least.
   CircularMenu({
-    @required this.items,
+    required this.items,
     this.alignment = Alignment.bottomCenter,
     this.radius = 100,
     this.backgroundWidget,
@@ -55,7 +55,7 @@ class CircularMenu extends StatefulWidget {
     this.reverseCurve = Curves.fastOutSlowIn,
     this.toggleButtonOnPressed,
     this.toggleButtonColor,
-    this.toggleButtonBoxShadow,
+    required this.toggleButtonBoxShadow,
     this.toggleButtonMargin = 10,
     this.toggleButtonPadding = 10,
     this.toggleButtonSize = 40,
@@ -63,7 +63,7 @@ class CircularMenu extends StatefulWidget {
     this.useInHorizontal = false,
     this.isExpanded = false,
     this.key,
-  })  : assert(items != null, 'items can not be empty list'),
+  })  : //assert(items != null, 'items can not be empty list'),
         assert(items.length > 1, 'if you have one item no need to use a Menu'),
         super(key: key);
 
@@ -73,8 +73,8 @@ class CircularMenu extends StatefulWidget {
 
 class CircularMenuState extends State<CircularMenu>
     with SingleTickerProviderStateMixin {
-  AnimationController _animationController;
-  Animation<double> _animation;
+  late AnimationController _animationController;
+  late Animation<double> _animation;
 
   /// forward animation
   void forwardAnimation() {
@@ -148,9 +148,7 @@ class CircularMenuState extends State<CircularMenu>
             _animationController.status == AnimationStatus.dismissed
                 ? (_animationController).forward()
                 : (_animationController).reverse();
-            if (widget.toggleButtonOnPressed != null) {
-              widget.toggleButtonOnPressed();
-            }
+            widget.toggleButtonOnPressed?.call();
           },
           boxShadow: widget.toggleButtonBoxShadow,
           animatedIcon: AnimatedIcon(
