@@ -9,33 +9,33 @@ enum CustomBubbleBottomBarFabLocation { end, center }
 // ignore: must_be_immutable
 class CustomBubbleBottomBar extends StatefulWidget {
   CustomBubbleBottomBar(
-      {Key key,
-      @required this.items,
+      {Key? key,
+      required this.items,
       this.onTap,
       this.currentIndex = 0,
-      @required this.opacity,
+      required this.opacity,
       this.iconSize = 24.0,
       this.borderRadius,
       this.elevation,
       this.backgroundColor,
       this.hasNotch = false,
       this.hasInk = false,
-      this.inkColor,
-      this.fabLocation})
-      : assert(items != null),
+      required this.inkColor,
+      required this.fabLocation})
+      : //assert(items != null),
         assert(items.length >= 2),
         assert(0 <= currentIndex && currentIndex < items.length),
-        assert(iconSize != null),
+        //assert(iconSize != null),
         super(key: key);
 
   final List<CustomBubbleBottomBarItem> items;
-  final ValueChanged<int> onTap;
+  final ValueChanged<int>? onTap;
   int currentIndex;
   final double iconSize;
   final double opacity;
-  final BorderRadius borderRadius;
-  final double elevation;
-  final Color backgroundColor;
+  final BorderRadius? borderRadius;
+  final double? elevation;
+  final Color? backgroundColor;
   final bool hasNotch;
   final bool hasInk;
   final CustomBubbleBottomBarFabLocation fabLocation;
@@ -47,28 +47,31 @@ class CustomBubbleBottomBar extends StatefulWidget {
 }
 
 class _CustomBottomNavigationTile extends StatelessWidget {
-  const _CustomBottomNavigationTile(
-      this.item, this.opacity, this.animation, this.iconSize,
-      {this.onTap,
-      this.colorTween,
-      this.flex,
-      this.selected = false,
-      this.indexLabel,
-      this.ink = false,
-      this.inkColor})
-      : assert(selected != null);
-
   final CustomBubbleBottomBarItem item;
   final Animation<double> animation;
   final double iconSize;
   final VoidCallback onTap;
-  final ColorTween colorTween;
+  final ColorTween? colorTween;
   final double flex;
   final bool selected;
   final String indexLabel;
   final double opacity;
   final bool ink;
-  final Color inkColor;
+  final Color? inkColor;
+
+  const _CustomBottomNavigationTile(
+    this.item,
+    this.opacity,
+    this.animation,
+    this.iconSize, {
+    required this.onTap,
+    this.colorTween,
+    required this.flex,
+    this.selected = false,
+    required this.indexLabel,
+    this.ink = false,
+    this.inkColor,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -139,20 +142,20 @@ class _CustomBottomNavigationTile extends StatelessWidget {
 }
 
 class _TileIcon extends StatelessWidget {
-  const _TileIcon({
-    Key key,
-    @required this.colorTween,
-    @required this.animation,
-    @required this.iconSize,
-    @required this.selected,
-    @required this.item,
-  }) : super(key: key);
-
-  final ColorTween colorTween;
+  final ColorTween? colorTween;
   final Animation<double> animation;
   final double iconSize;
   final bool selected;
   final CustomBubbleBottomBarItem item;
+
+  const _TileIcon({
+    Key? key,
+    this.colorTween,
+    required this.animation,
+    required this.iconSize,
+    required this.selected,
+    required this.item,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -175,16 +178,15 @@ class _TileIcon extends StatelessWidget {
 }
 
 class _Label extends StatelessWidget {
-  _Label({
-    Key key,
-    @required this.animation,
-    @required this.item,
-    @required this.color,
-  }) : super(key: key);
-
   final Animation<double> animation;
   final CustomBubbleBottomBarItem item;
   final Color color;
+  _Label({
+    Key? key,
+    required this.animation,
+    required this.item,
+    required this.color,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -203,13 +205,16 @@ class _Label extends StatelessWidget {
 
 class _CustomBottomNavigationBarState extends State<CustomBubbleBottomBar>
     with TickerProviderStateMixin {
-  List<AnimationController> _controllers = <AnimationController>[];
-  List<CurvedAnimation> _animations;
-  Color _backgroundColor;
-  ValueListenable<ScaffoldGeometry> geometryListenable;
+  /// list
+  var _controllers = <AnimationController>[];
+  var _animations = <CurvedAnimation>[];
+  Color? _backgroundColor;
   bool fabExists = false;
-  CustomBubbleBottomBar holder;
-  Animatable<double> _flexTween;
+
+  /// late
+  late ValueListenable<ScaffoldGeometry> geometryListenable;
+  late Animatable<double> _flexTween;
+  //CustomBubbleBottomBar holder;
 
   @override
   void didChangeDependencies() {
@@ -297,7 +302,7 @@ class _CustomBottomNavigationBarState extends State<CustomBubbleBottomBar>
   List<Widget> _createTiles() {
     final MaterialLocalizations localizations =
         MaterialLocalizations.of(context);
-    assert(localizations != null);
+
     final List<Widget> children = <Widget>[];
     for (int i = 0; i < widget.items.length; i += 1) {
       children.add(
@@ -307,7 +312,7 @@ class _CustomBottomNavigationBarState extends State<CustomBubbleBottomBar>
           _animations[i],
           widget.iconSize,
           onTap: () {
-            if (widget.onTap != null) widget.onTap(i);
+            /* if (widget.onTap != null) */ widget.onTap?.call(i);
           },
           flex: _evaluateFlex(_animations[i]),
           selected: i == widget.currentIndex,
@@ -369,10 +374,8 @@ class _CustomBottomNavigationBarState extends State<CustomBubbleBottomBar>
         explicitChildNodes: true,
         child: widget.hasNotch
             ? PhysicalShape(
-                elevation: widget.elevation != null ? widget.elevation : 8.0,
-                color: widget.backgroundColor != null
-                    ? widget.backgroundColor
-                    : Colors.white,
+                elevation: widget.elevation ?? 8.0,
+                color: widget.backgroundColor ?? Colors.white,
                 clipper: _CustomBubbleBottomBarClipper(
                   shape: CircularNotchedRectangle(),
                   geometry: geometryListenable,
@@ -381,26 +384,21 @@ class _CustomBottomNavigationBarState extends State<CustomBubbleBottomBar>
                 child: _inner(additionalBottomPadding),
               )
             : Material(
-                elevation: widget.elevation != null ? widget.elevation : 8.0,
-                color: widget.backgroundColor != null
-                    ? widget.backgroundColor
-                    : Colors.white,
+                elevation: widget.elevation ?? 8.0,
+                color: widget.backgroundColor ?? Colors.white,
                 child: _inner(additionalBottomPadding),
-                borderRadius: widget.borderRadius != null
-                    ? widget.borderRadius
-                    : BorderRadius.zero,
+                borderRadius: widget.borderRadius ?? BorderRadius.zero,
               ));
   }
 }
 
 class CustomBubbleBottomBarItem {
   const CustomBubbleBottomBarItem({
-    @required this.icon,
+    required this.icon,
 //    this.title,
-    Widget activeIcon,
-    this.backgroundColor,
-  })  : activeIcon = activeIcon ?? icon,
-        assert(icon != null);
+    Widget? activeIcon,
+    required this.backgroundColor,
+  }) : activeIcon = activeIcon ?? icon;
   final Widget icon;
   final Widget activeIcon;
 //  final Widget title;
@@ -408,24 +406,23 @@ class CustomBubbleBottomBarItem {
 }
 
 class _CustomBubbleBottomBarClipper extends CustomClipper<Path> {
-  const _CustomBubbleBottomBarClipper({
-    @required this.geometry,
-    @required this.shape,
-    @required this.notchMargin,
-  })  : assert(geometry != null),
-        assert(shape != null),
-        assert(notchMargin != null),
-        super(reclip: geometry);
-
   final ValueListenable<ScaffoldGeometry> geometry;
   final NotchedShape shape;
   final double notchMargin;
+  const _CustomBubbleBottomBarClipper({
+    required this.geometry,
+    required this.shape,
+    required this.notchMargin,
+  }) : /*  : assert(geometry != null),
+        assert(shape != null),
+        assert(notchMargin != null), */
+        super(reclip: geometry);
 
   @override
   Path getClip(Size size) {
-    final Rect button = geometry.value.floatingActionButtonArea?.translate(
+    final Rect? button = geometry.value.floatingActionButtonArea?.translate(
       0.0,
-      geometry.value.bottomNavigationBarTop * -1.0,
+      (geometry.value.bottomNavigationBarTop ?? .0) * -1.0,
     );
 
     return shape.getOuterPath(Offset.zero & size, button?.inflate(notchMargin));
