@@ -7,8 +7,8 @@ class MultiCircularMenu extends StatelessWidget {
   final List<CircularMenu> menus;
 
   /// widget holds actual page content
-  final Widget backgroundWidget;
-  const MultiCircularMenu({@required this.menus, this.backgroundWidget})
+  final Widget? backgroundWidget;
+  const MultiCircularMenu({required this.menus, this.backgroundWidget})
       : assert(menus.length != 0, 'menus can not be empty list'),
         assert(menus.length > 1,
             'no need to use MultiCircularMenu you can directly use CircularMenu');
