@@ -4,6 +4,6 @@ import '../entities/entities.dart';
 
 abstract class SecretKeyAuthentication {
   Future<SecretKeyEntity> auth({
-    @required String pin,
+    required String pin,
   });
 }
