@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -19,576 +19,670 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
-  static m0(email) => "An access key was sent to ${email}";
+  static String m0(email) => "An access key was sent to ${email}";
 
-  static m1(howMany) => "${Intl.plural(howMany, zero: 'No photos selected', one: '1 photo selected', other: '${howMany} photos selected')}";
+  static String m1(howMany) =>
+      "${Intl.plural(howMany, zero: 'No photos selected', one: '1 photo selected', other: '${howMany} photos selected')}";
 
-  static m2(number) => "You completed your ${number} free daily pics, do you want to continue?";
+  static String m2(number) =>
+      "You completed your ${number} free daily pics, do you want to continue?";
 
-  static m3(url) => "To organize all your photos go to ${url}";
+  static String m3(url) => "To organize all your photos go to ${url}";
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "access_code" : MessageLookupByLibrary.simpleMessage("Access code"),
-    "access_code_sent" : m0,
-    "add_multiple_tags" : MessageLookupByLibrary.simpleMessage("Adicione múltiplas tags"),
-    "add_tag" : MessageLookupByLibrary.simpleMessage("Add tag"),
-    "add_tags" : MessageLookupByLibrary.simpleMessage("Add tags"),
-    "aerospace_engineering_tag" : MessageLookupByLibrary.simpleMessage("Aerospace engineering"),
-    "aircraft_tag" : MessageLookupByLibrary.simpleMessage("Aircraft"),
-    "airliner_tag" : MessageLookupByLibrary.simpleMessage("Airliner"),
-    "airplane_tag" : MessageLookupByLibrary.simpleMessage("Airplane"),
-    "alcohol_tag" : MessageLookupByLibrary.simpleMessage("Alcohol"),
-    "all_at_once" : MessageLookupByLibrary.simpleMessage("Organize multiple photos at once"),
-    "all_photos_were_tagged" : MessageLookupByLibrary.simpleMessage("There are no more photos to organize."),
-    "all_search_tags" : MessageLookupByLibrary.simpleMessage("All Search Tags"),
-    "always" : MessageLookupByLibrary.simpleMessage("Always"),
-    "aquarium_tag" : MessageLookupByLibrary.simpleMessage("Aquarium"),
-    "archery_tag" : MessageLookupByLibrary.simpleMessage("Archery"),
-    "armrest_tag" : MessageLookupByLibrary.simpleMessage("Armrest"),
-    "army_tag" : MessageLookupByLibrary.simpleMessage("Army"),
-    "ask_photo_library_permission" : MessageLookupByLibrary.simpleMessage("We need access to your photo library so you can start organizing your photos with picPics. Don\'t worry your data will never leave your device!"),
-    "asphalt_tag" : MessageLookupByLibrary.simpleMessage("Asphalt"),
-    "aurora_tag" : MessageLookupByLibrary.simpleMessage("Aurora"),
-    "auto_renewable_first_part" : MessageLookupByLibrary.simpleMessage("The subscription is "),
-    "auto_renewable_second_part" : MessageLookupByLibrary.simpleMessage("auto-renewable."),
-    "aviation_tag" : MessageLookupByLibrary.simpleMessage("Aviation"),
-    "baby_tag" : MessageLookupByLibrary.simpleMessage("Baby"),
-    "backpacking_tag" : MessageLookupByLibrary.simpleMessage("Backpacking"),
-    "badminton_tag" : MessageLookupByLibrary.simpleMessage("Badminton"),
-    "bag_tag" : MessageLookupByLibrary.simpleMessage("Bag"),
-    "balance_tag" : MessageLookupByLibrary.simpleMessage("Balance"),
-    "balloon_tag" : MessageLookupByLibrary.simpleMessage("Balloon"),
-    "ballroom_tag" : MessageLookupByLibrary.simpleMessage("Ballroom"),
-    "bangle_tag" : MessageLookupByLibrary.simpleMessage("Bangle"),
-    "bangs_tag" : MessageLookupByLibrary.simpleMessage("Bangs"),
-    "bar_tag" : MessageLookupByLibrary.simpleMessage("Bar"),
-    "barn_tag" : MessageLookupByLibrary.simpleMessage("Barn"),
-    "basset_hound_tag" : MessageLookupByLibrary.simpleMessage("Basset hound"),
-    "bathing_tag" : MessageLookupByLibrary.simpleMessage("Bathing"),
-    "bathroom_tag" : MessageLookupByLibrary.simpleMessage("Bathroom"),
-    "beach_tag" : MessageLookupByLibrary.simpleMessage("Beach"),
-    "beanie_tag" : MessageLookupByLibrary.simpleMessage("Beanie"),
-    "bear_tag" : MessageLookupByLibrary.simpleMessage("Bear"),
-    "beard_tag" : MessageLookupByLibrary.simpleMessage("Beard"),
-    "bedroom_tag" : MessageLookupByLibrary.simpleMessage("Bedroom"),
-    "bench_tag" : MessageLookupByLibrary.simpleMessage("Bench"),
-    "bento_tag" : MessageLookupByLibrary.simpleMessage("Bento"),
-    "bicycle_tag" : MessageLookupByLibrary.simpleMessage("Bicycle"),
-    "bird_tag" : MessageLookupByLibrary.simpleMessage("Bird"),
-    "blackboard_tag" : MessageLookupByLibrary.simpleMessage("Blackboard"),
-    "blazer_tag" : MessageLookupByLibrary.simpleMessage("Blazer"),
-    "boat_tag" : MessageLookupByLibrary.simpleMessage("Boat"),
-    "bonfire_tag" : MessageLookupByLibrary.simpleMessage("Bonfire"),
-    "boxer_tag" : MessageLookupByLibrary.simpleMessage("Boxer"),
-    "bracelet_tag" : MessageLookupByLibrary.simpleMessage("Bracelet"),
-    "branch_tag" : MessageLookupByLibrary.simpleMessage("Branch"),
-    "bread_tag" : MessageLookupByLibrary.simpleMessage("Bread"),
-    "brick_tag" : MessageLookupByLibrary.simpleMessage("Brick"),
-    "bride_tag" : MessageLookupByLibrary.simpleMessage("Bride"),
-    "bridge_tag" : MessageLookupByLibrary.simpleMessage("Bridge"),
-    "bridle_tag" : MessageLookupByLibrary.simpleMessage("Bridle"),
-    "brig_tag" : MessageLookupByLibrary.simpleMessage("Brig"),
-    "building_tag" : MessageLookupByLibrary.simpleMessage("Building"),
-    "bull_tag" : MessageLookupByLibrary.simpleMessage("Bull"),
-    "bullfighting_tag" : MessageLookupByLibrary.simpleMessage("Bullfighting"),
-    "bumper_tag" : MessageLookupByLibrary.simpleMessage("Bumper"),
-    "bunk_bed_tag" : MessageLookupByLibrary.simpleMessage("Bunk bed"),
-    "bus_tag" : MessageLookupByLibrary.simpleMessage("Bus"),
-    "butterfly_tag" : MessageLookupByLibrary.simpleMessage("Butterfly"),
-    "cabinetry_tag" : MessageLookupByLibrary.simpleMessage("Cabinetry"),
-    "cage_tag" : MessageLookupByLibrary.simpleMessage("Cage"),
-    "cairn_terrier_tag" : MessageLookupByLibrary.simpleMessage("Cairn terrier"),
-    "cake_tag" : MessageLookupByLibrary.simpleMessage("Cake"),
-    "camping_tag" : MessageLookupByLibrary.simpleMessage("Camping"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Cancel"),
-    "cancel_anytime" : MessageLookupByLibrary.simpleMessage("Cancel anytime"),
-    "canoe_tag" : MessageLookupByLibrary.simpleMessage("Canoe"),
-    "canyon_tag" : MessageLookupByLibrary.simpleMessage("Canyon"),
-    "cap_tag" : MessageLookupByLibrary.simpleMessage("Cap"),
-    "cappuccino_tag" : MessageLookupByLibrary.simpleMessage("Cappuccino"),
-    "car_tag" : MessageLookupByLibrary.simpleMessage("Car"),
-    "carnival_tag" : MessageLookupByLibrary.simpleMessage("Carnival"),
-    "casino_tag" : MessageLookupByLibrary.simpleMessage("Casino"),
-    "castle_tag" : MessageLookupByLibrary.simpleMessage("Castle"),
-    "cat_tag" : MessageLookupByLibrary.simpleMessage("Cat"),
-    "cathedral_tag" : MessageLookupByLibrary.simpleMessage("Cathedral"),
-    "cattle_tag" : MessageLookupByLibrary.simpleMessage("Cattle"),
-    "cavalier_tag" : MessageLookupByLibrary.simpleMessage("Cavalier"),
-    "cave_tag" : MessageLookupByLibrary.simpleMessage("Cave"),
-    "caving_tag" : MessageLookupByLibrary.simpleMessage("Caving"),
-    "centrepiece_tag" : MessageLookupByLibrary.simpleMessage("Centrepiece"),
-    "chain_tag" : MessageLookupByLibrary.simpleMessage("Chain"),
-    "chair_tag" : MessageLookupByLibrary.simpleMessage("Chair"),
-    "cheeseburger_tag" : MessageLookupByLibrary.simpleMessage("Cheeseburger"),
-    "christmas_tag" : MessageLookupByLibrary.simpleMessage("Christmas"),
-    "church_tag" : MessageLookupByLibrary.simpleMessage("Church"),
-    "circus_tag" : MessageLookupByLibrary.simpleMessage("Circus"),
-    "class_tag" : MessageLookupByLibrary.simpleMessage("Class"),
-    "cliff_tag" : MessageLookupByLibrary.simpleMessage("Cliff"),
-    "clipper_tag" : MessageLookupByLibrary.simpleMessage("Clipper"),
-    "clock_tag" : MessageLookupByLibrary.simpleMessage("Clock"),
-    "close" : MessageLookupByLibrary.simpleMessage("Close"),
-    "clown_tag" : MessageLookupByLibrary.simpleMessage("Clown"),
-    "coffee_tag" : MessageLookupByLibrary.simpleMessage("Coffee"),
-    "cola_tag" : MessageLookupByLibrary.simpleMessage("Cola"),
-    "comet_tag" : MessageLookupByLibrary.simpleMessage("Comet"),
-    "comics_tag" : MessageLookupByLibrary.simpleMessage("Comics"),
-    "community_tag" : MessageLookupByLibrary.simpleMessage("Community"),
-    "competition_tag" : MessageLookupByLibrary.simpleMessage("Competition"),
-    "computer_tag" : MessageLookupByLibrary.simpleMessage("Computer"),
-    "concert_tag" : MessageLookupByLibrary.simpleMessage("Concert"),
-    "confirm_email" : MessageLookupByLibrary.simpleMessage("Confirm your registration email to receive your access code"),
-    "confirm_secret_key" : MessageLookupByLibrary.simpleMessage("Confirm secret key"),
-    "construction_tag" : MessageLookupByLibrary.simpleMessage("Construction"),
-    "continue_string" : MessageLookupByLibrary.simpleMessage("Continue"),
-    "cookie_tag" : MessageLookupByLibrary.simpleMessage("Cookie"),
-    "cookware_and_bakeware_tag" : MessageLookupByLibrary.simpleMessage("Cookware and bakeware"),
-    "cool_tag" : MessageLookupByLibrary.simpleMessage("Cool"),
-    "cotton_tag" : MessageLookupByLibrary.simpleMessage("Cotton"),
-    "couch_tag" : MessageLookupByLibrary.simpleMessage("Couch"),
-    "countertop_tag" : MessageLookupByLibrary.simpleMessage("Countertop"),
-    "country" : MessageLookupByLibrary.simpleMessage("country"),
-    "couscous_tag" : MessageLookupByLibrary.simpleMessage("Couscous"),
-    "crew_tag" : MessageLookupByLibrary.simpleMessage("Crew"),
-    "crochet_tag" : MessageLookupByLibrary.simpleMessage("Crochet"),
-    "crocodile_tag" : MessageLookupByLibrary.simpleMessage("Crocodile"),
-    "crowd_tag" : MessageLookupByLibrary.simpleMessage("Crowd"),
-    "cuisine_tag" : MessageLookupByLibrary.simpleMessage("Cuisine"),
-    "cup_tag" : MessageLookupByLibrary.simpleMessage("Cup"),
-    "curling_tag" : MessageLookupByLibrary.simpleMessage("Curling"),
-    "curtain_tag" : MessageLookupByLibrary.simpleMessage("Curtain"),
-    "cushion_tag" : MessageLookupByLibrary.simpleMessage("Cushion"),
-    "cutlery_tag" : MessageLookupByLibrary.simpleMessage("Cutlery"),
-    "cycling_tag" : MessageLookupByLibrary.simpleMessage("Cycling"),
-    "daily_challenge" : MessageLookupByLibrary.simpleMessage("Daily challenge"),
-    "daily_challenge_permission_description" : MessageLookupByLibrary.simpleMessage("For us to be able to send your daily challenges we need authorization to send notifications, so, it is necessary that you authorize the notifications in the options of your cell phone"),
-    "daily_goal" : MessageLookupByLibrary.simpleMessage("Daily goal"),
-    "daily_notification_description" : MessageLookupByLibrary.simpleMessage("It\'s time to complete your picPics daily challenge!"),
-    "daily_notification_title" : MessageLookupByLibrary.simpleMessage("Daily challenge"),
-    "dalmatian_tag" : MessageLookupByLibrary.simpleMessage("Dalmatian"),
-    "dam_tag" : MessageLookupByLibrary.simpleMessage("Dam"),
-    "dance_tag" : MessageLookupByLibrary.simpleMessage("Dance"),
-    "deejay_tag" : MessageLookupByLibrary.simpleMessage("Deejay"),
-    "delete" : MessageLookupByLibrary.simpleMessage("Delete"),
-    "denim_tag" : MessageLookupByLibrary.simpleMessage("Denim"),
-    "desert_tag" : MessageLookupByLibrary.simpleMessage("Desert"),
-    "desk_tag" : MessageLookupByLibrary.simpleMessage("Desk"),
-    "device_has_no_pics" : MessageLookupByLibrary.simpleMessage("This device has no photo in the gallery, so there is no photo that can be tagged."),
-    "dinosaur_tag" : MessageLookupByLibrary.simpleMessage("Dinosaur"),
-    "disable_secret" : MessageLookupByLibrary.simpleMessage("Do you want to unhide this photo?"),
-    "dog_tag" : MessageLookupByLibrary.simpleMessage("Dog"),
-    "doily_tag" : MessageLookupByLibrary.simpleMessage("Doily"),
-    "dont_ask_again" : MessageLookupByLibrary.simpleMessage("Don\'t ask again"),
-    "dragon_tag" : MessageLookupByLibrary.simpleMessage("Dragon"),
-    "drawer_tag" : MessageLookupByLibrary.simpleMessage("Drawer"),
-    "dress_tag" : MessageLookupByLibrary.simpleMessage("Dress"),
-    "duck_tag" : MessageLookupByLibrary.simpleMessage("Duck"),
-    "dude_tag" : MessageLookupByLibrary.simpleMessage("Dude"),
-    "dune_tag" : MessageLookupByLibrary.simpleMessage("Dune"),
-    "ear_tag" : MessageLookupByLibrary.simpleMessage("Ear"),
-    "eating_tag" : MessageLookupByLibrary.simpleMessage("Eating"),
-    "edit_tag" : MessageLookupByLibrary.simpleMessage("Edit tag"),
-    "email" : MessageLookupByLibrary.simpleMessage("Email"),
-    "enable_faceid" : MessageLookupByLibrary.simpleMessage("Enable Face ID"),
-    "enable_fingerprint" : MessageLookupByLibrary.simpleMessage("Enable Fingerprint"),
-    "enable_irisscanner" : MessageLookupByLibrary.simpleMessage("Enable Iris Scanner"),
-    "enable_touchid" : MessageLookupByLibrary.simpleMessage("Enable Touch ID"),
-    "event_tag" : MessageLookupByLibrary.simpleMessage("Event"),
-    "export_all_gallery" : MessageLookupByLibrary.simpleMessage("Export all gallery"),
-    "export_library" : MessageLookupByLibrary.simpleMessage("Export Library"),
-    "eyelash_tag" : MessageLookupByLibrary.simpleMessage("Eyelash"),
-    "factory_tag" : MessageLookupByLibrary.simpleMessage("Factory"),
-    "family_tag" : MessageLookupByLibrary.simpleMessage("Family"),
-    "farm_tag" : MessageLookupByLibrary.simpleMessage("Farm"),
-    "fast_food_tag" : MessageLookupByLibrary.simpleMessage("Fast food"),
-    "feedback_bug_report" : MessageLookupByLibrary.simpleMessage("Feedback & bug report"),
-    "ferris_wheel_tag" : MessageLookupByLibrary.simpleMessage("Ferris wheel"),
-    "fiction_tag" : MessageLookupByLibrary.simpleMessage("Fiction"),
-    "field_tag" : MessageLookupByLibrary.simpleMessage("Field"),
-    "find_easily" : MessageLookupByLibrary.simpleMessage("Easily find your photos"),
-    "fire_tag" : MessageLookupByLibrary.simpleMessage("Fire"),
-    "fireworks_tag" : MessageLookupByLibrary.simpleMessage("Fireworks"),
-    "fishing_tag" : MessageLookupByLibrary.simpleMessage("Fishing"),
-    "flag_tag" : MessageLookupByLibrary.simpleMessage("Flag"),
-    "flap_tag" : MessageLookupByLibrary.simpleMessage("Flap"),
-    "flesh_tag" : MessageLookupByLibrary.simpleMessage("Flesh"),
-    "flora_tag" : MessageLookupByLibrary.simpleMessage("Flora"),
-    "flower_tag" : MessageLookupByLibrary.simpleMessage("Flower"),
-    "flowerpot_tag" : MessageLookupByLibrary.simpleMessage("Flowerpot"),
-    "fog_tag" : MessageLookupByLibrary.simpleMessage("Fog"),
-    "food_tag" : MessageLookupByLibrary.simpleMessage("Food"),
-    "foods_tag" : MessageLookupByLibrary.simpleMessage("Foods"),
-    "foot_tag" : MessageLookupByLibrary.simpleMessage("Foot"),
-    "forest_tag" : MessageLookupByLibrary.simpleMessage("Forest"),
-    "forgot_secret_key" : MessageLookupByLibrary.simpleMessage("Forgot secret key?"),
-    "frigate_tag" : MessageLookupByLibrary.simpleMessage("Frigate"),
-    "fruit_tag" : MessageLookupByLibrary.simpleMessage("Fruit"),
-    "full_screen" : MessageLookupByLibrary.simpleMessage("Full Screen"),
-    "fun_tag" : MessageLookupByLibrary.simpleMessage("Fun"),
-    "fur_tag" : MessageLookupByLibrary.simpleMessage("Fur"),
-    "gallery_access_permission" : MessageLookupByLibrary.simpleMessage("Access permissions"),
-    "gallery_access_permission_description" : MessageLookupByLibrary.simpleMessage("To start organizing your photos, we need authorization to access them"),
-    "gallery_access_reason" : MessageLookupByLibrary.simpleMessage("To organize your photos we need access to your photo gallery"),
-    "garden_tag" : MessageLookupByLibrary.simpleMessage("Garden"),
-    "gelato_tag" : MessageLookupByLibrary.simpleMessage("Gelato"),
-    "gerbil_tag" : MessageLookupByLibrary.simpleMessage("Gerbil"),
-    "get_premium_description" : MessageLookupByLibrary.simpleMessage("TO HAVE ALL THESE FEATURES"),
-    "get_premium_now" : MessageLookupByLibrary.simpleMessage("Get premium now!"),
-    "get_premium_title" : MessageLookupByLibrary.simpleMessage("GET PREMIUM"),
-    "glacier_tag" : MessageLookupByLibrary.simpleMessage("Glacier"),
-    "glasses_tag" : MessageLookupByLibrary.simpleMessage("Glasses"),
-    "glitter_tag" : MessageLookupByLibrary.simpleMessage("Glitter"),
-    "goggles_tag" : MessageLookupByLibrary.simpleMessage("Goggles"),
-    "gown_tag" : MessageLookupByLibrary.simpleMessage("Gown"),
-    "graduation_tag" : MessageLookupByLibrary.simpleMessage("Graduation"),
-    "grandparent_tag" : MessageLookupByLibrary.simpleMessage("Grandparent"),
-    "groom_tag" : MessageLookupByLibrary.simpleMessage("Groom"),
-    "gymnastics_tag" : MessageLookupByLibrary.simpleMessage("Gymnastics"),
-    "hair_tag" : MessageLookupByLibrary.simpleMessage("Hair"),
-    "hand_tag" : MessageLookupByLibrary.simpleMessage("Hand"),
-    "handbag_tag" : MessageLookupByLibrary.simpleMessage("Handbag"),
-    "handrail_tag" : MessageLookupByLibrary.simpleMessage("Handrail"),
-    "hanukkah_tag" : MessageLookupByLibrary.simpleMessage("Hanukkah"),
-    "hat_tag" : MessageLookupByLibrary.simpleMessage("Hat"),
-    "heart_tag" : MessageLookupByLibrary.simpleMessage("Heart"),
-    "helicopter_tag" : MessageLookupByLibrary.simpleMessage("Helicopter"),
-    "helmet_tag" : MessageLookupByLibrary.simpleMessage("Helmet"),
-    "herd_tag" : MessageLookupByLibrary.simpleMessage("Herd"),
-    "himalayan_tag" : MessageLookupByLibrary.simpleMessage("Himalayan"),
-    "home_tag" : MessageLookupByLibrary.simpleMessage("Home"),
-    "horn_tag" : MessageLookupByLibrary.simpleMessage("Horn"),
-    "horse_tag" : MessageLookupByLibrary.simpleMessage("Horse"),
-    "hot_dog_tag" : MessageLookupByLibrary.simpleMessage("Hot dog"),
-    "how_many_pics" : MessageLookupByLibrary.simpleMessage("How many pics"),
-    "ice_tag" : MessageLookupByLibrary.simpleMessage("Ice"),
-    "iceberg_tag" : MessageLookupByLibrary.simpleMessage("Iceberg"),
-    "icicle_tag" : MessageLookupByLibrary.simpleMessage("Icicle"),
-    "icing_tag" : MessageLookupByLibrary.simpleMessage("Icing"),
-    "icon_tag" : MessageLookupByLibrary.simpleMessage("Icon"),
-    "infinite_tags" : MessageLookupByLibrary.simpleMessage("Infinite tags"),
-    "infrastructure_tag" : MessageLookupByLibrary.simpleMessage("Infrastructure"),
-    "insect_tag" : MessageLookupByLibrary.simpleMessage("Insect"),
-    "interaction_tag" : MessageLookupByLibrary.simpleMessage("Interaction"),
-    "jacket_tag" : MessageLookupByLibrary.simpleMessage("Jacket"),
-    "jeans_tag" : MessageLookupByLibrary.simpleMessage("Jeans"),
-    "jersey_tag" : MessageLookupByLibrary.simpleMessage("Jersey"),
-    "jewellery_tag" : MessageLookupByLibrary.simpleMessage("Jewellery"),
-    "joker_tag" : MessageLookupByLibrary.simpleMessage("Joker"),
-    "juice_tag" : MessageLookupByLibrary.simpleMessage("Juice"),
-    "jungle_tag" : MessageLookupByLibrary.simpleMessage("Jungle"),
-    "junk_tag" : MessageLookupByLibrary.simpleMessage("Junk"),
-    "kayak_tag" : MessageLookupByLibrary.simpleMessage("Kayak"),
-    "keep_asking" : MessageLookupByLibrary.simpleMessage("Keep asking"),
-    "keep_safe" : MessageLookupByLibrary.simpleMessage("Your photo is now safe on picPics. Do you want to delete it from your camera roll?"),
-    "kitchen_tag" : MessageLookupByLibrary.simpleMessage("Kitchen"),
-    "knitting_tag" : MessageLookupByLibrary.simpleMessage("Knitting"),
-    "lake_tag" : MessageLookupByLibrary.simpleMessage("Lake"),
-    "lampshade_tag" : MessageLookupByLibrary.simpleMessage("Lampshade"),
-    "language" : MessageLookupByLibrary.simpleMessage("Language"),
-    "larva_tag" : MessageLookupByLibrary.simpleMessage("Larva"),
-    "laugh_tag" : MessageLookupByLibrary.simpleMessage("Laugh"),
-    "leather_tag" : MessageLookupByLibrary.simpleMessage("Leather"),
-    "leggings_tag" : MessageLookupByLibrary.simpleMessage("Leggings"),
-    "lego_tag" : MessageLookupByLibrary.simpleMessage("Lego"),
-    "leisure_tag" : MessageLookupByLibrary.simpleMessage("Leisure"),
-    "lighthouse_tag" : MessageLookupByLibrary.simpleMessage("Lighthouse"),
-    "lightning_tag" : MessageLookupByLibrary.simpleMessage("Lightning"),
-    "lipstick_tag" : MessageLookupByLibrary.simpleMessage("Lipstick"),
-    "lock_with_pin" : MessageLookupByLibrary.simpleMessage("Lock down your private photos with a PIN passcode."),
-    "lock_your_photos" : MessageLookupByLibrary.simpleMessage("Lock your photos"),
-    "longboard_tag" : MessageLookupByLibrary.simpleMessage("Longboard"),
-    "love_tag" : MessageLookupByLibrary.simpleMessage("Love"),
-    "loveseat_tag" : MessageLookupByLibrary.simpleMessage("Loveseat"),
-    "lunch_tag" : MessageLookupByLibrary.simpleMessage("Lunch"),
-    "marathon_tag" : MessageLookupByLibrary.simpleMessage("Marathon"),
-    "marriage_tag" : MessageLookupByLibrary.simpleMessage("Marriage"),
-    "meal_tag" : MessageLookupByLibrary.simpleMessage("Meal"),
-    "menu_tag" : MessageLookupByLibrary.simpleMessage("Menu"),
-    "metal_tag" : MessageLookupByLibrary.simpleMessage("Metal"),
-    "militia_tag" : MessageLookupByLibrary.simpleMessage("Militia"),
-    "miniature_tag" : MessageLookupByLibrary.simpleMessage("Miniature"),
-    "mobile_phone_tag" : MessageLookupByLibrary.simpleMessage("Mobile phone"),
-    "model_tag" : MessageLookupByLibrary.simpleMessage("Model"),
-    "money_tag" : MessageLookupByLibrary.simpleMessage("Money"),
-    "monochrome_tag" : MessageLookupByLibrary.simpleMessage("Monochrome"),
-    "month" : MessageLookupByLibrary.simpleMessage("month"),
-    "monument_tag" : MessageLookupByLibrary.simpleMessage("Monument"),
-    "moon_tag" : MessageLookupByLibrary.simpleMessage("Moon"),
-    "mortarboard_tag" : MessageLookupByLibrary.simpleMessage("Mortarboard"),
-    "mosque_tag" : MessageLookupByLibrary.simpleMessage("Mosque"),
-    "motorcycle_tag" : MessageLookupByLibrary.simpleMessage("Motorcycle"),
-    "mountain_tag" : MessageLookupByLibrary.simpleMessage("Mountain"),
-    "moustache_tag" : MessageLookupByLibrary.simpleMessage("Moustache"),
-    "mouth_tag" : MessageLookupByLibrary.simpleMessage("Mouth"),
-    "mufti_tag" : MessageLookupByLibrary.simpleMessage("Mufti"),
-    "muscle_tag" : MessageLookupByLibrary.simpleMessage("Muscle"),
-    "museum_tag" : MessageLookupByLibrary.simpleMessage("Museum"),
-    "musical_instrument_tag" : MessageLookupByLibrary.simpleMessage("Musical instrument"),
-    "musical_tag" : MessageLookupByLibrary.simpleMessage("Musical"),
-    "musician_tag" : MessageLookupByLibrary.simpleMessage("Musician"),
-    "nail_tag" : MessageLookupByLibrary.simpleMessage("Nail"),
-    "nebula_tag" : MessageLookupByLibrary.simpleMessage("Nebula"),
-    "necklace_tag" : MessageLookupByLibrary.simpleMessage("Necklace"),
-    "needlework_tag" : MessageLookupByLibrary.simpleMessage("Needlework"),
-    "neon_tag" : MessageLookupByLibrary.simpleMessage("Neon"),
-    "net_tag" : MessageLookupByLibrary.simpleMessage("Net"),
-    "new_secret_key" : MessageLookupByLibrary.simpleMessage("New secret key"),
-    "news_tag" : MessageLookupByLibrary.simpleMessage("News"),
-    "newspaper_tag" : MessageLookupByLibrary.simpleMessage("Newspaper"),
-    "next" : MessageLookupByLibrary.simpleMessage("Next"),
-    "nightclub_tag" : MessageLookupByLibrary.simpleMessage("Nightclub"),
-    "no" : MessageLookupByLibrary.simpleMessage("No"),
-    "no_ads" : MessageLookupByLibrary.simpleMessage("No Ads"),
-    "no_previous_purchase" : MessageLookupByLibrary.simpleMessage("No Previous Purchase"),
-    "no_tagged_photos" : MessageLookupByLibrary.simpleMessage("You don\'t have any tagged photos yet"),
-    "no_tags_found" : MessageLookupByLibrary.simpleMessage("No tags found"),
-    "no_valid_subscription" : MessageLookupByLibrary.simpleMessage("Could not find a valid subscription purchase."),
-    "notification_time" : MessageLookupByLibrary.simpleMessage("Notification time"),
-    "notifications" : MessageLookupByLibrary.simpleMessage("Notifications"),
-    "ny_tag" : MessageLookupByLibrary.simpleMessage("NY"),
-    "odometer_tag" : MessageLookupByLibrary.simpleMessage("Odometer"),
-    "ok" : MessageLookupByLibrary.simpleMessage("OK"),
-    "open_gallery" : MessageLookupByLibrary.simpleMessage("Open gallery"),
-    "organized_photos_description" : MessageLookupByLibrary.simpleMessage("Photos that have already been tagged"),
-    "organized_photos_title" : MessageLookupByLibrary.simpleMessage("Organized Photos"),
-    "outerwear_tag" : MessageLookupByLibrary.simpleMessage("Outerwear"),
-    "palace_tag" : MessageLookupByLibrary.simpleMessage("Palace"),
-    "paper_tag" : MessageLookupByLibrary.simpleMessage("Paper"),
-    "park_tag" : MessageLookupByLibrary.simpleMessage("Park"),
-    "parties_tag" : MessageLookupByLibrary.simpleMessage("Parties"),
-    "party_tag" : MessageLookupByLibrary.simpleMessage("Party"),
-    "passport_tag" : MessageLookupByLibrary.simpleMessage("Passport"),
-    "pasteles_tag" : MessageLookupByLibrary.simpleMessage("Pasteles"),
-    "pattern_tag" : MessageLookupByLibrary.simpleMessage("Pattern"),
-    "penguin_tag" : MessageLookupByLibrary.simpleMessage("Penguin"),
-    "person_tag" : MessageLookupByLibrary.simpleMessage("Person"),
-    "pet_tag" : MessageLookupByLibrary.simpleMessage("Pet"),
-    "petal_tag" : MessageLookupByLibrary.simpleMessage("Petal"),
-    "pets_tag" : MessageLookupByLibrary.simpleMessage("Pets"),
-    "pho_tag" : MessageLookupByLibrary.simpleMessage("Pho"),
-    "photo_gallery_count" : m1,
-    "photo_gallery_description" : MessageLookupByLibrary.simpleMessage("Photos not yet organized"),
-    "photo_gallery_title" : MessageLookupByLibrary.simpleMessage("Photo Gallery"),
-    "photo_location" : MessageLookupByLibrary.simpleMessage("Photo location"),
-    "photos_always_organized" : MessageLookupByLibrary.simpleMessage("Now your photos will be always organized"),
-    "piano_tag" : MessageLookupByLibrary.simpleMessage("Piano"),
-    "picnic_tag" : MessageLookupByLibrary.simpleMessage("Picnic"),
-    "picpics_photo_manager" : MessageLookupByLibrary.simpleMessage("picPics - Photo Manager"),
-    "pie_tag" : MessageLookupByLibrary.simpleMessage("Pie"),
-    "pier_tag" : MessageLookupByLibrary.simpleMessage("Pier"),
-    "pillow_tag" : MessageLookupByLibrary.simpleMessage("Pillow"),
-    "pitch_tag" : MessageLookupByLibrary.simpleMessage("Pitch"),
-    "pixie_bob_tag" : MessageLookupByLibrary.simpleMessage("Pixie-bob"),
-    "pizza_tag" : MessageLookupByLibrary.simpleMessage("Pizza"),
-    "placemat_tag" : MessageLookupByLibrary.simpleMessage("Placemat"),
-    "plant_tag" : MessageLookupByLibrary.simpleMessage("Plant"),
-    "playground_tag" : MessageLookupByLibrary.simpleMessage("Playground"),
-    "plush_tag" : MessageLookupByLibrary.simpleMessage("Plush"),
-    "pocket_tag" : MessageLookupByLibrary.simpleMessage("Pocket"),
-    "polo_tag" : MessageLookupByLibrary.simpleMessage("Polo"),
-    "pomacentridae_tag" : MessageLookupByLibrary.simpleMessage("Pomacentridae"),
-    "pool_tag" : MessageLookupByLibrary.simpleMessage("Pool"),
-    "pop_music_tag" : MessageLookupByLibrary.simpleMessage("Pop music"),
-    "porcelain_tag" : MessageLookupByLibrary.simpleMessage("Porcelain"),
-    "poster_tag" : MessageLookupByLibrary.simpleMessage("Poster"),
-    "prairie_tag" : MessageLookupByLibrary.simpleMessage("Prairie"),
-    "premium_modal_description" : m2,
-    "premium_modal_get_premium_description" : MessageLookupByLibrary.simpleMessage("ALL FEATURES WITHOUT ADS"),
-    "premium_modal_get_premium_title" : MessageLookupByLibrary.simpleMessage("Get Premium Account"),
-    "premium_modal_watch_ad" : MessageLookupByLibrary.simpleMessage("Watch video ad to continue"),
-    "presentation_tag" : MessageLookupByLibrary.simpleMessage("Presentation"),
-    "privacy_policy" : MessageLookupByLibrary.simpleMessage("Privacy Policy"),
-    "private_photos" : MessageLookupByLibrary.simpleMessage("Private Photos"),
-    "product_tag" : MessageLookupByLibrary.simpleMessage("Product"),
-    "prom_tag" : MessageLookupByLibrary.simpleMessage("Prom"),
-    "protect_with_encryption" : MessageLookupByLibrary.simpleMessage("Protect your personal photos with encryption only accessible with a PIN password."),
-    "quilting_tag" : MessageLookupByLibrary.simpleMessage("Quilting"),
-    "race_tag" : MessageLookupByLibrary.simpleMessage("Race"),
-    "racing_tag" : MessageLookupByLibrary.simpleMessage("Racing"),
-    "rafting_tag" : MessageLookupByLibrary.simpleMessage("Rafting"),
-    "ragdoll_tag" : MessageLookupByLibrary.simpleMessage("Ragdoll"),
-    "rainbow_tag" : MessageLookupByLibrary.simpleMessage("Rainbow"),
-    "ranch_tag" : MessageLookupByLibrary.simpleMessage("Ranch"),
-    "rate_this_app" : MessageLookupByLibrary.simpleMessage("Rate this app"),
-    "receipt_tag" : MessageLookupByLibrary.simpleMessage("Receipt"),
-    "recent_tags" : MessageLookupByLibrary.simpleMessage("Recent Tags"),
-    "reef_tag" : MessageLookupByLibrary.simpleMessage("Reef"),
-    "rein_tag" : MessageLookupByLibrary.simpleMessage("Rein"),
-    "require_secret_key" : MessageLookupByLibrary.simpleMessage("Require secret key"),
-    "restore_purchase" : MessageLookupByLibrary.simpleMessage("Restore purchase"),
-    "rickshaw_tag" : MessageLookupByLibrary.simpleMessage("Rickshaw"),
-    "ring_tag" : MessageLookupByLibrary.simpleMessage("Ring"),
-    "river_tag" : MessageLookupByLibrary.simpleMessage("River"),
-    "road_tag" : MessageLookupByLibrary.simpleMessage("Road"),
-    "rock_tag" : MessageLookupByLibrary.simpleMessage("Rock"),
-    "rocket_tag" : MessageLookupByLibrary.simpleMessage("Rocket"),
-    "rodeo_tag" : MessageLookupByLibrary.simpleMessage("Rodeo"),
-    "roller_tag" : MessageLookupByLibrary.simpleMessage("Roller"),
-    "roof_tag" : MessageLookupByLibrary.simpleMessage("Roof"),
-    "room_tag" : MessageLookupByLibrary.simpleMessage("Room"),
-    "rowing_tag" : MessageLookupByLibrary.simpleMessage("Rowing"),
-    "rugby_tag" : MessageLookupByLibrary.simpleMessage("Rugby"),
-    "ruins_tag" : MessageLookupByLibrary.simpleMessage("Ruins"),
-    "running_tag" : MessageLookupByLibrary.simpleMessage("Running"),
-    "safari_tag" : MessageLookupByLibrary.simpleMessage("Safari"),
-    "sailboat_tag" : MessageLookupByLibrary.simpleMessage("Sailboat"),
-    "sand_tag" : MessageLookupByLibrary.simpleMessage("Sand"),
-    "santa_claus_tag" : MessageLookupByLibrary.simpleMessage("Santa claus"),
-    "sari_tag" : MessageLookupByLibrary.simpleMessage("Sari"),
-    "saucer_tag" : MessageLookupByLibrary.simpleMessage("Saucer"),
-    "save" : MessageLookupByLibrary.simpleMessage("save"),
-    "save_location" : MessageLookupByLibrary.simpleMessage("Save location"),
-    "scarf_tag" : MessageLookupByLibrary.simpleMessage("Scarf"),
-    "school_tag" : MessageLookupByLibrary.simpleMessage("School"),
-    "screenshot_tag" : MessageLookupByLibrary.simpleMessage("Screenshot"),
-    "screenshots_tag" : MessageLookupByLibrary.simpleMessage("Screenshots"),
-    "scuba_diving_tag" : MessageLookupByLibrary.simpleMessage("Scuba diving"),
-    "seal_tag" : MessageLookupByLibrary.simpleMessage("Seal"),
-    "search" : MessageLookupByLibrary.simpleMessage("Search..."),
-    "search_all_tags_not_found" : MessageLookupByLibrary.simpleMessage("No pictures were found with all tags on it."),
-    "search_results" : MessageLookupByLibrary.simpleMessage("Search results"),
-    "secret_key_created" : MessageLookupByLibrary.simpleMessage("Secret key successfully created!"),
-    "secret_photos" : MessageLookupByLibrary.simpleMessage("Secret Photos"),
-    "selfie_tag" : MessageLookupByLibrary.simpleMessage("Selfie"),
-    "selfies_tag" : MessageLookupByLibrary.simpleMessage("Selfies"),
-    "settings" : MessageLookupByLibrary.simpleMessage("Settings"),
-    "share_with_friends" : MessageLookupByLibrary.simpleMessage("Share with friends"),
-    "shelf_tag" : MessageLookupByLibrary.simpleMessage("Shelf"),
-    "shell_tag" : MessageLookupByLibrary.simpleMessage("Shell"),
-    "shetland_sheepdog_tag" : MessageLookupByLibrary.simpleMessage("Shetland sheepdog"),
-    "shikoku_tag" : MessageLookupByLibrary.simpleMessage("Shikoku"),
-    "shipwreck_tag" : MessageLookupByLibrary.simpleMessage("Shipwreck"),
-    "shoe_tag" : MessageLookupByLibrary.simpleMessage("Shoe"),
-    "shorts_tag" : MessageLookupByLibrary.simpleMessage("Shorts"),
-    "sign" : MessageLookupByLibrary.simpleMessage("Sign"),
-    "singer_tag" : MessageLookupByLibrary.simpleMessage("Singer"),
-    "sink_tag" : MessageLookupByLibrary.simpleMessage("Sink"),
-    "sitting_tag" : MessageLookupByLibrary.simpleMessage("Sitting"),
-    "skateboard_tag" : MessageLookupByLibrary.simpleMessage("Skateboard"),
-    "skateboarder_tag" : MessageLookupByLibrary.simpleMessage("Skateboarder"),
-    "skiff_tag" : MessageLookupByLibrary.simpleMessage("Skiff"),
-    "skiing_tag" : MessageLookupByLibrary.simpleMessage("Skiing"),
-    "skin_tag" : MessageLookupByLibrary.simpleMessage("Skin"),
-    "sky_tag" : MessageLookupByLibrary.simpleMessage("Sky"),
-    "skyline_tag" : MessageLookupByLibrary.simpleMessage("Skyline"),
-    "skyscraper_tag" : MessageLookupByLibrary.simpleMessage("Skyscraper"),
-    "sled_tag" : MessageLookupByLibrary.simpleMessage("Sled"),
-    "sledding_tag" : MessageLookupByLibrary.simpleMessage("Sledding"),
-    "sleep_tag" : MessageLookupByLibrary.simpleMessage("Sleep"),
-    "smile_tag" : MessageLookupByLibrary.simpleMessage("Smile"),
-    "sneakers_tag" : MessageLookupByLibrary.simpleMessage("Sneakers"),
-    "snorkeling_tag" : MessageLookupByLibrary.simpleMessage("Snorkeling"),
-    "snowboarding_tag" : MessageLookupByLibrary.simpleMessage("Snowboarding"),
-    "soccer_tag" : MessageLookupByLibrary.simpleMessage("Soccer"),
-    "softball_tag" : MessageLookupByLibrary.simpleMessage("Softball"),
-    "soil_tag" : MessageLookupByLibrary.simpleMessage("Soil"),
-    "song_tag" : MessageLookupByLibrary.simpleMessage("Song"),
-    "space_tag" : MessageLookupByLibrary.simpleMessage("Space"),
-    "sparkler_tag" : MessageLookupByLibrary.simpleMessage("Sparkler"),
-    "speedboat_tag" : MessageLookupByLibrary.simpleMessage("Speedboat"),
-    "sphynx_tag" : MessageLookupByLibrary.simpleMessage("Sphynx"),
-    "sports_tag" : MessageLookupByLibrary.simpleMessage("Sports"),
-    "stadium_tag" : MessageLookupByLibrary.simpleMessage("Stadium"),
-    "stairs_tag" : MessageLookupByLibrary.simpleMessage("Stairs"),
-    "standing_tag" : MessageLookupByLibrary.simpleMessage("Standing"),
-    "star_tag" : MessageLookupByLibrary.simpleMessage("Star"),
-    "start" : MessageLookupByLibrary.simpleMessage("Start"),
-    "start_tagging" : MessageLookupByLibrary.simpleMessage("Start tagging"),
-    "statue_tag" : MessageLookupByLibrary.simpleMessage("Statue"),
-    "steaming_tag" : MessageLookupByLibrary.simpleMessage("Steaming"),
-    "storm_tag" : MessageLookupByLibrary.simpleMessage("Storm"),
-    "strap_tag" : MessageLookupByLibrary.simpleMessage("Strap"),
-    "stuffed_toy_tag" : MessageLookupByLibrary.simpleMessage("Stuffed toy"),
-    "submarine_tag" : MessageLookupByLibrary.simpleMessage("Submarine"),
-    "subwoofer_tag" : MessageLookupByLibrary.simpleMessage("Subwoofer"),
-    "suggestions" : MessageLookupByLibrary.simpleMessage("Suggestions"),
-    "sunglasses_tag" : MessageLookupByLibrary.simpleMessage("Sunglasses"),
-    "sunset_tag" : MessageLookupByLibrary.simpleMessage("Sunset"),
-    "superhero_tag" : MessageLookupByLibrary.simpleMessage("Superhero"),
-    "superman_tag" : MessageLookupByLibrary.simpleMessage("Superman"),
-    "supervillain_tag" : MessageLookupByLibrary.simpleMessage("Supervillain"),
-    "supper_tag" : MessageLookupByLibrary.simpleMessage("Supper"),
-    "surfboard_tag" : MessageLookupByLibrary.simpleMessage("Surfboard"),
-    "surfing_tag" : MessageLookupByLibrary.simpleMessage("Surfing"),
-    "sushi_tag" : MessageLookupByLibrary.simpleMessage("Sushi"),
-    "swamp_tag" : MessageLookupByLibrary.simpleMessage("Swamp"),
-    "swimming_tag" : MessageLookupByLibrary.simpleMessage("Swimming"),
-    "swimwear_tag" : MessageLookupByLibrary.simpleMessage("Swimwear"),
-    "swing_tag" : MessageLookupByLibrary.simpleMessage("Swing"),
-    "tablecloth_tag" : MessageLookupByLibrary.simpleMessage("Tablecloth"),
-    "tableware_tag" : MessageLookupByLibrary.simpleMessage("Tableware"),
-    "tag_multiple_photos_at_once" : MessageLookupByLibrary.simpleMessage("Tag multiple photos at once"),
-    "take_a_look" : MessageLookupByLibrary.simpleMessage("Take a look at this app!"),
-    "take_a_look_description" : m3,
-    "tattoo_tag" : MessageLookupByLibrary.simpleMessage("Tattoo"),
-    "team_tag" : MessageLookupByLibrary.simpleMessage("Team"),
-    "television_tag" : MessageLookupByLibrary.simpleMessage("Television"),
-    "temple_tag" : MessageLookupByLibrary.simpleMessage("Temple"),
-    "terms_of_use" : MessageLookupByLibrary.simpleMessage("Terms of Use"),
-    "textile_tag" : MessageLookupByLibrary.simpleMessage("Textile"),
-    "thanksgiving_tag" : MessageLookupByLibrary.simpleMessage("Thanksgiving"),
-    "tie_tag" : MessageLookupByLibrary.simpleMessage("Tie"),
-    "tights_tag" : MessageLookupByLibrary.simpleMessage("Tights"),
-    "tile_tag" : MessageLookupByLibrary.simpleMessage("Tile"),
-    "time" : MessageLookupByLibrary.simpleMessage("Time"),
-    "tire_tag" : MessageLookupByLibrary.simpleMessage("Tire"),
-    "toe_tag" : MessageLookupByLibrary.simpleMessage("Toe"),
-    "toggle_date" : MessageLookupByLibrary.simpleMessage("Date"),
-    "toggle_days" : MessageLookupByLibrary.simpleMessage("Days"),
-    "toggle_months" : MessageLookupByLibrary.simpleMessage("Months"),
-    "toggle_tags" : MessageLookupByLibrary.simpleMessage("Tags"),
-    "tower_tag" : MessageLookupByLibrary.simpleMessage("Tower"),
-    "toy_tag" : MessageLookupByLibrary.simpleMessage("Toy"),
-    "track_tag" : MessageLookupByLibrary.simpleMessage("Track"),
-    "tractor_tag" : MessageLookupByLibrary.simpleMessage("Tractor"),
-    "train_tag" : MessageLookupByLibrary.simpleMessage("Train"),
-    "travel_tag" : MessageLookupByLibrary.simpleMessage("Travel"),
-    "trunk_tag" : MessageLookupByLibrary.simpleMessage("Trunk"),
-    "tsqr_tag" : MessageLookupByLibrary.simpleMessage("Times Square"),
-    "tubing_tag" : MessageLookupByLibrary.simpleMessage("Tubing"),
-    "turtle_tag" : MessageLookupByLibrary.simpleMessage("Turtle"),
-    "tutorial_daily_package" : MessageLookupByLibrary.simpleMessage("We bring a daily package for you to gradually organize your library."),
-    "tutorial_however_you_want" : MessageLookupByLibrary.simpleMessage("Organize your photos by adding tags, like \"family\", \"pets\", or whatever you want."),
-    "tutorial_just_swipe" : MessageLookupByLibrary.simpleMessage("After adding the tags to your photo, just swipe to go to the next one."),
-    "tutorial_multiselect" : MessageLookupByLibrary.simpleMessage("You can \"tap & hold\" on your photos to tag multiple photos at once."),
-    "tutorial_secret" : MessageLookupByLibrary.simpleMessage("Hide your private photos with a pin code protection, keeping them safe."),
-    "tuxedo_tag" : MessageLookupByLibrary.simpleMessage("Tuxedo"),
-    "twig_tag" : MessageLookupByLibrary.simpleMessage("Twig"),
-    "umbrella_tag" : MessageLookupByLibrary.simpleMessage("Umbrella"),
-    "underwater_tag" : MessageLookupByLibrary.simpleMessage("Underwater"),
-    "unicycle_tag" : MessageLookupByLibrary.simpleMessage("Unicycle"),
-    "unlimited_private_pics" : MessageLookupByLibrary.simpleMessage("Unlimited private photos"),
-    "vacation_tag" : MessageLookupByLibrary.simpleMessage("Vacation"),
-    "van_tag" : MessageLookupByLibrary.simpleMessage("Van"),
-    "vegetable_tag" : MessageLookupByLibrary.simpleMessage("Vegetable"),
-    "vehicle_tag" : MessageLookupByLibrary.simpleMessage("Vehicle"),
-    "veil_tag" : MessageLookupByLibrary.simpleMessage("Veil"),
-    "view_hidden_photos" : MessageLookupByLibrary.simpleMessage("To view your hidden photos, release the lock in the application settings. You can unlock them with your PIN."),
-    "volcano_tag" : MessageLookupByLibrary.simpleMessage("Volcano"),
-    "wakeboarding_tag" : MessageLookupByLibrary.simpleMessage("Wakeboarding"),
-    "wall_tag" : MessageLookupByLibrary.simpleMessage("Wall"),
-    "watercolor_paint_tag" : MessageLookupByLibrary.simpleMessage("Watercolor paint"),
-    "waterfall_tag" : MessageLookupByLibrary.simpleMessage("Waterfall"),
-    "waterfowl_tag" : MessageLookupByLibrary.simpleMessage("Waterfowl"),
-    "waterskiing_tag" : MessageLookupByLibrary.simpleMessage("Waterskiing"),
-    "web_page_tag" : MessageLookupByLibrary.simpleMessage("Web page"),
-    "welcome" : MessageLookupByLibrary.simpleMessage("Welcome!"),
-    "wetsuit_tag" : MessageLookupByLibrary.simpleMessage("Wetsuit"),
-    "wheel_tag" : MessageLookupByLibrary.simpleMessage("Wheel"),
-    "wheelbarrow_tag" : MessageLookupByLibrary.simpleMessage("Wheelbarrow"),
-    "whiteboard_tag" : MessageLookupByLibrary.simpleMessage("Whiteboard"),
-    "windshield_tag" : MessageLookupByLibrary.simpleMessage("Windshield"),
-    "windsurfing_tag" : MessageLookupByLibrary.simpleMessage("Windsurfing"),
-    "wine_tag" : MessageLookupByLibrary.simpleMessage("Wine"),
-    "wing_tag" : MessageLookupByLibrary.simpleMessage("Wing"),
-    "wool_tag" : MessageLookupByLibrary.simpleMessage("Wool"),
-    "work_tag" : MessageLookupByLibrary.simpleMessage("Work"),
-    "wreath_tag" : MessageLookupByLibrary.simpleMessage("Wreath"),
-    "x_minutes" : MessageLookupByLibrary.simpleMessage("20 min"),
-    "year" : MessageLookupByLibrary.simpleMessage("year"),
-    "yes" : MessageLookupByLibrary.simpleMessage("Yes"),
-    "you_are_premium" : MessageLookupByLibrary.simpleMessage("You are premium!"),
-    "your_secret_key" : MessageLookupByLibrary.simpleMessage("Your secret key")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "access_code": MessageLookupByLibrary.simpleMessage("Access code"),
+        "access_code_sent": m0,
+        "add_multiple_tags":
+            MessageLookupByLibrary.simpleMessage("Adicione múltiplas tags"),
+        "add_tag": MessageLookupByLibrary.simpleMessage("Add tag"),
+        "add_tags": MessageLookupByLibrary.simpleMessage("Add tags"),
+        "aerospace_engineering_tag":
+            MessageLookupByLibrary.simpleMessage("Aerospace engineering"),
+        "aircraft_tag": MessageLookupByLibrary.simpleMessage("Aircraft"),
+        "airliner_tag": MessageLookupByLibrary.simpleMessage("Airliner"),
+        "airplane_tag": MessageLookupByLibrary.simpleMessage("Airplane"),
+        "alcohol_tag": MessageLookupByLibrary.simpleMessage("Alcohol"),
+        "all_at_once": MessageLookupByLibrary.simpleMessage(
+            "Organize multiple photos at once"),
+        "all_photos_were_tagged": MessageLookupByLibrary.simpleMessage(
+            "There are no more photos to organize."),
+        "all_search_tags":
+            MessageLookupByLibrary.simpleMessage("All Search Tags"),
+        "always": MessageLookupByLibrary.simpleMessage("Always"),
+        "aquarium_tag": MessageLookupByLibrary.simpleMessage("Aquarium"),
+        "archery_tag": MessageLookupByLibrary.simpleMessage("Archery"),
+        "armrest_tag": MessageLookupByLibrary.simpleMessage("Armrest"),
+        "army_tag": MessageLookupByLibrary.simpleMessage("Army"),
+        "ask_photo_library_permission": MessageLookupByLibrary.simpleMessage(
+            "We need access to your photo library so you can start organizing your photos with picPics. Don\'t worry your data will never leave your device!"),
+        "asphalt_tag": MessageLookupByLibrary.simpleMessage("Asphalt"),
+        "aurora_tag": MessageLookupByLibrary.simpleMessage("Aurora"),
+        "auto_renewable_first_part":
+            MessageLookupByLibrary.simpleMessage("The subscription is "),
+        "auto_renewable_second_part":
+            MessageLookupByLibrary.simpleMessage("auto-renewable."),
+        "aviation_tag": MessageLookupByLibrary.simpleMessage("Aviation"),
+        "baby_tag": MessageLookupByLibrary.simpleMessage("Baby"),
+        "backpacking_tag": MessageLookupByLibrary.simpleMessage("Backpacking"),
+        "badminton_tag": MessageLookupByLibrary.simpleMessage("Badminton"),
+        "bag_tag": MessageLookupByLibrary.simpleMessage("Bag"),
+        "balance_tag": MessageLookupByLibrary.simpleMessage("Balance"),
+        "balloon_tag": MessageLookupByLibrary.simpleMessage("Balloon"),
+        "ballroom_tag": MessageLookupByLibrary.simpleMessage("Ballroom"),
+        "bangle_tag": MessageLookupByLibrary.simpleMessage("Bangle"),
+        "bangs_tag": MessageLookupByLibrary.simpleMessage("Bangs"),
+        "bar_tag": MessageLookupByLibrary.simpleMessage("Bar"),
+        "barn_tag": MessageLookupByLibrary.simpleMessage("Barn"),
+        "basset_hound_tag":
+            MessageLookupByLibrary.simpleMessage("Basset hound"),
+        "bathing_tag": MessageLookupByLibrary.simpleMessage("Bathing"),
+        "bathroom_tag": MessageLookupByLibrary.simpleMessage("Bathroom"),
+        "beach_tag": MessageLookupByLibrary.simpleMessage("Beach"),
+        "beanie_tag": MessageLookupByLibrary.simpleMessage("Beanie"),
+        "bear_tag": MessageLookupByLibrary.simpleMessage("Bear"),
+        "beard_tag": MessageLookupByLibrary.simpleMessage("Beard"),
+        "bedroom_tag": MessageLookupByLibrary.simpleMessage("Bedroom"),
+        "bench_tag": MessageLookupByLibrary.simpleMessage("Bench"),
+        "bento_tag": MessageLookupByLibrary.simpleMessage("Bento"),
+        "bicycle_tag": MessageLookupByLibrary.simpleMessage("Bicycle"),
+        "bird_tag": MessageLookupByLibrary.simpleMessage("Bird"),
+        "blackboard_tag": MessageLookupByLibrary.simpleMessage("Blackboard"),
+        "blazer_tag": MessageLookupByLibrary.simpleMessage("Blazer"),
+        "boat_tag": MessageLookupByLibrary.simpleMessage("Boat"),
+        "bonfire_tag": MessageLookupByLibrary.simpleMessage("Bonfire"),
+        "boxer_tag": MessageLookupByLibrary.simpleMessage("Boxer"),
+        "bracelet_tag": MessageLookupByLibrary.simpleMessage("Bracelet"),
+        "branch_tag": MessageLookupByLibrary.simpleMessage("Branch"),
+        "bread_tag": MessageLookupByLibrary.simpleMessage("Bread"),
+        "brick_tag": MessageLookupByLibrary.simpleMessage("Brick"),
+        "bride_tag": MessageLookupByLibrary.simpleMessage("Bride"),
+        "bridge_tag": MessageLookupByLibrary.simpleMessage("Bridge"),
+        "bridle_tag": MessageLookupByLibrary.simpleMessage("Bridle"),
+        "brig_tag": MessageLookupByLibrary.simpleMessage("Brig"),
+        "building_tag": MessageLookupByLibrary.simpleMessage("Building"),
+        "bull_tag": MessageLookupByLibrary.simpleMessage("Bull"),
+        "bullfighting_tag":
+            MessageLookupByLibrary.simpleMessage("Bullfighting"),
+        "bumper_tag": MessageLookupByLibrary.simpleMessage("Bumper"),
+        "bunk_bed_tag": MessageLookupByLibrary.simpleMessage("Bunk bed"),
+        "bus_tag": MessageLookupByLibrary.simpleMessage("Bus"),
+        "butterfly_tag": MessageLookupByLibrary.simpleMessage("Butterfly"),
+        "cabinetry_tag": MessageLookupByLibrary.simpleMessage("Cabinetry"),
+        "cage_tag": MessageLookupByLibrary.simpleMessage("Cage"),
+        "cairn_terrier_tag":
+            MessageLookupByLibrary.simpleMessage("Cairn terrier"),
+        "cake_tag": MessageLookupByLibrary.simpleMessage("Cake"),
+        "camping_tag": MessageLookupByLibrary.simpleMessage("Camping"),
+        "cancel": MessageLookupByLibrary.simpleMessage("Cancel"),
+        "cancel_anytime":
+            MessageLookupByLibrary.simpleMessage("Cancel anytime"),
+        "canoe_tag": MessageLookupByLibrary.simpleMessage("Canoe"),
+        "canyon_tag": MessageLookupByLibrary.simpleMessage("Canyon"),
+        "cap_tag": MessageLookupByLibrary.simpleMessage("Cap"),
+        "cappuccino_tag": MessageLookupByLibrary.simpleMessage("Cappuccino"),
+        "car_tag": MessageLookupByLibrary.simpleMessage("Car"),
+        "carnival_tag": MessageLookupByLibrary.simpleMessage("Carnival"),
+        "casino_tag": MessageLookupByLibrary.simpleMessage("Casino"),
+        "castle_tag": MessageLookupByLibrary.simpleMessage("Castle"),
+        "cat_tag": MessageLookupByLibrary.simpleMessage("Cat"),
+        "cathedral_tag": MessageLookupByLibrary.simpleMessage("Cathedral"),
+        "cattle_tag": MessageLookupByLibrary.simpleMessage("Cattle"),
+        "cavalier_tag": MessageLookupByLibrary.simpleMessage("Cavalier"),
+        "cave_tag": MessageLookupByLibrary.simpleMessage("Cave"),
+        "caving_tag": MessageLookupByLibrary.simpleMessage("Caving"),
+        "centrepiece_tag": MessageLookupByLibrary.simpleMessage("Centrepiece"),
+        "chain_tag": MessageLookupByLibrary.simpleMessage("Chain"),
+        "chair_tag": MessageLookupByLibrary.simpleMessage("Chair"),
+        "cheeseburger_tag":
+            MessageLookupByLibrary.simpleMessage("Cheeseburger"),
+        "christmas_tag": MessageLookupByLibrary.simpleMessage("Christmas"),
+        "church_tag": MessageLookupByLibrary.simpleMessage("Church"),
+        "circus_tag": MessageLookupByLibrary.simpleMessage("Circus"),
+        "class_tag": MessageLookupByLibrary.simpleMessage("Class"),
+        "cliff_tag": MessageLookupByLibrary.simpleMessage("Cliff"),
+        "clipper_tag": MessageLookupByLibrary.simpleMessage("Clipper"),
+        "clock_tag": MessageLookupByLibrary.simpleMessage("Clock"),
+        "close": MessageLookupByLibrary.simpleMessage("Close"),
+        "clown_tag": MessageLookupByLibrary.simpleMessage("Clown"),
+        "coffee_tag": MessageLookupByLibrary.simpleMessage("Coffee"),
+        "cola_tag": MessageLookupByLibrary.simpleMessage("Cola"),
+        "comet_tag": MessageLookupByLibrary.simpleMessage("Comet"),
+        "comics_tag": MessageLookupByLibrary.simpleMessage("Comics"),
+        "community_tag": MessageLookupByLibrary.simpleMessage("Community"),
+        "competition_tag": MessageLookupByLibrary.simpleMessage("Competition"),
+        "computer_tag": MessageLookupByLibrary.simpleMessage("Computer"),
+        "concert_tag": MessageLookupByLibrary.simpleMessage("Concert"),
+        "confirm_email": MessageLookupByLibrary.simpleMessage(
+            "Confirm your registration email to receive your access code"),
+        "confirm_secret_key":
+            MessageLookupByLibrary.simpleMessage("Confirm secret key"),
+        "construction_tag":
+            MessageLookupByLibrary.simpleMessage("Construction"),
+        "continue_string": MessageLookupByLibrary.simpleMessage("Continue"),
+        "cookie_tag": MessageLookupByLibrary.simpleMessage("Cookie"),
+        "cookware_and_bakeware_tag":
+            MessageLookupByLibrary.simpleMessage("Cookware and bakeware"),
+        "cool_tag": MessageLookupByLibrary.simpleMessage("Cool"),
+        "cotton_tag": MessageLookupByLibrary.simpleMessage("Cotton"),
+        "couch_tag": MessageLookupByLibrary.simpleMessage("Couch"),
+        "countertop_tag": MessageLookupByLibrary.simpleMessage("Countertop"),
+        "country": MessageLookupByLibrary.simpleMessage("country"),
+        "couscous_tag": MessageLookupByLibrary.simpleMessage("Couscous"),
+        "crew_tag": MessageLookupByLibrary.simpleMessage("Crew"),
+        "crochet_tag": MessageLookupByLibrary.simpleMessage("Crochet"),
+        "crocodile_tag": MessageLookupByLibrary.simpleMessage("Crocodile"),
+        "crowd_tag": MessageLookupByLibrary.simpleMessage("Crowd"),
+        "cuisine_tag": MessageLookupByLibrary.simpleMessage("Cuisine"),
+        "cup_tag": MessageLookupByLibrary.simpleMessage("Cup"),
+        "curling_tag": MessageLookupByLibrary.simpleMessage("Curling"),
+        "curtain_tag": MessageLookupByLibrary.simpleMessage("Curtain"),
+        "cushion_tag": MessageLookupByLibrary.simpleMessage("Cushion"),
+        "cutlery_tag": MessageLookupByLibrary.simpleMessage("Cutlery"),
+        "cycling_tag": MessageLookupByLibrary.simpleMessage("Cycling"),
+        "daily_challenge":
+            MessageLookupByLibrary.simpleMessage("Daily challenge"),
+        "daily_challenge_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "For us to be able to send your daily challenges we need authorization to send notifications, so, it is necessary that you authorize the notifications in the options of your cell phone"),
+        "daily_goal": MessageLookupByLibrary.simpleMessage("Daily goal"),
+        "daily_notification_description": MessageLookupByLibrary.simpleMessage(
+            "It\'s time to complete your picPics daily challenge!"),
+        "daily_notification_title":
+            MessageLookupByLibrary.simpleMessage("Daily challenge"),
+        "dalmatian_tag": MessageLookupByLibrary.simpleMessage("Dalmatian"),
+        "dam_tag": MessageLookupByLibrary.simpleMessage("Dam"),
+        "dance_tag": MessageLookupByLibrary.simpleMessage("Dance"),
+        "deejay_tag": MessageLookupByLibrary.simpleMessage("Deejay"),
+        "delete": MessageLookupByLibrary.simpleMessage("Delete"),
+        "denim_tag": MessageLookupByLibrary.simpleMessage("Denim"),
+        "desert_tag": MessageLookupByLibrary.simpleMessage("Desert"),
+        "desk_tag": MessageLookupByLibrary.simpleMessage("Desk"),
+        "device_has_no_pics": MessageLookupByLibrary.simpleMessage(
+            "This device has no photo in the gallery, so there is no photo that can be tagged."),
+        "dinosaur_tag": MessageLookupByLibrary.simpleMessage("Dinosaur"),
+        "disable_secret": MessageLookupByLibrary.simpleMessage(
+            "Do you want to unhide this photo?"),
+        "dog_tag": MessageLookupByLibrary.simpleMessage("Dog"),
+        "doily_tag": MessageLookupByLibrary.simpleMessage("Doily"),
+        "dont_ask_again":
+            MessageLookupByLibrary.simpleMessage("Don\'t ask again"),
+        "dragon_tag": MessageLookupByLibrary.simpleMessage("Dragon"),
+        "drawer_tag": MessageLookupByLibrary.simpleMessage("Drawer"),
+        "dress_tag": MessageLookupByLibrary.simpleMessage("Dress"),
+        "duck_tag": MessageLookupByLibrary.simpleMessage("Duck"),
+        "dude_tag": MessageLookupByLibrary.simpleMessage("Dude"),
+        "dune_tag": MessageLookupByLibrary.simpleMessage("Dune"),
+        "ear_tag": MessageLookupByLibrary.simpleMessage("Ear"),
+        "eating_tag": MessageLookupByLibrary.simpleMessage("Eating"),
+        "edit_tag": MessageLookupByLibrary.simpleMessage("Edit tag"),
+        "email": MessageLookupByLibrary.simpleMessage("Email"),
+        "enable_faceid": MessageLookupByLibrary.simpleMessage("Enable Face ID"),
+        "enable_fingerprint":
+            MessageLookupByLibrary.simpleMessage("Enable Fingerprint"),
+        "enable_irisscanner":
+            MessageLookupByLibrary.simpleMessage("Enable Iris Scanner"),
+        "enable_touchid":
+            MessageLookupByLibrary.simpleMessage("Enable Touch ID"),
+        "event_tag": MessageLookupByLibrary.simpleMessage("Event"),
+        "export_all_gallery":
+            MessageLookupByLibrary.simpleMessage("Export all gallery"),
+        "export_library":
+            MessageLookupByLibrary.simpleMessage("Export Library"),
+        "eyelash_tag": MessageLookupByLibrary.simpleMessage("Eyelash"),
+        "factory_tag": MessageLookupByLibrary.simpleMessage("Factory"),
+        "family_tag": MessageLookupByLibrary.simpleMessage("Family"),
+        "farm_tag": MessageLookupByLibrary.simpleMessage("Farm"),
+        "fast_food_tag": MessageLookupByLibrary.simpleMessage("Fast food"),
+        "feedback_bug_report":
+            MessageLookupByLibrary.simpleMessage("Feedback & bug report"),
+        "ferris_wheel_tag":
+            MessageLookupByLibrary.simpleMessage("Ferris wheel"),
+        "fiction_tag": MessageLookupByLibrary.simpleMessage("Fiction"),
+        "field_tag": MessageLookupByLibrary.simpleMessage("Field"),
+        "find_easily":
+            MessageLookupByLibrary.simpleMessage("Easily find your photos"),
+        "fire_tag": MessageLookupByLibrary.simpleMessage("Fire"),
+        "fireworks_tag": MessageLookupByLibrary.simpleMessage("Fireworks"),
+        "fishing_tag": MessageLookupByLibrary.simpleMessage("Fishing"),
+        "flag_tag": MessageLookupByLibrary.simpleMessage("Flag"),
+        "flap_tag": MessageLookupByLibrary.simpleMessage("Flap"),
+        "flesh_tag": MessageLookupByLibrary.simpleMessage("Flesh"),
+        "flora_tag": MessageLookupByLibrary.simpleMessage("Flora"),
+        "flower_tag": MessageLookupByLibrary.simpleMessage("Flower"),
+        "flowerpot_tag": MessageLookupByLibrary.simpleMessage("Flowerpot"),
+        "fog_tag": MessageLookupByLibrary.simpleMessage("Fog"),
+        "food_tag": MessageLookupByLibrary.simpleMessage("Food"),
+        "foods_tag": MessageLookupByLibrary.simpleMessage("Foods"),
+        "foot_tag": MessageLookupByLibrary.simpleMessage("Foot"),
+        "forest_tag": MessageLookupByLibrary.simpleMessage("Forest"),
+        "forgot_secret_key":
+            MessageLookupByLibrary.simpleMessage("Forgot secret key?"),
+        "frigate_tag": MessageLookupByLibrary.simpleMessage("Frigate"),
+        "fruit_tag": MessageLookupByLibrary.simpleMessage("Fruit"),
+        "full_screen": MessageLookupByLibrary.simpleMessage("Full Screen"),
+        "fun_tag": MessageLookupByLibrary.simpleMessage("Fun"),
+        "fur_tag": MessageLookupByLibrary.simpleMessage("Fur"),
+        "gallery_access_permission":
+            MessageLookupByLibrary.simpleMessage("Access permissions"),
+        "gallery_access_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "To start organizing your photos, we need authorization to access them"),
+        "gallery_access_reason": MessageLookupByLibrary.simpleMessage(
+            "To organize your photos we need access to your photo gallery"),
+        "garden_tag": MessageLookupByLibrary.simpleMessage("Garden"),
+        "gelato_tag": MessageLookupByLibrary.simpleMessage("Gelato"),
+        "gerbil_tag": MessageLookupByLibrary.simpleMessage("Gerbil"),
+        "get_premium_description":
+            MessageLookupByLibrary.simpleMessage("TO HAVE ALL THESE FEATURES"),
+        "get_premium_now":
+            MessageLookupByLibrary.simpleMessage("Get premium now!"),
+        "get_premium_title":
+            MessageLookupByLibrary.simpleMessage("GET PREMIUM"),
+        "glacier_tag": MessageLookupByLibrary.simpleMessage("Glacier"),
+        "glasses_tag": MessageLookupByLibrary.simpleMessage("Glasses"),
+        "glitter_tag": MessageLookupByLibrary.simpleMessage("Glitter"),
+        "goggles_tag": MessageLookupByLibrary.simpleMessage("Goggles"),
+        "gown_tag": MessageLookupByLibrary.simpleMessage("Gown"),
+        "graduation_tag": MessageLookupByLibrary.simpleMessage("Graduation"),
+        "grandparent_tag": MessageLookupByLibrary.simpleMessage("Grandparent"),
+        "groom_tag": MessageLookupByLibrary.simpleMessage("Groom"),
+        "gymnastics_tag": MessageLookupByLibrary.simpleMessage("Gymnastics"),
+        "hair_tag": MessageLookupByLibrary.simpleMessage("Hair"),
+        "hand_tag": MessageLookupByLibrary.simpleMessage("Hand"),
+        "handbag_tag": MessageLookupByLibrary.simpleMessage("Handbag"),
+        "handrail_tag": MessageLookupByLibrary.simpleMessage("Handrail"),
+        "hanukkah_tag": MessageLookupByLibrary.simpleMessage("Hanukkah"),
+        "hat_tag": MessageLookupByLibrary.simpleMessage("Hat"),
+        "heart_tag": MessageLookupByLibrary.simpleMessage("Heart"),
+        "helicopter_tag": MessageLookupByLibrary.simpleMessage("Helicopter"),
+        "helmet_tag": MessageLookupByLibrary.simpleMessage("Helmet"),
+        "herd_tag": MessageLookupByLibrary.simpleMessage("Herd"),
+        "himalayan_tag": MessageLookupByLibrary.simpleMessage("Himalayan"),
+        "home_tag": MessageLookupByLibrary.simpleMessage("Home"),
+        "horn_tag": MessageLookupByLibrary.simpleMessage("Horn"),
+        "horse_tag": MessageLookupByLibrary.simpleMessage("Horse"),
+        "hot_dog_tag": MessageLookupByLibrary.simpleMessage("Hot dog"),
+        "how_many_pics": MessageLookupByLibrary.simpleMessage("How many pics"),
+        "ice_tag": MessageLookupByLibrary.simpleMessage("Ice"),
+        "iceberg_tag": MessageLookupByLibrary.simpleMessage("Iceberg"),
+        "icicle_tag": MessageLookupByLibrary.simpleMessage("Icicle"),
+        "icing_tag": MessageLookupByLibrary.simpleMessage("Icing"),
+        "icon_tag": MessageLookupByLibrary.simpleMessage("Icon"),
+        "infinite_tags": MessageLookupByLibrary.simpleMessage("Infinite tags"),
+        "infrastructure_tag":
+            MessageLookupByLibrary.simpleMessage("Infrastructure"),
+        "insect_tag": MessageLookupByLibrary.simpleMessage("Insect"),
+        "interaction_tag": MessageLookupByLibrary.simpleMessage("Interaction"),
+        "jacket_tag": MessageLookupByLibrary.simpleMessage("Jacket"),
+        "jeans_tag": MessageLookupByLibrary.simpleMessage("Jeans"),
+        "jersey_tag": MessageLookupByLibrary.simpleMessage("Jersey"),
+        "jewellery_tag": MessageLookupByLibrary.simpleMessage("Jewellery"),
+        "joker_tag": MessageLookupByLibrary.simpleMessage("Joker"),
+        "juice_tag": MessageLookupByLibrary.simpleMessage("Juice"),
+        "jungle_tag": MessageLookupByLibrary.simpleMessage("Jungle"),
+        "junk_tag": MessageLookupByLibrary.simpleMessage("Junk"),
+        "kayak_tag": MessageLookupByLibrary.simpleMessage("Kayak"),
+        "keep_asking": MessageLookupByLibrary.simpleMessage("Keep asking"),
+        "keep_safe": MessageLookupByLibrary.simpleMessage(
+            "Your photo is now safe on picPics. Do you want to delete it from your camera roll?"),
+        "kitchen_tag": MessageLookupByLibrary.simpleMessage("Kitchen"),
+        "knitting_tag": MessageLookupByLibrary.simpleMessage("Knitting"),
+        "lake_tag": MessageLookupByLibrary.simpleMessage("Lake"),
+        "lampshade_tag": MessageLookupByLibrary.simpleMessage("Lampshade"),
+        "language": MessageLookupByLibrary.simpleMessage("Language"),
+        "larva_tag": MessageLookupByLibrary.simpleMessage("Larva"),
+        "laugh_tag": MessageLookupByLibrary.simpleMessage("Laugh"),
+        "leather_tag": MessageLookupByLibrary.simpleMessage("Leather"),
+        "leggings_tag": MessageLookupByLibrary.simpleMessage("Leggings"),
+        "lego_tag": MessageLookupByLibrary.simpleMessage("Lego"),
+        "leisure_tag": MessageLookupByLibrary.simpleMessage("Leisure"),
+        "lighthouse_tag": MessageLookupByLibrary.simpleMessage("Lighthouse"),
+        "lightning_tag": MessageLookupByLibrary.simpleMessage("Lightning"),
+        "lipstick_tag": MessageLookupByLibrary.simpleMessage("Lipstick"),
+        "lock_with_pin": MessageLookupByLibrary.simpleMessage(
+            "Lock down your private photos with a PIN passcode."),
+        "lock_your_photos":
+            MessageLookupByLibrary.simpleMessage("Lock your photos"),
+        "longboard_tag": MessageLookupByLibrary.simpleMessage("Longboard"),
+        "love_tag": MessageLookupByLibrary.simpleMessage("Love"),
+        "loveseat_tag": MessageLookupByLibrary.simpleMessage("Loveseat"),
+        "lunch_tag": MessageLookupByLibrary.simpleMessage("Lunch"),
+        "marathon_tag": MessageLookupByLibrary.simpleMessage("Marathon"),
+        "marriage_tag": MessageLookupByLibrary.simpleMessage("Marriage"),
+        "meal_tag": MessageLookupByLibrary.simpleMessage("Meal"),
+        "menu_tag": MessageLookupByLibrary.simpleMessage("Menu"),
+        "metal_tag": MessageLookupByLibrary.simpleMessage("Metal"),
+        "militia_tag": MessageLookupByLibrary.simpleMessage("Militia"),
+        "miniature_tag": MessageLookupByLibrary.simpleMessage("Miniature"),
+        "mobile_phone_tag":
+            MessageLookupByLibrary.simpleMessage("Mobile phone"),
+        "model_tag": MessageLookupByLibrary.simpleMessage("Model"),
+        "money_tag": MessageLookupByLibrary.simpleMessage("Money"),
+        "monochrome_tag": MessageLookupByLibrary.simpleMessage("Monochrome"),
+        "month": MessageLookupByLibrary.simpleMessage("month"),
+        "monument_tag": MessageLookupByLibrary.simpleMessage("Monument"),
+        "moon_tag": MessageLookupByLibrary.simpleMessage("Moon"),
+        "mortarboard_tag": MessageLookupByLibrary.simpleMessage("Mortarboard"),
+        "mosque_tag": MessageLookupByLibrary.simpleMessage("Mosque"),
+        "motorcycle_tag": MessageLookupByLibrary.simpleMessage("Motorcycle"),
+        "mountain_tag": MessageLookupByLibrary.simpleMessage("Mountain"),
+        "moustache_tag": MessageLookupByLibrary.simpleMessage("Moustache"),
+        "mouth_tag": MessageLookupByLibrary.simpleMessage("Mouth"),
+        "mufti_tag": MessageLookupByLibrary.simpleMessage("Mufti"),
+        "muscle_tag": MessageLookupByLibrary.simpleMessage("Muscle"),
+        "museum_tag": MessageLookupByLibrary.simpleMessage("Museum"),
+        "musical_instrument_tag":
+            MessageLookupByLibrary.simpleMessage("Musical instrument"),
+        "musical_tag": MessageLookupByLibrary.simpleMessage("Musical"),
+        "musician_tag": MessageLookupByLibrary.simpleMessage("Musician"),
+        "nail_tag": MessageLookupByLibrary.simpleMessage("Nail"),
+        "nebula_tag": MessageLookupByLibrary.simpleMessage("Nebula"),
+        "necklace_tag": MessageLookupByLibrary.simpleMessage("Necklace"),
+        "needlework_tag": MessageLookupByLibrary.simpleMessage("Needlework"),
+        "neon_tag": MessageLookupByLibrary.simpleMessage("Neon"),
+        "net_tag": MessageLookupByLibrary.simpleMessage("Net"),
+        "new_secret_key":
+            MessageLookupByLibrary.simpleMessage("New secret key"),
+        "news_tag": MessageLookupByLibrary.simpleMessage("News"),
+        "newspaper_tag": MessageLookupByLibrary.simpleMessage("Newspaper"),
+        "next": MessageLookupByLibrary.simpleMessage("Next"),
+        "nightclub_tag": MessageLookupByLibrary.simpleMessage("Nightclub"),
+        "no": MessageLookupByLibrary.simpleMessage("No"),
+        "no_ads": MessageLookupByLibrary.simpleMessage("No Ads"),
+        "no_previous_purchase":
+            MessageLookupByLibrary.simpleMessage("No Previous Purchase"),
+        "no_tagged_photos": MessageLookupByLibrary.simpleMessage(
+            "You don\'t have any tagged photos yet"),
+        "no_tags_found": MessageLookupByLibrary.simpleMessage("No tags found"),
+        "no_valid_subscription": MessageLookupByLibrary.simpleMessage(
+            "Could not find a valid subscription purchase."),
+        "notification_time":
+            MessageLookupByLibrary.simpleMessage("Notification time"),
+        "notifications": MessageLookupByLibrary.simpleMessage("Notifications"),
+        "ny_tag": MessageLookupByLibrary.simpleMessage("NY"),
+        "odometer_tag": MessageLookupByLibrary.simpleMessage("Odometer"),
+        "ok": MessageLookupByLibrary.simpleMessage("OK"),
+        "open_gallery": MessageLookupByLibrary.simpleMessage("Open gallery"),
+        "organized_photos_description": MessageLookupByLibrary.simpleMessage(
+            "Photos that have already been tagged"),
+        "organized_photos_title":
+            MessageLookupByLibrary.simpleMessage("Organized Photos"),
+        "outerwear_tag": MessageLookupByLibrary.simpleMessage("Outerwear"),
+        "palace_tag": MessageLookupByLibrary.simpleMessage("Palace"),
+        "paper_tag": MessageLookupByLibrary.simpleMessage("Paper"),
+        "park_tag": MessageLookupByLibrary.simpleMessage("Park"),
+        "parties_tag": MessageLookupByLibrary.simpleMessage("Parties"),
+        "party_tag": MessageLookupByLibrary.simpleMessage("Party"),
+        "passport_tag": MessageLookupByLibrary.simpleMessage("Passport"),
+        "pasteles_tag": MessageLookupByLibrary.simpleMessage("Pasteles"),
+        "pattern_tag": MessageLookupByLibrary.simpleMessage("Pattern"),
+        "penguin_tag": MessageLookupByLibrary.simpleMessage("Penguin"),
+        "person_tag": MessageLookupByLibrary.simpleMessage("Person"),
+        "pet_tag": MessageLookupByLibrary.simpleMessage("Pet"),
+        "petal_tag": MessageLookupByLibrary.simpleMessage("Petal"),
+        "pets_tag": MessageLookupByLibrary.simpleMessage("Pets"),
+        "pho_tag": MessageLookupByLibrary.simpleMessage("Pho"),
+        "photo_gallery_count": m1,
+        "photo_gallery_description":
+            MessageLookupByLibrary.simpleMessage("Photos not yet organized"),
+        "photo_gallery_title":
+            MessageLookupByLibrary.simpleMessage("Photo Gallery"),
+        "photo_location":
+            MessageLookupByLibrary.simpleMessage("Photo location"),
+        "photos_always_organized": MessageLookupByLibrary.simpleMessage(
+            "Now your photos will be always organized"),
+        "piano_tag": MessageLookupByLibrary.simpleMessage("Piano"),
+        "picnic_tag": MessageLookupByLibrary.simpleMessage("Picnic"),
+        "picpics_photo_manager":
+            MessageLookupByLibrary.simpleMessage("picPics - Photo Manager"),
+        "pie_tag": MessageLookupByLibrary.simpleMessage("Pie"),
+        "pier_tag": MessageLookupByLibrary.simpleMessage("Pier"),
+        "pillow_tag": MessageLookupByLibrary.simpleMessage("Pillow"),
+        "pitch_tag": MessageLookupByLibrary.simpleMessage("Pitch"),
+        "pixie_bob_tag": MessageLookupByLibrary.simpleMessage("Pixie-bob"),
+        "pizza_tag": MessageLookupByLibrary.simpleMessage("Pizza"),
+        "placemat_tag": MessageLookupByLibrary.simpleMessage("Placemat"),
+        "plant_tag": MessageLookupByLibrary.simpleMessage("Plant"),
+        "playground_tag": MessageLookupByLibrary.simpleMessage("Playground"),
+        "plush_tag": MessageLookupByLibrary.simpleMessage("Plush"),
+        "pocket_tag": MessageLookupByLibrary.simpleMessage("Pocket"),
+        "polo_tag": MessageLookupByLibrary.simpleMessage("Polo"),
+        "pomacentridae_tag":
+            MessageLookupByLibrary.simpleMessage("Pomacentridae"),
+        "pool_tag": MessageLookupByLibrary.simpleMessage("Pool"),
+        "pop_music_tag": MessageLookupByLibrary.simpleMessage("Pop music"),
+        "porcelain_tag": MessageLookupByLibrary.simpleMessage("Porcelain"),
+        "poster_tag": MessageLookupByLibrary.simpleMessage("Poster"),
+        "prairie_tag": MessageLookupByLibrary.simpleMessage("Prairie"),
+        "premium_modal_description": m2,
+        "premium_modal_get_premium_description":
+            MessageLookupByLibrary.simpleMessage("ALL FEATURES WITHOUT ADS"),
+        "premium_modal_get_premium_title":
+            MessageLookupByLibrary.simpleMessage("Get Premium Account"),
+        "premium_modal_watch_ad":
+            MessageLookupByLibrary.simpleMessage("Watch video ad to continue"),
+        "presentation_tag":
+            MessageLookupByLibrary.simpleMessage("Presentation"),
+        "privacy_policy":
+            MessageLookupByLibrary.simpleMessage("Privacy Policy"),
+        "private_photos":
+            MessageLookupByLibrary.simpleMessage("Private Photos"),
+        "product_tag": MessageLookupByLibrary.simpleMessage("Product"),
+        "prom_tag": MessageLookupByLibrary.simpleMessage("Prom"),
+        "protect_with_encryption": MessageLookupByLibrary.simpleMessage(
+            "Protect your personal photos with encryption only accessible with a PIN password."),
+        "quilting_tag": MessageLookupByLibrary.simpleMessage("Quilting"),
+        "race_tag": MessageLookupByLibrary.simpleMessage("Race"),
+        "racing_tag": MessageLookupByLibrary.simpleMessage("Racing"),
+        "rafting_tag": MessageLookupByLibrary.simpleMessage("Rafting"),
+        "ragdoll_tag": MessageLookupByLibrary.simpleMessage("Ragdoll"),
+        "rainbow_tag": MessageLookupByLibrary.simpleMessage("Rainbow"),
+        "ranch_tag": MessageLookupByLibrary.simpleMessage("Ranch"),
+        "rate_this_app": MessageLookupByLibrary.simpleMessage("Rate this app"),
+        "receipt_tag": MessageLookupByLibrary.simpleMessage("Receipt"),
+        "recent_tags": MessageLookupByLibrary.simpleMessage("Recent Tags"),
+        "reef_tag": MessageLookupByLibrary.simpleMessage("Reef"),
+        "rein_tag": MessageLookupByLibrary.simpleMessage("Rein"),
+        "require_secret_key":
+            MessageLookupByLibrary.simpleMessage("Require secret key"),
+        "restore_purchase":
+            MessageLookupByLibrary.simpleMessage("Restore purchase"),
+        "rickshaw_tag": MessageLookupByLibrary.simpleMessage("Rickshaw"),
+        "ring_tag": MessageLookupByLibrary.simpleMessage("Ring"),
+        "river_tag": MessageLookupByLibrary.simpleMessage("River"),
+        "road_tag": MessageLookupByLibrary.simpleMessage("Road"),
+        "rock_tag": MessageLookupByLibrary.simpleMessage("Rock"),
+        "rocket_tag": MessageLookupByLibrary.simpleMessage("Rocket"),
+        "rodeo_tag": MessageLookupByLibrary.simpleMessage("Rodeo"),
+        "roller_tag": MessageLookupByLibrary.simpleMessage("Roller"),
+        "roof_tag": MessageLookupByLibrary.simpleMessage("Roof"),
+        "room_tag": MessageLookupByLibrary.simpleMessage("Room"),
+        "rowing_tag": MessageLookupByLibrary.simpleMessage("Rowing"),
+        "rugby_tag": MessageLookupByLibrary.simpleMessage("Rugby"),
+        "ruins_tag": MessageLookupByLibrary.simpleMessage("Ruins"),
+        "running_tag": MessageLookupByLibrary.simpleMessage("Running"),
+        "safari_tag": MessageLookupByLibrary.simpleMessage("Safari"),
+        "sailboat_tag": MessageLookupByLibrary.simpleMessage("Sailboat"),
+        "sand_tag": MessageLookupByLibrary.simpleMessage("Sand"),
+        "santa_claus_tag": MessageLookupByLibrary.simpleMessage("Santa claus"),
+        "sari_tag": MessageLookupByLibrary.simpleMessage("Sari"),
+        "saucer_tag": MessageLookupByLibrary.simpleMessage("Saucer"),
+        "save": MessageLookupByLibrary.simpleMessage("save"),
+        "save_location": MessageLookupByLibrary.simpleMessage("Save location"),
+        "scarf_tag": MessageLookupByLibrary.simpleMessage("Scarf"),
+        "school_tag": MessageLookupByLibrary.simpleMessage("School"),
+        "screenshot_tag": MessageLookupByLibrary.simpleMessage("Screenshot"),
+        "screenshots_tag": MessageLookupByLibrary.simpleMessage("Screenshots"),
+        "scuba_diving_tag":
+            MessageLookupByLibrary.simpleMessage("Scuba diving"),
+        "seal_tag": MessageLookupByLibrary.simpleMessage("Seal"),
+        "search": MessageLookupByLibrary.simpleMessage("Search..."),
+        "search_all_tags_not_found": MessageLookupByLibrary.simpleMessage(
+            "No pictures were found with all tags on it."),
+        "search_results":
+            MessageLookupByLibrary.simpleMessage("Search results"),
+        "secret_key_created": MessageLookupByLibrary.simpleMessage(
+            "Secret key successfully created!"),
+        "secret_photos": MessageLookupByLibrary.simpleMessage("Secret Photos"),
+        "selfie_tag": MessageLookupByLibrary.simpleMessage("Selfie"),
+        "selfies_tag": MessageLookupByLibrary.simpleMessage("Selfies"),
+        "settings": MessageLookupByLibrary.simpleMessage("Settings"),
+        "share_with_friends":
+            MessageLookupByLibrary.simpleMessage("Share with friends"),
+        "shelf_tag": MessageLookupByLibrary.simpleMessage("Shelf"),
+        "shell_tag": MessageLookupByLibrary.simpleMessage("Shell"),
+        "shetland_sheepdog_tag":
+            MessageLookupByLibrary.simpleMessage("Shetland sheepdog"),
+        "shikoku_tag": MessageLookupByLibrary.simpleMessage("Shikoku"),
+        "shipwreck_tag": MessageLookupByLibrary.simpleMessage("Shipwreck"),
+        "shoe_tag": MessageLookupByLibrary.simpleMessage("Shoe"),
+        "shorts_tag": MessageLookupByLibrary.simpleMessage("Shorts"),
+        "sign": MessageLookupByLibrary.simpleMessage("Sign"),
+        "singer_tag": MessageLookupByLibrary.simpleMessage("Singer"),
+        "sink_tag": MessageLookupByLibrary.simpleMessage("Sink"),
+        "sitting_tag": MessageLookupByLibrary.simpleMessage("Sitting"),
+        "skateboard_tag": MessageLookupByLibrary.simpleMessage("Skateboard"),
+        "skateboarder_tag":
+            MessageLookupByLibrary.simpleMessage("Skateboarder"),
+        "skiff_tag": MessageLookupByLibrary.simpleMessage("Skiff"),
+        "skiing_tag": MessageLookupByLibrary.simpleMessage("Skiing"),
+        "skin_tag": MessageLookupByLibrary.simpleMessage("Skin"),
+        "sky_tag": MessageLookupByLibrary.simpleMessage("Sky"),
+        "skyline_tag": MessageLookupByLibrary.simpleMessage("Skyline"),
+        "skyscraper_tag": MessageLookupByLibrary.simpleMessage("Skyscraper"),
+        "sled_tag": MessageLookupByLibrary.simpleMessage("Sled"),
+        "sledding_tag": MessageLookupByLibrary.simpleMessage("Sledding"),
+        "sleep_tag": MessageLookupByLibrary.simpleMessage("Sleep"),
+        "smile_tag": MessageLookupByLibrary.simpleMessage("Smile"),
+        "sneakers_tag": MessageLookupByLibrary.simpleMessage("Sneakers"),
+        "snorkeling_tag": MessageLookupByLibrary.simpleMessage("Snorkeling"),
+        "snowboarding_tag":
+            MessageLookupByLibrary.simpleMessage("Snowboarding"),
+        "soccer_tag": MessageLookupByLibrary.simpleMessage("Soccer"),
+        "softball_tag": MessageLookupByLibrary.simpleMessage("Softball"),
+        "soil_tag": MessageLookupByLibrary.simpleMessage("Soil"),
+        "song_tag": MessageLookupByLibrary.simpleMessage("Song"),
+        "space_tag": MessageLookupByLibrary.simpleMessage("Space"),
+        "sparkler_tag": MessageLookupByLibrary.simpleMessage("Sparkler"),
+        "speedboat_tag": MessageLookupByLibrary.simpleMessage("Speedboat"),
+        "sphynx_tag": MessageLookupByLibrary.simpleMessage("Sphynx"),
+        "sports_tag": MessageLookupByLibrary.simpleMessage("Sports"),
+        "stadium_tag": MessageLookupByLibrary.simpleMessage("Stadium"),
+        "stairs_tag": MessageLookupByLibrary.simpleMessage("Stairs"),
+        "standing_tag": MessageLookupByLibrary.simpleMessage("Standing"),
+        "star_tag": MessageLookupByLibrary.simpleMessage("Star"),
+        "start": MessageLookupByLibrary.simpleMessage("Start"),
+        "start_tagging": MessageLookupByLibrary.simpleMessage("Start tagging"),
+        "statue_tag": MessageLookupByLibrary.simpleMessage("Statue"),
+        "steaming_tag": MessageLookupByLibrary.simpleMessage("Steaming"),
+        "storm_tag": MessageLookupByLibrary.simpleMessage("Storm"),
+        "strap_tag": MessageLookupByLibrary.simpleMessage("Strap"),
+        "stuffed_toy_tag": MessageLookupByLibrary.simpleMessage("Stuffed toy"),
+        "submarine_tag": MessageLookupByLibrary.simpleMessage("Submarine"),
+        "subwoofer_tag": MessageLookupByLibrary.simpleMessage("Subwoofer"),
+        "suggestions": MessageLookupByLibrary.simpleMessage("Suggestions"),
+        "sunglasses_tag": MessageLookupByLibrary.simpleMessage("Sunglasses"),
+        "sunset_tag": MessageLookupByLibrary.simpleMessage("Sunset"),
+        "superhero_tag": MessageLookupByLibrary.simpleMessage("Superhero"),
+        "superman_tag": MessageLookupByLibrary.simpleMessage("Superman"),
+        "supervillain_tag":
+            MessageLookupByLibrary.simpleMessage("Supervillain"),
+        "supper_tag": MessageLookupByLibrary.simpleMessage("Supper"),
+        "surfboard_tag": MessageLookupByLibrary.simpleMessage("Surfboard"),
+        "surfing_tag": MessageLookupByLibrary.simpleMessage("Surfing"),
+        "sushi_tag": MessageLookupByLibrary.simpleMessage("Sushi"),
+        "swamp_tag": MessageLookupByLibrary.simpleMessage("Swamp"),
+        "swimming_tag": MessageLookupByLibrary.simpleMessage("Swimming"),
+        "swimwear_tag": MessageLookupByLibrary.simpleMessage("Swimwear"),
+        "swing_tag": MessageLookupByLibrary.simpleMessage("Swing"),
+        "tablecloth_tag": MessageLookupByLibrary.simpleMessage("Tablecloth"),
+        "tableware_tag": MessageLookupByLibrary.simpleMessage("Tableware"),
+        "tag_multiple_photos_at_once":
+            MessageLookupByLibrary.simpleMessage("Tag multiple photos at once"),
+        "take_a_look":
+            MessageLookupByLibrary.simpleMessage("Take a look at this app!"),
+        "take_a_look_description": m3,
+        "tattoo_tag": MessageLookupByLibrary.simpleMessage("Tattoo"),
+        "team_tag": MessageLookupByLibrary.simpleMessage("Team"),
+        "television_tag": MessageLookupByLibrary.simpleMessage("Television"),
+        "temple_tag": MessageLookupByLibrary.simpleMessage("Temple"),
+        "terms_of_use": MessageLookupByLibrary.simpleMessage("Terms of Use"),
+        "textile_tag": MessageLookupByLibrary.simpleMessage("Textile"),
+        "thanksgiving_tag":
+            MessageLookupByLibrary.simpleMessage("Thanksgiving"),
+        "tie_tag": MessageLookupByLibrary.simpleMessage("Tie"),
+        "tights_tag": MessageLookupByLibrary.simpleMessage("Tights"),
+        "tile_tag": MessageLookupByLibrary.simpleMessage("Tile"),
+        "time": MessageLookupByLibrary.simpleMessage("Time"),
+        "tire_tag": MessageLookupByLibrary.simpleMessage("Tire"),
+        "toe_tag": MessageLookupByLibrary.simpleMessage("Toe"),
+        "toggle_date": MessageLookupByLibrary.simpleMessage("Date"),
+        "toggle_days": MessageLookupByLibrary.simpleMessage("Days"),
+        "toggle_months": MessageLookupByLibrary.simpleMessage("Months"),
+        "toggle_tags": MessageLookupByLibrary.simpleMessage("Tags"),
+        "tower_tag": MessageLookupByLibrary.simpleMessage("Tower"),
+        "toy_tag": MessageLookupByLibrary.simpleMessage("Toy"),
+        "track_tag": MessageLookupByLibrary.simpleMessage("Track"),
+        "tractor_tag": MessageLookupByLibrary.simpleMessage("Tractor"),
+        "train_tag": MessageLookupByLibrary.simpleMessage("Train"),
+        "travel_tag": MessageLookupByLibrary.simpleMessage("Travel"),
+        "trunk_tag": MessageLookupByLibrary.simpleMessage("Trunk"),
+        "tsqr_tag": MessageLookupByLibrary.simpleMessage("Times Square"),
+        "tubing_tag": MessageLookupByLibrary.simpleMessage("Tubing"),
+        "turtle_tag": MessageLookupByLibrary.simpleMessage("Turtle"),
+        "tutorial_daily_package": MessageLookupByLibrary.simpleMessage(
+            "We bring a daily package for you to gradually organize your library."),
+        "tutorial_however_you_want": MessageLookupByLibrary.simpleMessage(
+            "Organize your photos by adding tags, like \"family\", \"pets\", or whatever you want."),
+        "tutorial_just_swipe": MessageLookupByLibrary.simpleMessage(
+            "After adding the tags to your photo, just swipe to go to the next one."),
+        "tutorial_multiselect": MessageLookupByLibrary.simpleMessage(
+            "You can \"tap & hold\" on your photos to tag multiple photos at once."),
+        "tutorial_secret": MessageLookupByLibrary.simpleMessage(
+            "Hide your private photos with a pin code protection, keeping them safe."),
+        "tuxedo_tag": MessageLookupByLibrary.simpleMessage("Tuxedo"),
+        "twig_tag": MessageLookupByLibrary.simpleMessage("Twig"),
+        "umbrella_tag": MessageLookupByLibrary.simpleMessage("Umbrella"),
+        "underwater_tag": MessageLookupByLibrary.simpleMessage("Underwater"),
+        "unicycle_tag": MessageLookupByLibrary.simpleMessage("Unicycle"),
+        "unlimited_private_pics":
+            MessageLookupByLibrary.simpleMessage("Unlimited private photos"),
+        "vacation_tag": MessageLookupByLibrary.simpleMessage("Vacation"),
+        "van_tag": MessageLookupByLibrary.simpleMessage("Van"),
+        "vegetable_tag": MessageLookupByLibrary.simpleMessage("Vegetable"),
+        "vehicle_tag": MessageLookupByLibrary.simpleMessage("Vehicle"),
+        "veil_tag": MessageLookupByLibrary.simpleMessage("Veil"),
+        "view_hidden_photos": MessageLookupByLibrary.simpleMessage(
+            "To view your hidden photos, release the lock in the application settings. You can unlock them with your PIN."),
+        "volcano_tag": MessageLookupByLibrary.simpleMessage("Volcano"),
+        "wakeboarding_tag":
+            MessageLookupByLibrary.simpleMessage("Wakeboarding"),
+        "wall_tag": MessageLookupByLibrary.simpleMessage("Wall"),
+        "watercolor_paint_tag":
+            MessageLookupByLibrary.simpleMessage("Watercolor paint"),
+        "waterfall_tag": MessageLookupByLibrary.simpleMessage("Waterfall"),
+        "waterfowl_tag": MessageLookupByLibrary.simpleMessage("Waterfowl"),
+        "waterskiing_tag": MessageLookupByLibrary.simpleMessage("Waterskiing"),
+        "web_page_tag": MessageLookupByLibrary.simpleMessage("Web page"),
+        "welcome": MessageLookupByLibrary.simpleMessage("Welcome!"),
+        "wetsuit_tag": MessageLookupByLibrary.simpleMessage("Wetsuit"),
+        "wheel_tag": MessageLookupByLibrary.simpleMessage("Wheel"),
+        "wheelbarrow_tag": MessageLookupByLibrary.simpleMessage("Wheelbarrow"),
+        "whiteboard_tag": MessageLookupByLibrary.simpleMessage("Whiteboard"),
+        "windshield_tag": MessageLookupByLibrary.simpleMessage("Windshield"),
+        "windsurfing_tag": MessageLookupByLibrary.simpleMessage("Windsurfing"),
+        "wine_tag": MessageLookupByLibrary.simpleMessage("Wine"),
+        "wing_tag": MessageLookupByLibrary.simpleMessage("Wing"),
+        "wool_tag": MessageLookupByLibrary.simpleMessage("Wool"),
+        "work_tag": MessageLookupByLibrary.simpleMessage("Work"),
+        "wreath_tag": MessageLookupByLibrary.simpleMessage("Wreath"),
+        "x_minutes": MessageLookupByLibrary.simpleMessage("20 min"),
+        "year": MessageLookupByLibrary.simpleMessage("year"),
+        "yes": MessageLookupByLibrary.simpleMessage("Yes"),
+        "you_are_premium":
+            MessageLookupByLibrary.simpleMessage("You are premium!"),
+        "your_secret_key":
+            MessageLookupByLibrary.simpleMessage("Your secret key")
+      };
 }
