@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -19,569 +19,681 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'es';
 
-  static m0(email) => "Se envió una clave de acceso a ${email}";
+  static String m0(email) => "Se envió una clave de acceso a ${email}";
 
-  static m1(howMany) => "${Intl.plural(howMany, zero: 'No hay ninguna foto seleccionada', one: '1 foto seleccionada', other: '${howMany} fotos seleccionadas')}";
+  static String m1(howMany) =>
+      "${Intl.plural(howMany, zero: 'No hay ninguna foto seleccionada', one: '1 foto seleccionada', other: '${howMany} fotos seleccionadas')}";
 
-  static m2(number) => "Has completado tu ${number} de fotos diarias gratis, ¿quieres continuar?";
+  static String m2(number) =>
+      "Has completado tu ${number} de fotos diarias gratis, ¿quieres continuar?";
 
-  static m3(url) => "Para organizar todas tus fotos, ve a ${url}";
+  static String m3(url) => "Para organizar todas tus fotos, ve a ${url}";
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "access_code" : MessageLookupByLibrary.simpleMessage("Código de acceso"),
-    "access_code_sent" : m0,
-    "add_multiple_tags" : MessageLookupByLibrary.simpleMessage("Añade múltiples etiquetas"),
-    "add_tag" : MessageLookupByLibrary.simpleMessage("Añadir etiqueta"),
-    "add_tags" : MessageLookupByLibrary.simpleMessage("Añadir etiquetas"),
-    "aerospace_engineering_tag" : MessageLookupByLibrary.simpleMessage("Ingeniería Aeroespacial"),
-    "aircraft_tag" : MessageLookupByLibrary.simpleMessage("Aeronave"),
-    "airliner_tag" : MessageLookupByLibrary.simpleMessage("Avión de línea"),
-    "airplane_tag" : MessageLookupByLibrary.simpleMessage("Avión"),
-    "alcohol_tag" : MessageLookupByLibrary.simpleMessage("Alcohol"),
-    "all_at_once" : MessageLookupByLibrary.simpleMessage("Organiza varias fotos a la vez"),
-    "all_photos_were_tagged" : MessageLookupByLibrary.simpleMessage("There are no more photos to organize."),
-    "all_search_tags" : MessageLookupByLibrary.simpleMessage("Todas las etiquetas de búsqueda"),
-    "always" : MessageLookupByLibrary.simpleMessage("Siempre"),
-    "aquarium_tag" : MessageLookupByLibrary.simpleMessage("Acuario"),
-    "archery_tag" : MessageLookupByLibrary.simpleMessage("Tiro al arco"),
-    "armrest_tag" : MessageLookupByLibrary.simpleMessage("Brazo"),
-    "army_tag" : MessageLookupByLibrary.simpleMessage("Ejército"),
-    "ask_photo_library_permission" : MessageLookupByLibrary.simpleMessage("Necesitamos acceso a su biblioteca de fotos para que pueda comenzar a organizar sus fotos con picPics. ¡No se preocupe, sus datos nunca saldrán de su dispositivo!"),
-    "asphalt_tag" : MessageLookupByLibrary.simpleMessage("Asfalto"),
-    "aurora_tag" : MessageLookupByLibrary.simpleMessage("Aurora"),
-    "auto_renewable_first_part" : MessageLookupByLibrary.simpleMessage("La suscripción se "),
-    "auto_renewable_second_part" : MessageLookupByLibrary.simpleMessage("renueva automáticamente."),
-    "aviation_tag" : MessageLookupByLibrary.simpleMessage("Aviación"),
-    "baby_tag" : MessageLookupByLibrary.simpleMessage("Bebé"),
-    "backpacking_tag" : MessageLookupByLibrary.simpleMessage("Mochilero"),
-    "badminton_tag" : MessageLookupByLibrary.simpleMessage("Bádminton"),
-    "bag_tag" : MessageLookupByLibrary.simpleMessage("Bolso"),
-    "balance_tag" : MessageLookupByLibrary.simpleMessage("Equilibrar"),
-    "balloon_tag" : MessageLookupByLibrary.simpleMessage("Globo"),
-    "ballroom_tag" : MessageLookupByLibrary.simpleMessage("Salón de baile"),
-    "bangle_tag" : MessageLookupByLibrary.simpleMessage("Brazalete"),
-    "bangs_tag" : MessageLookupByLibrary.simpleMessage("Golpes"),
-    "bar_tag" : MessageLookupByLibrary.simpleMessage("Bar"),
-    "barn_tag" : MessageLookupByLibrary.simpleMessage("Granero"),
-    "basset_hound_tag" : MessageLookupByLibrary.simpleMessage("Perro basset"),
-    "bathing_tag" : MessageLookupByLibrary.simpleMessage("Baños"),
-    "bathroom_tag" : MessageLookupByLibrary.simpleMessage("Baño"),
-    "beach_tag" : MessageLookupByLibrary.simpleMessage("Playa"),
-    "beanie_tag" : MessageLookupByLibrary.simpleMessage("Gorro"),
-    "bear_tag" : MessageLookupByLibrary.simpleMessage("Oso"),
-    "beard_tag" : MessageLookupByLibrary.simpleMessage("Barba"),
-    "bedroom_tag" : MessageLookupByLibrary.simpleMessage("Habitación"),
-    "bench_tag" : MessageLookupByLibrary.simpleMessage("Banco"),
-    "bento_tag" : MessageLookupByLibrary.simpleMessage("Bento"),
-    "bicycle_tag" : MessageLookupByLibrary.simpleMessage("Bicicleta"),
-    "bird_tag" : MessageLookupByLibrary.simpleMessage("Pájaro"),
-    "blazer_tag" : MessageLookupByLibrary.simpleMessage("Chaqueta de sport"),
-    "boat_tag" : MessageLookupByLibrary.simpleMessage("Barco"),
-    "bonfire_tag" : MessageLookupByLibrary.simpleMessage("Hoguera"),
-    "boxer_tag" : MessageLookupByLibrary.simpleMessage("Boxer"),
-    "bracelet_tag" : MessageLookupByLibrary.simpleMessage("Pulsera"),
-    "branch_tag" : MessageLookupByLibrary.simpleMessage("Rama"),
-    "bread_tag" : MessageLookupByLibrary.simpleMessage("Pan"),
-    "brick_tag" : MessageLookupByLibrary.simpleMessage("Ladrillo"),
-    "bride_tag" : MessageLookupByLibrary.simpleMessage("Novia"),
-    "bridge_tag" : MessageLookupByLibrary.simpleMessage("Puente"),
-    "bridle_tag" : MessageLookupByLibrary.simpleMessage("Brida"),
-    "brig_tag" : MessageLookupByLibrary.simpleMessage("Bergantín"),
-    "building_tag" : MessageLookupByLibrary.simpleMessage("Edificio"),
-    "bull_tag" : MessageLookupByLibrary.simpleMessage("Toro"),
-    "bullfighting_tag" : MessageLookupByLibrary.simpleMessage("Toreo"),
-    "bumper_tag" : MessageLookupByLibrary.simpleMessage("Parachoque"),
-    "bunk_bed_tag" : MessageLookupByLibrary.simpleMessage("Litera"),
-    "bus_tag" : MessageLookupByLibrary.simpleMessage("Autobús"),
-    "butterfly_tag" : MessageLookupByLibrary.simpleMessage("Mariposa"),
-    "cabinetry_tag" : MessageLookupByLibrary.simpleMessage("Ebanistería"),
-    "cage_tag" : MessageLookupByLibrary.simpleMessage("Jaula"),
-    "cairn_terrier_tag" : MessageLookupByLibrary.simpleMessage("Cairn terrier"),
-    "cake_tag" : MessageLookupByLibrary.simpleMessage("Pastel"),
-    "camping_tag" : MessageLookupByLibrary.simpleMessage("Cámping"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Cancelar"),
-    "cancel_anytime" : MessageLookupByLibrary.simpleMessage("Puedes cancelar en cualquier momento"),
-    "canoe_tag" : MessageLookupByLibrary.simpleMessage("Canoa"),
-    "canyon_tag" : MessageLookupByLibrary.simpleMessage("Cañón"),
-    "cap_tag" : MessageLookupByLibrary.simpleMessage("Gorra"),
-    "cappuccino_tag" : MessageLookupByLibrary.simpleMessage("Capuchino"),
-    "car_tag" : MessageLookupByLibrary.simpleMessage("Coche"),
-    "carnival_tag" : MessageLookupByLibrary.simpleMessage("Carnaval"),
-    "casino_tag" : MessageLookupByLibrary.simpleMessage("Casino"),
-    "castle_tag" : MessageLookupByLibrary.simpleMessage("Castillo"),
-    "cat_tag" : MessageLookupByLibrary.simpleMessage("Gato"),
-    "cathedral_tag" : MessageLookupByLibrary.simpleMessage("Catedral"),
-    "cattle_tag" : MessageLookupByLibrary.simpleMessage("Vacas"),
-    "cavalier_tag" : MessageLookupByLibrary.simpleMessage("Caballero"),
-    "cave_tag" : MessageLookupByLibrary.simpleMessage("Cueva"),
-    "caving_tag" : MessageLookupByLibrary.simpleMessage("Espeleología"),
-    "centrepiece_tag" : MessageLookupByLibrary.simpleMessage("Habitación central"),
-    "chain_tag" : MessageLookupByLibrary.simpleMessage("Cadena"),
-    "chair_tag" : MessageLookupByLibrary.simpleMessage("Silla"),
-    "cheeseburger_tag" : MessageLookupByLibrary.simpleMessage("Hamburguesa con queso"),
-    "christmas_tag" : MessageLookupByLibrary.simpleMessage("Navidad"),
-    "church_tag" : MessageLookupByLibrary.simpleMessage("Iglesia"),
-    "circus_tag" : MessageLookupByLibrary.simpleMessage("Circo"),
-    "class_tag" : MessageLookupByLibrary.simpleMessage("Clase"),
-    "cliff_tag" : MessageLookupByLibrary.simpleMessage("Acantilado"),
-    "clipper_tag" : MessageLookupByLibrary.simpleMessage("Clíper"),
-    "clock_tag" : MessageLookupByLibrary.simpleMessage("Reloj"),
-    "close" : MessageLookupByLibrary.simpleMessage("Cerrar"),
-    "clown_tag" : MessageLookupByLibrary.simpleMessage("Payaso"),
-    "coffee_tag" : MessageLookupByLibrary.simpleMessage("Café"),
-    "cola_tag" : MessageLookupByLibrary.simpleMessage("Reajuste salarial"),
-    "comet_tag" : MessageLookupByLibrary.simpleMessage("Cometa"),
-    "comics_tag" : MessageLookupByLibrary.simpleMessage("Historietas"),
-    "community_tag" : MessageLookupByLibrary.simpleMessage("Comunidad"),
-    "computer_tag" : MessageLookupByLibrary.simpleMessage("Computadora"),
-    "concert_tag" : MessageLookupByLibrary.simpleMessage("Concierto"),
-    "confirm_email" : MessageLookupByLibrary.simpleMessage("Confirma tu correo electrónico de registro para recibir tu código de acceso"),
-    "confirm_secret_key" : MessageLookupByLibrary.simpleMessage("Confirmar clave secreta"),
-    "construction_tag" : MessageLookupByLibrary.simpleMessage("Construcción"),
-    "continue_string" : MessageLookupByLibrary.simpleMessage("Continuar"),
-    "cookie_tag" : MessageLookupByLibrary.simpleMessage("Galleta"),
-    "cookware_and_bakeware_tag" : MessageLookupByLibrary.simpleMessage("Utensilios de cocina y para hornear"),
-    "cool_tag" : MessageLookupByLibrary.simpleMessage("Frio"),
-    "cotton_tag" : MessageLookupByLibrary.simpleMessage("Algodón"),
-    "couch_tag" : MessageLookupByLibrary.simpleMessage("Sofá"),
-    "countertop_tag" : MessageLookupByLibrary.simpleMessage("Encimera"),
-    "country" : MessageLookupByLibrary.simpleMessage("país"),
-    "couscous_tag" : MessageLookupByLibrary.simpleMessage("Cuscús"),
-    "crew_tag" : MessageLookupByLibrary.simpleMessage("Tripulación"),
-    "crochet_tag" : MessageLookupByLibrary.simpleMessage("Tejer"),
-    "crocodile_tag" : MessageLookupByLibrary.simpleMessage("Cocodrilo"),
-    "crowd_tag" : MessageLookupByLibrary.simpleMessage("Multitud"),
-    "cuisine_tag" : MessageLookupByLibrary.simpleMessage("Cocina"),
-    "cup_tag" : MessageLookupByLibrary.simpleMessage("Taza"),
-    "curling_tag" : MessageLookupByLibrary.simpleMessage("Curling"),
-    "curtain_tag" : MessageLookupByLibrary.simpleMessage("Cortina"),
-    "cushion_tag" : MessageLookupByLibrary.simpleMessage("Amortiguar"),
-    "cutlery_tag" : MessageLookupByLibrary.simpleMessage("Cuchillería"),
-    "cycling_tag" : MessageLookupByLibrary.simpleMessage("Ciclismo"),
-    "daily_challenge" : MessageLookupByLibrary.simpleMessage("Desafío diario"),
-    "daily_challenge_permission_description" : MessageLookupByLibrary.simpleMessage("Para que podamos enviar tus desafíos diarios, necesitamos autorización para enviar notificaciones, por lo tanto, es necesario que autorices las notificaciones en las opciones de tu teléfono celular"),
-    "daily_goal" : MessageLookupByLibrary.simpleMessage("Objetivo diario"),
-    "daily_notification_description" : MessageLookupByLibrary.simpleMessage("It\'s time to complete your picPics daily challenge!"),
-    "daily_notification_title" : MessageLookupByLibrary.simpleMessage("Daily challenge"),
-    "dalmatian_tag" : MessageLookupByLibrary.simpleMessage("Dálmata"),
-    "dam_tag" : MessageLookupByLibrary.simpleMessage("Represa"),
-    "dance_tag" : MessageLookupByLibrary.simpleMessage("Danza"),
-    "deejay_tag" : MessageLookupByLibrary.simpleMessage("Deejay"),
-    "delete" : MessageLookupByLibrary.simpleMessage("Eliminar"),
-    "denim_tag" : MessageLookupByLibrary.simpleMessage("Mezclilla"),
-    "desert_tag" : MessageLookupByLibrary.simpleMessage("Desierto"),
-    "desk_tag" : MessageLookupByLibrary.simpleMessage("Escritorio"),
-    "device_has_no_pics" : MessageLookupByLibrary.simpleMessage("Este dispositivo no tiene ninguna foto en la galería, por lo que no hay fotos que se puedan etiquetar."),
-    "dinosaur_tag" : MessageLookupByLibrary.simpleMessage("Dinosaurio"),
-    "disable_secret" : MessageLookupByLibrary.simpleMessage("¿Quieres mostrar esta foto?"),
-    "dog_tag" : MessageLookupByLibrary.simpleMessage("Perro"),
-    "doily_tag" : MessageLookupByLibrary.simpleMessage("Mantelito"),
-    "dont_ask_again" : MessageLookupByLibrary.simpleMessage("No volver a preguntar"),
-    "dragon_tag" : MessageLookupByLibrary.simpleMessage("Continuar"),
-    "drawer_tag" : MessageLookupByLibrary.simpleMessage("Cajón"),
-    "dress_tag" : MessageLookupByLibrary.simpleMessage("Vestido"),
-    "duck_tag" : MessageLookupByLibrary.simpleMessage("Pato"),
-    "dude_tag" : MessageLookupByLibrary.simpleMessage("Tipo"),
-    "dune_tag" : MessageLookupByLibrary.simpleMessage("Duna"),
-    "ear_tag" : MessageLookupByLibrary.simpleMessage("Oreja"),
-    "eating_tag" : MessageLookupByLibrary.simpleMessage("Comiendo"),
-    "edit_tag" : MessageLookupByLibrary.simpleMessage("Editar etiqueta"),
-    "email" : MessageLookupByLibrary.simpleMessage("Correo electrónico"),
-    "enable_faceid" : MessageLookupByLibrary.simpleMessage("Habilitar Face ID"),
-    "enable_fingerprint" : MessageLookupByLibrary.simpleMessage("Habilitar Huella Digital"),
-    "enable_irisscanner" : MessageLookupByLibrary.simpleMessage("habilitar Escáner de Iris"),
-    "enable_touchid" : MessageLookupByLibrary.simpleMessage("Habilitar Touch ID"),
-    "event_tag" : MessageLookupByLibrary.simpleMessage("Evento"),
-    "export_all_gallery" : MessageLookupByLibrary.simpleMessage("Exportar toda la galería"),
-    "export_library" : MessageLookupByLibrary.simpleMessage("Exportar biblioteca"),
-    "eyelash_tag" : MessageLookupByLibrary.simpleMessage("Pestaña"),
-    "factory_tag" : MessageLookupByLibrary.simpleMessage("Fábrica"),
-    "family_tag" : MessageLookupByLibrary.simpleMessage("Familia"),
-    "farm_tag" : MessageLookupByLibrary.simpleMessage("Granja"),
-    "fast_food_tag" : MessageLookupByLibrary.simpleMessage("Comida rápida"),
-    "feedback_bug_report" : MessageLookupByLibrary.simpleMessage("Informe de errores y comentarios"),
-    "ferris_wheel_tag" : MessageLookupByLibrary.simpleMessage("Rueda de la fortuna"),
-    "fiction_tag" : MessageLookupByLibrary.simpleMessage("Ficción"),
-    "field_tag" : MessageLookupByLibrary.simpleMessage("Campo"),
-    "find_easily" : MessageLookupByLibrary.simpleMessage("Encuentra tus fotos fácilmente"),
-    "fire_tag" : MessageLookupByLibrary.simpleMessage("Fuego"),
-    "fireworks_tag" : MessageLookupByLibrary.simpleMessage("Fuegos artificiales"),
-    "fishing_tag" : MessageLookupByLibrary.simpleMessage("Pescar"),
-    "flag_tag" : MessageLookupByLibrary.simpleMessage("Bandera"),
-    "flap_tag" : MessageLookupByLibrary.simpleMessage("Solapa"),
-    "flesh_tag" : MessageLookupByLibrary.simpleMessage("Carne"),
-    "flora_tag" : MessageLookupByLibrary.simpleMessage("Flora"),
-    "flower_tag" : MessageLookupByLibrary.simpleMessage("Flor"),
-    "flowerpot_tag" : MessageLookupByLibrary.simpleMessage("Maceta"),
-    "fog_tag" : MessageLookupByLibrary.simpleMessage("Niebla"),
-    "food_tag" : MessageLookupByLibrary.simpleMessage("Comida"),
-    "foods_tag" : MessageLookupByLibrary.simpleMessage("Alimentos"),
-    "foot_tag" : MessageLookupByLibrary.simpleMessage("Pie"),
-    "forest_tag" : MessageLookupByLibrary.simpleMessage("Bosque"),
-    "forgot_secret_key" : MessageLookupByLibrary.simpleMessage("¿Olvidaste la clave secreta?"),
-    "frigate_tag" : MessageLookupByLibrary.simpleMessage("Fragata"),
-    "fruit_tag" : MessageLookupByLibrary.simpleMessage("Fruta"),
-    "full_screen" : MessageLookupByLibrary.simpleMessage("Pantalla completa"),
-    "fun_tag" : MessageLookupByLibrary.simpleMessage("Divertido"),
-    "fur_tag" : MessageLookupByLibrary.simpleMessage("Pelaje"),
-    "gallery_access_permission" : MessageLookupByLibrary.simpleMessage("Permisos de acceso"),
-    "gallery_access_permission_description" : MessageLookupByLibrary.simpleMessage("Para comenzar a organizar tus fotos, necesitamos autorización para acceder a ellas."),
-    "gallery_access_reason" : MessageLookupByLibrary.simpleMessage("Para organizar tus fotos, necesitamos acceder a tu galería de fotos."),
-    "garden_tag" : MessageLookupByLibrary.simpleMessage("Jardín"),
-    "gelato_tag" : MessageLookupByLibrary.simpleMessage("Helado"),
-    "gerbil_tag" : MessageLookupByLibrary.simpleMessage("Jerbo"),
-    "get_premium_description" : MessageLookupByLibrary.simpleMessage("PARA TENER TODAS ESTAS FUNCIONES"),
-    "get_premium_now" : MessageLookupByLibrary.simpleMessage("¡Hazte Premium ahora!"),
-    "get_premium_title" : MessageLookupByLibrary.simpleMessage("HAZTE PREMIUM"),
-    "glacier_tag" : MessageLookupByLibrary.simpleMessage("Glaciar"),
-    "glasses_tag" : MessageLookupByLibrary.simpleMessage("Gafas"),
-    "glitter_tag" : MessageLookupByLibrary.simpleMessage("Brillantina"),
-    "goggles_tag" : MessageLookupByLibrary.simpleMessage("Gafas de protección"),
-    "gown_tag" : MessageLookupByLibrary.simpleMessage("Vestido"),
-    "graduation_tag" : MessageLookupByLibrary.simpleMessage("Graduación"),
-    "grandparent_tag" : MessageLookupByLibrary.simpleMessage("Abuelo"),
-    "groom_tag" : MessageLookupByLibrary.simpleMessage("Novio"),
-    "gymnastics_tag" : MessageLookupByLibrary.simpleMessage("Gimnasia"),
-    "hair_tag" : MessageLookupByLibrary.simpleMessage("Pelo"),
-    "hand_tag" : MessageLookupByLibrary.simpleMessage("Mano"),
-    "handbag_tag" : MessageLookupByLibrary.simpleMessage("Bolso"),
-    "handrail_tag" : MessageLookupByLibrary.simpleMessage("Pretil"),
-    "hanukkah_tag" : MessageLookupByLibrary.simpleMessage("Hanukkah"),
-    "hat_tag" : MessageLookupByLibrary.simpleMessage("Sombrero"),
-    "heart_tag" : MessageLookupByLibrary.simpleMessage("Corazón"),
-    "helicopter_tag" : MessageLookupByLibrary.simpleMessage("Helicóptero"),
-    "helmet_tag" : MessageLookupByLibrary.simpleMessage("Casco"),
-    "herd_tag" : MessageLookupByLibrary.simpleMessage("Manada"),
-    "himalayan_tag" : MessageLookupByLibrary.simpleMessage("Himalaya"),
-    "home_tag" : MessageLookupByLibrary.simpleMessage("Casa"),
-    "horn_tag" : MessageLookupByLibrary.simpleMessage("Cuerno"),
-    "horse_tag" : MessageLookupByLibrary.simpleMessage("Caballo"),
-    "hot_dog_tag" : MessageLookupByLibrary.simpleMessage("Pancho"),
-    "how_many_pics" : MessageLookupByLibrary.simpleMessage("Cuántas fotos"),
-    "ice_tag" : MessageLookupByLibrary.simpleMessage("Hielo"),
-    "iceberg_tag" : MessageLookupByLibrary.simpleMessage("Iceberg"),
-    "icicle_tag" : MessageLookupByLibrary.simpleMessage("Carámbano"),
-    "icing_tag" : MessageLookupByLibrary.simpleMessage("Formación de hielo"),
-    "icon_tag" : MessageLookupByLibrary.simpleMessage("Icono"),
-    "infinite_tags" : MessageLookupByLibrary.simpleMessage("Etiquetas infinitas"),
-    "infrastructure_tag" : MessageLookupByLibrary.simpleMessage("Infraestructura"),
-    "insect_tag" : MessageLookupByLibrary.simpleMessage("Insecto"),
-    "jacket_tag" : MessageLookupByLibrary.simpleMessage("Chaqueta"),
-    "jeans_tag" : MessageLookupByLibrary.simpleMessage("Pantalones"),
-    "jersey_tag" : MessageLookupByLibrary.simpleMessage("Jersey"),
-    "jewellery_tag" : MessageLookupByLibrary.simpleMessage("Joyería"),
-    "joker_tag" : MessageLookupByLibrary.simpleMessage("Bufón"),
-    "juice_tag" : MessageLookupByLibrary.simpleMessage("Jugo"),
-    "jungle_tag" : MessageLookupByLibrary.simpleMessage("Selva"),
-    "junk_tag" : MessageLookupByLibrary.simpleMessage("Basura"),
-    "kayak_tag" : MessageLookupByLibrary.simpleMessage("Kayac"),
-    "keep_asking" : MessageLookupByLibrary.simpleMessage("Seguir preguntando"),
-    "keep_safe" : MessageLookupByLibrary.simpleMessage("Tu foto ahora está segura en picPics. ¿Quieres eliminarla del carrete de tu cámara?"),
-    "kitchen_tag" : MessageLookupByLibrary.simpleMessage("Cocina"),
-    "knitting_tag" : MessageLookupByLibrary.simpleMessage("Tejido de punto"),
-    "lake_tag" : MessageLookupByLibrary.simpleMessage("Lago"),
-    "lampshade_tag" : MessageLookupByLibrary.simpleMessage("Pantalla de lámpara"),
-    "language" : MessageLookupByLibrary.simpleMessage("Idioma"),
-    "larva_tag" : MessageLookupByLibrary.simpleMessage("Larva"),
-    "laugh_tag" : MessageLookupByLibrary.simpleMessage("Risa"),
-    "leather_tag" : MessageLookupByLibrary.simpleMessage("Cuero"),
-    "leggings_tag" : MessageLookupByLibrary.simpleMessage("Polainas"),
-    "lego_tag" : MessageLookupByLibrary.simpleMessage("Lego"),
-    "leisure_tag" : MessageLookupByLibrary.simpleMessage("Ocio"),
-    "lighthouse_tag" : MessageLookupByLibrary.simpleMessage("Faro"),
-    "lightning_tag" : MessageLookupByLibrary.simpleMessage("Relámpago"),
-    "lipstick_tag" : MessageLookupByLibrary.simpleMessage("Lápiz labial"),
-    "lock_with_pin" : MessageLookupByLibrary.simpleMessage("Bloquea tus fotos privadas con un código PIN."),
-    "lock_your_photos" : MessageLookupByLibrary.simpleMessage("Bloquea tus fotos"),
-    "longboard_tag" : MessageLookupByLibrary.simpleMessage("Longboard"),
-    "love_tag" : MessageLookupByLibrary.simpleMessage("Amor"),
-    "loveseat_tag" : MessageLookupByLibrary.simpleMessage("Sofá de dos plazas"),
-    "lunch_tag" : MessageLookupByLibrary.simpleMessage("Almuerzo"),
-    "marathon_tag" : MessageLookupByLibrary.simpleMessage("Maratón"),
-    "meal_tag" : MessageLookupByLibrary.simpleMessage("Comida"),
-    "menu_tag" : MessageLookupByLibrary.simpleMessage("Menú"),
-    "metal_tag" : MessageLookupByLibrary.simpleMessage("Metal"),
-    "militia_tag" : MessageLookupByLibrary.simpleMessage("Milicia"),
-    "miniature_tag" : MessageLookupByLibrary.simpleMessage("Miniatura"),
-    "mobile_phone_tag" : MessageLookupByLibrary.simpleMessage("Teléfono móvil"),
-    "model_tag" : MessageLookupByLibrary.simpleMessage("Modelo"),
-    "money_tag" : MessageLookupByLibrary.simpleMessage("Dinero"),
-    "monochrome_tag" : MessageLookupByLibrary.simpleMessage("Monocromo"),
-    "month" : MessageLookupByLibrary.simpleMessage("mes"),
-    "monument_tag" : MessageLookupByLibrary.simpleMessage("Monumento"),
-    "moon_tag" : MessageLookupByLibrary.simpleMessage("Luna"),
-    "mortarboard_tag" : MessageLookupByLibrary.simpleMessage("Birrete"),
-    "mosque_tag" : MessageLookupByLibrary.simpleMessage("Mezquita"),
-    "motorcycle_tag" : MessageLookupByLibrary.simpleMessage("Motocicleta"),
-    "mountain_tag" : MessageLookupByLibrary.simpleMessage("Montaña"),
-    "moustache_tag" : MessageLookupByLibrary.simpleMessage("Bigote"),
-    "mouth_tag" : MessageLookupByLibrary.simpleMessage("Boca"),
-    "mufti_tag" : MessageLookupByLibrary.simpleMessage("Muftí"),
-    "muscle_tag" : MessageLookupByLibrary.simpleMessage("Músculo"),
-    "museum_tag" : MessageLookupByLibrary.simpleMessage("Museo"),
-    "musical_instrument_tag" : MessageLookupByLibrary.simpleMessage("Instrumento musical"),
-    "musical_tag" : MessageLookupByLibrary.simpleMessage("Musical"),
-    "musician_tag" : MessageLookupByLibrary.simpleMessage("Músico"),
-    "nail_tag" : MessageLookupByLibrary.simpleMessage("Uña"),
-    "nebula_tag" : MessageLookupByLibrary.simpleMessage("Nebulosa"),
-    "necklace_tag" : MessageLookupByLibrary.simpleMessage("Collar"),
-    "neon_tag" : MessageLookupByLibrary.simpleMessage("Neón"),
-    "net_tag" : MessageLookupByLibrary.simpleMessage("Red"),
-    "new_secret_key" : MessageLookupByLibrary.simpleMessage("Nueva clave secreta"),
-    "news_tag" : MessageLookupByLibrary.simpleMessage("Noticias"),
-    "newspaper_tag" : MessageLookupByLibrary.simpleMessage("Periódico"),
-    "next" : MessageLookupByLibrary.simpleMessage("Siguiente"),
-    "nightclub_tag" : MessageLookupByLibrary.simpleMessage("Club nocturno"),
-    "no" : MessageLookupByLibrary.simpleMessage("No"),
-    "no_ads" : MessageLookupByLibrary.simpleMessage("Sin anuncios"),
-    "no_previous_purchase" : MessageLookupByLibrary.simpleMessage("No hay ninguna compra previa"),
-    "no_tagged_photos" : MessageLookupByLibrary.simpleMessage("Aún no tienes ninguna foto etiquetada"),
-    "no_tags_found" : MessageLookupByLibrary.simpleMessage("No se encontró ninguna etiqueta"),
-    "no_valid_subscription" : MessageLookupByLibrary.simpleMessage("No se pudo encontrar ninguna compra de suscripción válida."),
-    "notification_time" : MessageLookupByLibrary.simpleMessage("Hora de la notificación"),
-    "notifications" : MessageLookupByLibrary.simpleMessage("Notificaciones"),
-    "ny_tag" : MessageLookupByLibrary.simpleMessage("Nueva York"),
-    "odometer_tag" : MessageLookupByLibrary.simpleMessage("Cuentakilómetros"),
-    "ok" : MessageLookupByLibrary.simpleMessage("Aceptar"),
-    "open_gallery" : MessageLookupByLibrary.simpleMessage("Abrir galería"),
-    "organized_photos_description" : MessageLookupByLibrary.simpleMessage("Fotos que ya se han etiquetado"),
-    "organized_photos_title" : MessageLookupByLibrary.simpleMessage("Fotos organizadas"),
-    "outerwear_tag" : MessageLookupByLibrary.simpleMessage("Ropa de calle"),
-    "palace_tag" : MessageLookupByLibrary.simpleMessage("Palacio"),
-    "paper_tag" : MessageLookupByLibrary.simpleMessage("Papel"),
-    "park_tag" : MessageLookupByLibrary.simpleMessage("Parque"),
-    "parties_tag" : MessageLookupByLibrary.simpleMessage("Fiestas"),
-    "party_tag" : MessageLookupByLibrary.simpleMessage("Fiesta"),
-    "passport_tag" : MessageLookupByLibrary.simpleMessage("Pasaporte"),
-    "pasteles_tag" : MessageLookupByLibrary.simpleMessage("Pasteles"),
-    "pattern_tag" : MessageLookupByLibrary.simpleMessage("Patrón"),
-    "penguin_tag" : MessageLookupByLibrary.simpleMessage("Pingüino"),
-    "person_tag" : MessageLookupByLibrary.simpleMessage("Persona"),
-    "pet_tag" : MessageLookupByLibrary.simpleMessage("Mascota"),
-    "petal_tag" : MessageLookupByLibrary.simpleMessage("Pétalo"),
-    "pets_tag" : MessageLookupByLibrary.simpleMessage("Mascotas"),
-    "pho_tag" : MessageLookupByLibrary.simpleMessage("Pho"),
-    "photo_gallery_count" : m1,
-    "photo_gallery_description" : MessageLookupByLibrary.simpleMessage("Fotos aún no organizadas"),
-    "photo_gallery_title" : MessageLookupByLibrary.simpleMessage("Galería de fotos"),
-    "photo_location" : MessageLookupByLibrary.simpleMessage("Ubicación de la foto"),
-    "photos_always_organized" : MessageLookupByLibrary.simpleMessage("Ahora tus fotos siempre estarán organizadas"),
-    "piano_tag" : MessageLookupByLibrary.simpleMessage("Piano"),
-    "picnic_tag" : MessageLookupByLibrary.simpleMessage("Picnic"),
-    "picpics_photo_manager" : MessageLookupByLibrary.simpleMessage("picPics: Administrador de fotos"),
-    "pie_tag" : MessageLookupByLibrary.simpleMessage("Tarta"),
-    "pier_tag" : MessageLookupByLibrary.simpleMessage("Muelle"),
-    "pillow_tag" : MessageLookupByLibrary.simpleMessage("Almohada"),
-    "pitch_tag" : MessageLookupByLibrary.simpleMessage("Tono"),
-    "pixie_bob_tag" : MessageLookupByLibrary.simpleMessage("Pixie-bob"),
-    "pizza_tag" : MessageLookupByLibrary.simpleMessage("Pizza"),
-    "placemat_tag" : MessageLookupByLibrary.simpleMessage("Mantel individual"),
-    "plant_tag" : MessageLookupByLibrary.simpleMessage("Planta"),
-    "playground_tag" : MessageLookupByLibrary.simpleMessage("Patio de recreo"),
-    "plush_tag" : MessageLookupByLibrary.simpleMessage("Felpa"),
-    "pocket_tag" : MessageLookupByLibrary.simpleMessage("Bolsillo"),
-    "polo_tag" : MessageLookupByLibrary.simpleMessage("Polo"),
-    "pomacentridae_tag" : MessageLookupByLibrary.simpleMessage("Pomacentridae"),
-    "pool_tag" : MessageLookupByLibrary.simpleMessage("Piscina"),
-    "pop_music_tag" : MessageLookupByLibrary.simpleMessage("Música pop"),
-    "porcelain_tag" : MessageLookupByLibrary.simpleMessage("Porcelana"),
-    "poster_tag" : MessageLookupByLibrary.simpleMessage("Póster"),
-    "prairie_tag" : MessageLookupByLibrary.simpleMessage("Pradera"),
-    "premium_modal_description" : m2,
-    "premium_modal_get_premium_description" : MessageLookupByLibrary.simpleMessage("TODAS LAS FUNCIONES SIN ANUNCIOS"),
-    "premium_modal_get_premium_title" : MessageLookupByLibrary.simpleMessage("Obtener cuenta Premium"),
-    "premium_modal_watch_ad" : MessageLookupByLibrary.simpleMessage("Ver anuncio de vídeo para continuar"),
-    "presentation_tag" : MessageLookupByLibrary.simpleMessage("Presentación"),
-    "privacy_policy" : MessageLookupByLibrary.simpleMessage("Política de privacidad"),
-    "private_photos" : MessageLookupByLibrary.simpleMessage("Fotos privadas"),
-    "product_tag" : MessageLookupByLibrary.simpleMessage("Producto"),
-    "prom_tag" : MessageLookupByLibrary.simpleMessage("Paseo"),
-    "protect_with_encryption" : MessageLookupByLibrary.simpleMessage("Protege tus fotos personales con cifrado al que sólo se puede acceder con una contraseña PIN."),
-    "quilting_tag" : MessageLookupByLibrary.simpleMessage("Acolchado"),
-    "race_tag" : MessageLookupByLibrary.simpleMessage("Carrera"),
-    "racing_tag" : MessageLookupByLibrary.simpleMessage("Carreras"),
-    "rafting_tag" : MessageLookupByLibrary.simpleMessage("Canotaje"),
-    "ragdoll_tag" : MessageLookupByLibrary.simpleMessage("Muñeca de trapo"),
-    "rainbow_tag" : MessageLookupByLibrary.simpleMessage("Arco iris"),
-    "ranch_tag" : MessageLookupByLibrary.simpleMessage("Rancho"),
-    "rate_this_app" : MessageLookupByLibrary.simpleMessage("Califica esta aplicación"),
-    "receipt_tag" : MessageLookupByLibrary.simpleMessage("Recibo"),
-    "recent_tags" : MessageLookupByLibrary.simpleMessage("Etiquetas recientes"),
-    "reef_tag" : MessageLookupByLibrary.simpleMessage("Arrecife"),
-    "rein_tag" : MessageLookupByLibrary.simpleMessage("Rienda"),
-    "require_secret_key" : MessageLookupByLibrary.simpleMessage("Requerir clave secreta"),
-    "restore_purchase" : MessageLookupByLibrary.simpleMessage("Restaurar compra"),
-    "rickshaw_tag" : MessageLookupByLibrary.simpleMessage("Bicitaxi"),
-    "ring_tag" : MessageLookupByLibrary.simpleMessage("Anillo"),
-    "river_tag" : MessageLookupByLibrary.simpleMessage("Río"),
-    "road_tag" : MessageLookupByLibrary.simpleMessage("La carretera "),
-    "rock_tag" : MessageLookupByLibrary.simpleMessage("Rock"),
-    "rocket_tag" : MessageLookupByLibrary.simpleMessage("Cohete"),
-    "rodeo_tag" : MessageLookupByLibrary.simpleMessage("Rodeo"),
-    "roller_tag" : MessageLookupByLibrary.simpleMessage("Rodillo"),
-    "roof_tag" : MessageLookupByLibrary.simpleMessage("Techo"),
-    "room_tag" : MessageLookupByLibrary.simpleMessage("Habitación"),
-    "rowing_tag" : MessageLookupByLibrary.simpleMessage("Remo"),
-    "rugby_tag" : MessageLookupByLibrary.simpleMessage("Rugby"),
-    "ruins_tag" : MessageLookupByLibrary.simpleMessage("Restos"),
-    "running_tag" : MessageLookupByLibrary.simpleMessage("Corriendo"),
-    "safari_tag" : MessageLookupByLibrary.simpleMessage("Safari"),
-    "sailboat_tag" : MessageLookupByLibrary.simpleMessage("Velero"),
-    "sand_tag" : MessageLookupByLibrary.simpleMessage("Arena"),
-    "santa_claus_tag" : MessageLookupByLibrary.simpleMessage("Papá Noel"),
-    "sari_tag" : MessageLookupByLibrary.simpleMessage("Sari"),
-    "saucer_tag" : MessageLookupByLibrary.simpleMessage("Platillo"),
-    "save" : MessageLookupByLibrary.simpleMessage("guardar"),
-    "save_location" : MessageLookupByLibrary.simpleMessage("Guardar dirección"),
-    "scarf_tag" : MessageLookupByLibrary.simpleMessage("Bufanda"),
-    "school_tag" : MessageLookupByLibrary.simpleMessage("Colegio"),
-    "screenshot_tag" : MessageLookupByLibrary.simpleMessage("Captura de pantalla"),
-    "screenshots_tag" : MessageLookupByLibrary.simpleMessage("Capturas de pantalla"),
-    "scuba_diving_tag" : MessageLookupByLibrary.simpleMessage("Submarinismo"),
-    "seal_tag" : MessageLookupByLibrary.simpleMessage("Sello"),
-    "search" : MessageLookupByLibrary.simpleMessage("Buscar..."),
-    "search_all_tags_not_found" : MessageLookupByLibrary.simpleMessage("No se encontraron imágenes con todas las etiquetas"),
-    "search_results" : MessageLookupByLibrary.simpleMessage("Resultados de la búsqueda"),
-    "secret_key_created" : MessageLookupByLibrary.simpleMessage("¡Clave secreta creada correctamente!"),
-    "secret_photos" : MessageLookupByLibrary.simpleMessage("Fotos secretas"),
-    "selfie_tag" : MessageLookupByLibrary.simpleMessage("Selfie"),
-    "selfies_tag" : MessageLookupByLibrary.simpleMessage("Selfies"),
-    "settings" : MessageLookupByLibrary.simpleMessage("Ajustes"),
-    "share_with_friends" : MessageLookupByLibrary.simpleMessage("Compartir con amigos"),
-    "shelf_tag" : MessageLookupByLibrary.simpleMessage("Estante"),
-    "shell_tag" : MessageLookupByLibrary.simpleMessage("Cáscara"),
-    "shetland_sheepdog_tag" : MessageLookupByLibrary.simpleMessage("Perro pastor de Shetland"),
-    "shikoku_tag" : MessageLookupByLibrary.simpleMessage("Shikoku"),
-    "shipwreck_tag" : MessageLookupByLibrary.simpleMessage("Naufragio"),
-    "shoe_tag" : MessageLookupByLibrary.simpleMessage("Zapato"),
-    "shorts_tag" : MessageLookupByLibrary.simpleMessage("Pantalones cortos"),
-    "sign" : MessageLookupByLibrary.simpleMessage("Firmar"),
-    "singer_tag" : MessageLookupByLibrary.simpleMessage("Cantante"),
-    "sink_tag" : MessageLookupByLibrary.simpleMessage("Lavabo"),
-    "sitting_tag" : MessageLookupByLibrary.simpleMessage("Sentado"),
-    "skateboard_tag" : MessageLookupByLibrary.simpleMessage("Patineta"),
-    "skateboarder_tag" : MessageLookupByLibrary.simpleMessage("Skater"),
-    "skiff_tag" : MessageLookupByLibrary.simpleMessage("Esquife"),
-    "skin_tag" : MessageLookupByLibrary.simpleMessage("Piel"),
-    "sky_tag" : MessageLookupByLibrary.simpleMessage("Cielo"),
-    "skyline_tag" : MessageLookupByLibrary.simpleMessage("Horizonte"),
-    "skyscraper_tag" : MessageLookupByLibrary.simpleMessage("Rascacielos"),
-    "sled_tag" : MessageLookupByLibrary.simpleMessage("Trineo"),
-    "sledding_tag" : MessageLookupByLibrary.simpleMessage("Trineo"),
-    "sleep_tag" : MessageLookupByLibrary.simpleMessage("Dormir"),
-    "smile_tag" : MessageLookupByLibrary.simpleMessage("Sonreír"),
-    "sneakers_tag" : MessageLookupByLibrary.simpleMessage("Zapatillas"),
-    "snorkeling_tag" : MessageLookupByLibrary.simpleMessage("Bucear"),
-    "snowboarding_tag" : MessageLookupByLibrary.simpleMessage("Snowboarding"),
-    "soccer_tag" : MessageLookupByLibrary.simpleMessage("Fútbol"),
-    "softball_tag" : MessageLookupByLibrary.simpleMessage("Sofbol"),
-    "soil_tag" : MessageLookupByLibrary.simpleMessage("Suelo"),
-    "song_tag" : MessageLookupByLibrary.simpleMessage("Canción"),
-    "space_tag" : MessageLookupByLibrary.simpleMessage("Espacio"),
-    "sparkler_tag" : MessageLookupByLibrary.simpleMessage("Bengala"),
-    "speedboat_tag" : MessageLookupByLibrary.simpleMessage("Lancha rápida"),
-    "sphynx_tag" : MessageLookupByLibrary.simpleMessage("Sphynx"),
-    "sports_tag" : MessageLookupByLibrary.simpleMessage("Deportes"),
-    "stadium_tag" : MessageLookupByLibrary.simpleMessage("Estadio"),
-    "stairs_tag" : MessageLookupByLibrary.simpleMessage("Escalera"),
-    "standing_tag" : MessageLookupByLibrary.simpleMessage("En pie"),
-    "star_tag" : MessageLookupByLibrary.simpleMessage("Estrella"),
-    "start" : MessageLookupByLibrary.simpleMessage("Iniciar"),
-    "start_tagging" : MessageLookupByLibrary.simpleMessage("Comienza a etiquetar"),
-    "statue_tag" : MessageLookupByLibrary.simpleMessage("Estatua"),
-    "steaming_tag" : MessageLookupByLibrary.simpleMessage("Humeante"),
-    "storm_tag" : MessageLookupByLibrary.simpleMessage("Tormenta"),
-    "strap_tag" : MessageLookupByLibrary.simpleMessage("Correa"),
-    "stuffed_toy_tag" : MessageLookupByLibrary.simpleMessage("Peluche"),
-    "submarine_tag" : MessageLookupByLibrary.simpleMessage("Submarino"),
-    "subwoofer_tag" : MessageLookupByLibrary.simpleMessage("Subwoofer"),
-    "suggestions" : MessageLookupByLibrary.simpleMessage("Sugerencias"),
-    "sunglasses_tag" : MessageLookupByLibrary.simpleMessage("Gafas de sol"),
-    "sunset_tag" : MessageLookupByLibrary.simpleMessage("Puesta de sol"),
-    "superhero_tag" : MessageLookupByLibrary.simpleMessage("Superhéroe"),
-    "superman_tag" : MessageLookupByLibrary.simpleMessage("Superhombre"),
-    "supervillain_tag" : MessageLookupByLibrary.simpleMessage("Supervillano"),
-    "supper_tag" : MessageLookupByLibrary.simpleMessage("Cena"),
-    "surfboard_tag" : MessageLookupByLibrary.simpleMessage("Tabla de surf"),
-    "surfing_tag" : MessageLookupByLibrary.simpleMessage("Surf"),
-    "sushi_tag" : MessageLookupByLibrary.simpleMessage("Sushi"),
-    "swamp_tag" : MessageLookupByLibrary.simpleMessage("Pantano"),
-    "swimming_tag" : MessageLookupByLibrary.simpleMessage("Nadando"),
-    "swimwear_tag" : MessageLookupByLibrary.simpleMessage("Trajes de baño"),
-    "swing_tag" : MessageLookupByLibrary.simpleMessage("Columpio"),
-    "tablecloth_tag" : MessageLookupByLibrary.simpleMessage("Manteles"),
-    "tableware_tag" : MessageLookupByLibrary.simpleMessage("Vajilla"),
-    "tag_multiple_photos_at_once" : MessageLookupByLibrary.simpleMessage("Etiqueta varias fotos a la vez"),
-    "take_a_look" : MessageLookupByLibrary.simpleMessage("¡Echa un vistazo a esta aplicación!"),
-    "take_a_look_description" : m3,
-    "tattoo_tag" : MessageLookupByLibrary.simpleMessage("Tatuaje"),
-    "team_tag" : MessageLookupByLibrary.simpleMessage("Equipo"),
-    "television_tag" : MessageLookupByLibrary.simpleMessage("Televisión"),
-    "temple_tag" : MessageLookupByLibrary.simpleMessage("Templo"),
-    "terms_of_use" : MessageLookupByLibrary.simpleMessage("Términos de uso"),
-    "textile_tag" : MessageLookupByLibrary.simpleMessage("Textil"),
-    "thanksgiving_tag" : MessageLookupByLibrary.simpleMessage("Acción de gracias"),
-    "tie_tag" : MessageLookupByLibrary.simpleMessage("es decir"),
-    "tights_tag" : MessageLookupByLibrary.simpleMessage("Medias"),
-    "tile_tag" : MessageLookupByLibrary.simpleMessage("Loseta"),
-    "time" : MessageLookupByLibrary.simpleMessage("Hora"),
-    "tire_tag" : MessageLookupByLibrary.simpleMessage("Neumático"),
-    "toe_tag" : MessageLookupByLibrary.simpleMessage("Dedo del pie"),
-    "toggle_date" : MessageLookupByLibrary.simpleMessage("Fecha"),
-    "toggle_days" : MessageLookupByLibrary.simpleMessage("Días"),
-    "toggle_months" : MessageLookupByLibrary.simpleMessage("Meses"),
-    "toggle_tags" : MessageLookupByLibrary.simpleMessage("Etiquetas"),
-    "tower_tag" : MessageLookupByLibrary.simpleMessage("Torre"),
-    "toy_tag" : MessageLookupByLibrary.simpleMessage("Juguete"),
-    "track_tag" : MessageLookupByLibrary.simpleMessage("Pista"),
-    "tractor_tag" : MessageLookupByLibrary.simpleMessage("Tractor"),
-    "train_tag" : MessageLookupByLibrary.simpleMessage("Entrenar"),
-    "travel_tag" : MessageLookupByLibrary.simpleMessage("Viajes"),
-    "trunk_tag" : MessageLookupByLibrary.simpleMessage("Maletero"),
-    "tsqr_tag" : MessageLookupByLibrary.simpleMessage("Times Square"),
-    "tubing_tag" : MessageLookupByLibrary.simpleMessage("Tubería"),
-    "turtle_tag" : MessageLookupByLibrary.simpleMessage("Tortuga"),
-    "tutorial_daily_package" : MessageLookupByLibrary.simpleMessage("Ofrecemos un paquete diario para que puedas organizar tu biblioteca gradualmente."),
-    "tutorial_however_you_want" : MessageLookupByLibrary.simpleMessage("Organiza tus fotos añadiendo etiquetas, como \'familia\', \'mascotas\' o lo que quieras."),
-    "tutorial_just_swipe" : MessageLookupByLibrary.simpleMessage("Después de añadir las etiquetas a tu foto, simplemente desliza para pasar a la siguiente."),
-    "tutorial_multiselect" : MessageLookupByLibrary.simpleMessage("Puede \"tocar y mantener\" en sus fotos para etiquetar varias fotos a la vez."),
-    "tutorial_secret" : MessageLookupByLibrary.simpleMessage("Oculte sus fotos privadas con una protección de código PIN, manteniéndolas seguras."),
-    "tuxedo_tag" : MessageLookupByLibrary.simpleMessage("Smoking"),
-    "twig_tag" : MessageLookupByLibrary.simpleMessage("Ramita"),
-    "umbrella_tag" : MessageLookupByLibrary.simpleMessage("Paraguas"),
-    "underwater_tag" : MessageLookupByLibrary.simpleMessage("Submarino"),
-    "unicycle_tag" : MessageLookupByLibrary.simpleMessage("Monociclo"),
-    "unlimited_private_pics" : MessageLookupByLibrary.simpleMessage("Fotos privadas ilimitadas"),
-    "vacation_tag" : MessageLookupByLibrary.simpleMessage("Vacaciones"),
-    "van_tag" : MessageLookupByLibrary.simpleMessage("Camioneta"),
-    "vegetable_tag" : MessageLookupByLibrary.simpleMessage("Vegetal"),
-    "vehicle_tag" : MessageLookupByLibrary.simpleMessage("Vehículo"),
-    "veil_tag" : MessageLookupByLibrary.simpleMessage("Velo"),
-    "view_hidden_photos" : MessageLookupByLibrary.simpleMessage("Para ver tus fotos ocultas, retira el bloqueo en los ajustes de la aplicación. Puedes desbloquearlas con tu PIN."),
-    "volcano_tag" : MessageLookupByLibrary.simpleMessage("Volcán"),
-    "wakeboarding_tag" : MessageLookupByLibrary.simpleMessage("Wakeboard"),
-    "wall_tag" : MessageLookupByLibrary.simpleMessage("Pared"),
-    "waterfall_tag" : MessageLookupByLibrary.simpleMessage("Cascada"),
-    "waterfowl_tag" : MessageLookupByLibrary.simpleMessage("Aves acuáticas"),
-    "waterskiing_tag" : MessageLookupByLibrary.simpleMessage("Esquí acuático"),
-    "web_page_tag" : MessageLookupByLibrary.simpleMessage("Página web"),
-    "welcome" : MessageLookupByLibrary.simpleMessage("¡Bienvenido!"),
-    "wetsuit_tag" : MessageLookupByLibrary.simpleMessage("Traje de neopreno"),
-    "wheel_tag" : MessageLookupByLibrary.simpleMessage("Rueda"),
-    "wheelbarrow_tag" : MessageLookupByLibrary.simpleMessage("Carretilla"),
-    "whiteboard_tag" : MessageLookupByLibrary.simpleMessage("Pizarron"),
-    "windshield_tag" : MessageLookupByLibrary.simpleMessage("Parabrisas"),
-    "windsurfing_tag" : MessageLookupByLibrary.simpleMessage("Windsurfing"),
-    "wine_tag" : MessageLookupByLibrary.simpleMessage("Vino"),
-    "wing_tag" : MessageLookupByLibrary.simpleMessage("Ala"),
-    "wool_tag" : MessageLookupByLibrary.simpleMessage("Lana"),
-    "work_tag" : MessageLookupByLibrary.simpleMessage("Trabajo"),
-    "wreath_tag" : MessageLookupByLibrary.simpleMessage("Guirnalda"),
-    "x_minutes" : MessageLookupByLibrary.simpleMessage("20 minutos"),
-    "year" : MessageLookupByLibrary.simpleMessage("año"),
-    "yes" : MessageLookupByLibrary.simpleMessage("Sí"),
-    "you_are_premium" : MessageLookupByLibrary.simpleMessage("¡Eres Premium!"),
-    "your_secret_key" : MessageLookupByLibrary.simpleMessage("Tu clave secreta")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "access_code": MessageLookupByLibrary.simpleMessage("Código de acceso"),
+        "access_code_sent": m0,
+        "add_multiple_tags":
+            MessageLookupByLibrary.simpleMessage("Añade múltiples etiquetas"),
+        "add_tag": MessageLookupByLibrary.simpleMessage("Añadir etiqueta"),
+        "add_tags": MessageLookupByLibrary.simpleMessage("Añadir etiquetas"),
+        "aerospace_engineering_tag":
+            MessageLookupByLibrary.simpleMessage("Ingeniería Aeroespacial"),
+        "aircraft_tag": MessageLookupByLibrary.simpleMessage("Aeronave"),
+        "airliner_tag": MessageLookupByLibrary.simpleMessage("Avión de línea"),
+        "airplane_tag": MessageLookupByLibrary.simpleMessage("Avión"),
+        "alcohol_tag": MessageLookupByLibrary.simpleMessage("Alcohol"),
+        "all_at_once": MessageLookupByLibrary.simpleMessage(
+            "Organiza varias fotos a la vez"),
+        "all_photos_were_tagged": MessageLookupByLibrary.simpleMessage(
+            "There are no more photos to organize."),
+        "all_search_tags": MessageLookupByLibrary.simpleMessage(
+            "Todas las etiquetas de búsqueda"),
+        "always": MessageLookupByLibrary.simpleMessage("Siempre"),
+        "aquarium_tag": MessageLookupByLibrary.simpleMessage("Acuario"),
+        "archery_tag": MessageLookupByLibrary.simpleMessage("Tiro al arco"),
+        "armrest_tag": MessageLookupByLibrary.simpleMessage("Brazo"),
+        "army_tag": MessageLookupByLibrary.simpleMessage("Ejército"),
+        "ask_photo_library_permission": MessageLookupByLibrary.simpleMessage(
+            "Necesitamos acceso a su biblioteca de fotos para que pueda comenzar a organizar sus fotos con picPics. ¡No se preocupe, sus datos nunca saldrán de su dispositivo!"),
+        "asphalt_tag": MessageLookupByLibrary.simpleMessage("Asfalto"),
+        "aurora_tag": MessageLookupByLibrary.simpleMessage("Aurora"),
+        "auto_renewable_first_part":
+            MessageLookupByLibrary.simpleMessage("La suscripción se "),
+        "auto_renewable_second_part":
+            MessageLookupByLibrary.simpleMessage("renueva automáticamente."),
+        "aviation_tag": MessageLookupByLibrary.simpleMessage("Aviación"),
+        "baby_tag": MessageLookupByLibrary.simpleMessage("Bebé"),
+        "backpacking_tag": MessageLookupByLibrary.simpleMessage("Mochilero"),
+        "badminton_tag": MessageLookupByLibrary.simpleMessage("Bádminton"),
+        "bag_tag": MessageLookupByLibrary.simpleMessage("Bolso"),
+        "balance_tag": MessageLookupByLibrary.simpleMessage("Equilibrar"),
+        "balloon_tag": MessageLookupByLibrary.simpleMessage("Globo"),
+        "ballroom_tag": MessageLookupByLibrary.simpleMessage("Salón de baile"),
+        "bangle_tag": MessageLookupByLibrary.simpleMessage("Brazalete"),
+        "bangs_tag": MessageLookupByLibrary.simpleMessage("Golpes"),
+        "bar_tag": MessageLookupByLibrary.simpleMessage("Bar"),
+        "barn_tag": MessageLookupByLibrary.simpleMessage("Granero"),
+        "basset_hound_tag":
+            MessageLookupByLibrary.simpleMessage("Perro basset"),
+        "bathing_tag": MessageLookupByLibrary.simpleMessage("Baños"),
+        "bathroom_tag": MessageLookupByLibrary.simpleMessage("Baño"),
+        "beach_tag": MessageLookupByLibrary.simpleMessage("Playa"),
+        "beanie_tag": MessageLookupByLibrary.simpleMessage("Gorro"),
+        "bear_tag": MessageLookupByLibrary.simpleMessage("Oso"),
+        "beard_tag": MessageLookupByLibrary.simpleMessage("Barba"),
+        "bedroom_tag": MessageLookupByLibrary.simpleMessage("Habitación"),
+        "bench_tag": MessageLookupByLibrary.simpleMessage("Banco"),
+        "bento_tag": MessageLookupByLibrary.simpleMessage("Bento"),
+        "bicycle_tag": MessageLookupByLibrary.simpleMessage("Bicicleta"),
+        "bird_tag": MessageLookupByLibrary.simpleMessage("Pájaro"),
+        "blazer_tag": MessageLookupByLibrary.simpleMessage("Chaqueta de sport"),
+        "boat_tag": MessageLookupByLibrary.simpleMessage("Barco"),
+        "bonfire_tag": MessageLookupByLibrary.simpleMessage("Hoguera"),
+        "boxer_tag": MessageLookupByLibrary.simpleMessage("Boxer"),
+        "bracelet_tag": MessageLookupByLibrary.simpleMessage("Pulsera"),
+        "branch_tag": MessageLookupByLibrary.simpleMessage("Rama"),
+        "bread_tag": MessageLookupByLibrary.simpleMessage("Pan"),
+        "brick_tag": MessageLookupByLibrary.simpleMessage("Ladrillo"),
+        "bride_tag": MessageLookupByLibrary.simpleMessage("Novia"),
+        "bridge_tag": MessageLookupByLibrary.simpleMessage("Puente"),
+        "bridle_tag": MessageLookupByLibrary.simpleMessage("Brida"),
+        "brig_tag": MessageLookupByLibrary.simpleMessage("Bergantín"),
+        "building_tag": MessageLookupByLibrary.simpleMessage("Edificio"),
+        "bull_tag": MessageLookupByLibrary.simpleMessage("Toro"),
+        "bullfighting_tag": MessageLookupByLibrary.simpleMessage("Toreo"),
+        "bumper_tag": MessageLookupByLibrary.simpleMessage("Parachoque"),
+        "bunk_bed_tag": MessageLookupByLibrary.simpleMessage("Litera"),
+        "bus_tag": MessageLookupByLibrary.simpleMessage("Autobús"),
+        "butterfly_tag": MessageLookupByLibrary.simpleMessage("Mariposa"),
+        "cabinetry_tag": MessageLookupByLibrary.simpleMessage("Ebanistería"),
+        "cage_tag": MessageLookupByLibrary.simpleMessage("Jaula"),
+        "cairn_terrier_tag":
+            MessageLookupByLibrary.simpleMessage("Cairn terrier"),
+        "cake_tag": MessageLookupByLibrary.simpleMessage("Pastel"),
+        "camping_tag": MessageLookupByLibrary.simpleMessage("Cámping"),
+        "cancel": MessageLookupByLibrary.simpleMessage("Cancelar"),
+        "cancel_anytime": MessageLookupByLibrary.simpleMessage(
+            "Puedes cancelar en cualquier momento"),
+        "canoe_tag": MessageLookupByLibrary.simpleMessage("Canoa"),
+        "canyon_tag": MessageLookupByLibrary.simpleMessage("Cañón"),
+        "cap_tag": MessageLookupByLibrary.simpleMessage("Gorra"),
+        "cappuccino_tag": MessageLookupByLibrary.simpleMessage("Capuchino"),
+        "car_tag": MessageLookupByLibrary.simpleMessage("Coche"),
+        "carnival_tag": MessageLookupByLibrary.simpleMessage("Carnaval"),
+        "casino_tag": MessageLookupByLibrary.simpleMessage("Casino"),
+        "castle_tag": MessageLookupByLibrary.simpleMessage("Castillo"),
+        "cat_tag": MessageLookupByLibrary.simpleMessage("Gato"),
+        "cathedral_tag": MessageLookupByLibrary.simpleMessage("Catedral"),
+        "cattle_tag": MessageLookupByLibrary.simpleMessage("Vacas"),
+        "cavalier_tag": MessageLookupByLibrary.simpleMessage("Caballero"),
+        "cave_tag": MessageLookupByLibrary.simpleMessage("Cueva"),
+        "caving_tag": MessageLookupByLibrary.simpleMessage("Espeleología"),
+        "centrepiece_tag":
+            MessageLookupByLibrary.simpleMessage("Habitación central"),
+        "chain_tag": MessageLookupByLibrary.simpleMessage("Cadena"),
+        "chair_tag": MessageLookupByLibrary.simpleMessage("Silla"),
+        "cheeseburger_tag":
+            MessageLookupByLibrary.simpleMessage("Hamburguesa con queso"),
+        "christmas_tag": MessageLookupByLibrary.simpleMessage("Navidad"),
+        "church_tag": MessageLookupByLibrary.simpleMessage("Iglesia"),
+        "circus_tag": MessageLookupByLibrary.simpleMessage("Circo"),
+        "class_tag": MessageLookupByLibrary.simpleMessage("Clase"),
+        "cliff_tag": MessageLookupByLibrary.simpleMessage("Acantilado"),
+        "clipper_tag": MessageLookupByLibrary.simpleMessage("Clíper"),
+        "clock_tag": MessageLookupByLibrary.simpleMessage("Reloj"),
+        "close": MessageLookupByLibrary.simpleMessage("Cerrar"),
+        "clown_tag": MessageLookupByLibrary.simpleMessage("Payaso"),
+        "coffee_tag": MessageLookupByLibrary.simpleMessage("Café"),
+        "cola_tag": MessageLookupByLibrary.simpleMessage("Reajuste salarial"),
+        "comet_tag": MessageLookupByLibrary.simpleMessage("Cometa"),
+        "comics_tag": MessageLookupByLibrary.simpleMessage("Historietas"),
+        "community_tag": MessageLookupByLibrary.simpleMessage("Comunidad"),
+        "computer_tag": MessageLookupByLibrary.simpleMessage("Computadora"),
+        "concert_tag": MessageLookupByLibrary.simpleMessage("Concierto"),
+        "confirm_email": MessageLookupByLibrary.simpleMessage(
+            "Confirma tu correo electrónico de registro para recibir tu código de acceso"),
+        "confirm_secret_key":
+            MessageLookupByLibrary.simpleMessage("Confirmar clave secreta"),
+        "construction_tag":
+            MessageLookupByLibrary.simpleMessage("Construcción"),
+        "continue_string": MessageLookupByLibrary.simpleMessage("Continuar"),
+        "cookie_tag": MessageLookupByLibrary.simpleMessage("Galleta"),
+        "cookware_and_bakeware_tag": MessageLookupByLibrary.simpleMessage(
+            "Utensilios de cocina y para hornear"),
+        "cool_tag": MessageLookupByLibrary.simpleMessage("Frio"),
+        "cotton_tag": MessageLookupByLibrary.simpleMessage("Algodón"),
+        "couch_tag": MessageLookupByLibrary.simpleMessage("Sofá"),
+        "countertop_tag": MessageLookupByLibrary.simpleMessage("Encimera"),
+        "country": MessageLookupByLibrary.simpleMessage("país"),
+        "couscous_tag": MessageLookupByLibrary.simpleMessage("Cuscús"),
+        "crew_tag": MessageLookupByLibrary.simpleMessage("Tripulación"),
+        "crochet_tag": MessageLookupByLibrary.simpleMessage("Tejer"),
+        "crocodile_tag": MessageLookupByLibrary.simpleMessage("Cocodrilo"),
+        "crowd_tag": MessageLookupByLibrary.simpleMessage("Multitud"),
+        "cuisine_tag": MessageLookupByLibrary.simpleMessage("Cocina"),
+        "cup_tag": MessageLookupByLibrary.simpleMessage("Taza"),
+        "curling_tag": MessageLookupByLibrary.simpleMessage("Curling"),
+        "curtain_tag": MessageLookupByLibrary.simpleMessage("Cortina"),
+        "cushion_tag": MessageLookupByLibrary.simpleMessage("Amortiguar"),
+        "cutlery_tag": MessageLookupByLibrary.simpleMessage("Cuchillería"),
+        "cycling_tag": MessageLookupByLibrary.simpleMessage("Ciclismo"),
+        "daily_challenge":
+            MessageLookupByLibrary.simpleMessage("Desafío diario"),
+        "daily_challenge_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Para que podamos enviar tus desafíos diarios, necesitamos autorización para enviar notificaciones, por lo tanto, es necesario que autorices las notificaciones en las opciones de tu teléfono celular"),
+        "daily_goal": MessageLookupByLibrary.simpleMessage("Objetivo diario"),
+        "daily_notification_description": MessageLookupByLibrary.simpleMessage(
+            "It\'s time to complete your picPics daily challenge!"),
+        "daily_notification_title":
+            MessageLookupByLibrary.simpleMessage("Daily challenge"),
+        "dalmatian_tag": MessageLookupByLibrary.simpleMessage("Dálmata"),
+        "dam_tag": MessageLookupByLibrary.simpleMessage("Represa"),
+        "dance_tag": MessageLookupByLibrary.simpleMessage("Danza"),
+        "deejay_tag": MessageLookupByLibrary.simpleMessage("Deejay"),
+        "delete": MessageLookupByLibrary.simpleMessage("Eliminar"),
+        "denim_tag": MessageLookupByLibrary.simpleMessage("Mezclilla"),
+        "desert_tag": MessageLookupByLibrary.simpleMessage("Desierto"),
+        "desk_tag": MessageLookupByLibrary.simpleMessage("Escritorio"),
+        "device_has_no_pics": MessageLookupByLibrary.simpleMessage(
+            "Este dispositivo no tiene ninguna foto en la galería, por lo que no hay fotos que se puedan etiquetar."),
+        "dinosaur_tag": MessageLookupByLibrary.simpleMessage("Dinosaurio"),
+        "disable_secret":
+            MessageLookupByLibrary.simpleMessage("¿Quieres mostrar esta foto?"),
+        "dog_tag": MessageLookupByLibrary.simpleMessage("Perro"),
+        "doily_tag": MessageLookupByLibrary.simpleMessage("Mantelito"),
+        "dont_ask_again":
+            MessageLookupByLibrary.simpleMessage("No volver a preguntar"),
+        "dragon_tag": MessageLookupByLibrary.simpleMessage("Continuar"),
+        "drawer_tag": MessageLookupByLibrary.simpleMessage("Cajón"),
+        "dress_tag": MessageLookupByLibrary.simpleMessage("Vestido"),
+        "duck_tag": MessageLookupByLibrary.simpleMessage("Pato"),
+        "dude_tag": MessageLookupByLibrary.simpleMessage("Tipo"),
+        "dune_tag": MessageLookupByLibrary.simpleMessage("Duna"),
+        "ear_tag": MessageLookupByLibrary.simpleMessage("Oreja"),
+        "eating_tag": MessageLookupByLibrary.simpleMessage("Comiendo"),
+        "edit_tag": MessageLookupByLibrary.simpleMessage("Editar etiqueta"),
+        "email": MessageLookupByLibrary.simpleMessage("Correo electrónico"),
+        "enable_faceid":
+            MessageLookupByLibrary.simpleMessage("Habilitar Face ID"),
+        "enable_fingerprint":
+            MessageLookupByLibrary.simpleMessage("Habilitar Huella Digital"),
+        "enable_irisscanner":
+            MessageLookupByLibrary.simpleMessage("habilitar Escáner de Iris"),
+        "enable_touchid":
+            MessageLookupByLibrary.simpleMessage("Habilitar Touch ID"),
+        "event_tag": MessageLookupByLibrary.simpleMessage("Evento"),
+        "export_all_gallery":
+            MessageLookupByLibrary.simpleMessage("Exportar toda la galería"),
+        "export_library":
+            MessageLookupByLibrary.simpleMessage("Exportar biblioteca"),
+        "eyelash_tag": MessageLookupByLibrary.simpleMessage("Pestaña"),
+        "factory_tag": MessageLookupByLibrary.simpleMessage("Fábrica"),
+        "family_tag": MessageLookupByLibrary.simpleMessage("Familia"),
+        "farm_tag": MessageLookupByLibrary.simpleMessage("Granja"),
+        "fast_food_tag": MessageLookupByLibrary.simpleMessage("Comida rápida"),
+        "feedback_bug_report": MessageLookupByLibrary.simpleMessage(
+            "Informe de errores y comentarios"),
+        "ferris_wheel_tag":
+            MessageLookupByLibrary.simpleMessage("Rueda de la fortuna"),
+        "fiction_tag": MessageLookupByLibrary.simpleMessage("Ficción"),
+        "field_tag": MessageLookupByLibrary.simpleMessage("Campo"),
+        "find_easily": MessageLookupByLibrary.simpleMessage(
+            "Encuentra tus fotos fácilmente"),
+        "fire_tag": MessageLookupByLibrary.simpleMessage("Fuego"),
+        "fireworks_tag":
+            MessageLookupByLibrary.simpleMessage("Fuegos artificiales"),
+        "fishing_tag": MessageLookupByLibrary.simpleMessage("Pescar"),
+        "flag_tag": MessageLookupByLibrary.simpleMessage("Bandera"),
+        "flap_tag": MessageLookupByLibrary.simpleMessage("Solapa"),
+        "flesh_tag": MessageLookupByLibrary.simpleMessage("Carne"),
+        "flora_tag": MessageLookupByLibrary.simpleMessage("Flora"),
+        "flower_tag": MessageLookupByLibrary.simpleMessage("Flor"),
+        "flowerpot_tag": MessageLookupByLibrary.simpleMessage("Maceta"),
+        "fog_tag": MessageLookupByLibrary.simpleMessage("Niebla"),
+        "food_tag": MessageLookupByLibrary.simpleMessage("Comida"),
+        "foods_tag": MessageLookupByLibrary.simpleMessage("Alimentos"),
+        "foot_tag": MessageLookupByLibrary.simpleMessage("Pie"),
+        "forest_tag": MessageLookupByLibrary.simpleMessage("Bosque"),
+        "forgot_secret_key": MessageLookupByLibrary.simpleMessage(
+            "¿Olvidaste la clave secreta?"),
+        "frigate_tag": MessageLookupByLibrary.simpleMessage("Fragata"),
+        "fruit_tag": MessageLookupByLibrary.simpleMessage("Fruta"),
+        "full_screen":
+            MessageLookupByLibrary.simpleMessage("Pantalla completa"),
+        "fun_tag": MessageLookupByLibrary.simpleMessage("Divertido"),
+        "fur_tag": MessageLookupByLibrary.simpleMessage("Pelaje"),
+        "gallery_access_permission":
+            MessageLookupByLibrary.simpleMessage("Permisos de acceso"),
+        "gallery_access_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Para comenzar a organizar tus fotos, necesitamos autorización para acceder a ellas."),
+        "gallery_access_reason": MessageLookupByLibrary.simpleMessage(
+            "Para organizar tus fotos, necesitamos acceder a tu galería de fotos."),
+        "garden_tag": MessageLookupByLibrary.simpleMessage("Jardín"),
+        "gelato_tag": MessageLookupByLibrary.simpleMessage("Helado"),
+        "gerbil_tag": MessageLookupByLibrary.simpleMessage("Jerbo"),
+        "get_premium_description": MessageLookupByLibrary.simpleMessage(
+            "PARA TENER TODAS ESTAS FUNCIONES"),
+        "get_premium_now":
+            MessageLookupByLibrary.simpleMessage("¡Hazte Premium ahora!"),
+        "get_premium_title":
+            MessageLookupByLibrary.simpleMessage("HAZTE PREMIUM"),
+        "glacier_tag": MessageLookupByLibrary.simpleMessage("Glaciar"),
+        "glasses_tag": MessageLookupByLibrary.simpleMessage("Gafas"),
+        "glitter_tag": MessageLookupByLibrary.simpleMessage("Brillantina"),
+        "goggles_tag":
+            MessageLookupByLibrary.simpleMessage("Gafas de protección"),
+        "gown_tag": MessageLookupByLibrary.simpleMessage("Vestido"),
+        "graduation_tag": MessageLookupByLibrary.simpleMessage("Graduación"),
+        "grandparent_tag": MessageLookupByLibrary.simpleMessage("Abuelo"),
+        "groom_tag": MessageLookupByLibrary.simpleMessage("Novio"),
+        "gymnastics_tag": MessageLookupByLibrary.simpleMessage("Gimnasia"),
+        "hair_tag": MessageLookupByLibrary.simpleMessage("Pelo"),
+        "hand_tag": MessageLookupByLibrary.simpleMessage("Mano"),
+        "handbag_tag": MessageLookupByLibrary.simpleMessage("Bolso"),
+        "handrail_tag": MessageLookupByLibrary.simpleMessage("Pretil"),
+        "hanukkah_tag": MessageLookupByLibrary.simpleMessage("Hanukkah"),
+        "hat_tag": MessageLookupByLibrary.simpleMessage("Sombrero"),
+        "heart_tag": MessageLookupByLibrary.simpleMessage("Corazón"),
+        "helicopter_tag": MessageLookupByLibrary.simpleMessage("Helicóptero"),
+        "helmet_tag": MessageLookupByLibrary.simpleMessage("Casco"),
+        "herd_tag": MessageLookupByLibrary.simpleMessage("Manada"),
+        "himalayan_tag": MessageLookupByLibrary.simpleMessage("Himalaya"),
+        "home_tag": MessageLookupByLibrary.simpleMessage("Casa"),
+        "horn_tag": MessageLookupByLibrary.simpleMessage("Cuerno"),
+        "horse_tag": MessageLookupByLibrary.simpleMessage("Caballo"),
+        "hot_dog_tag": MessageLookupByLibrary.simpleMessage("Pancho"),
+        "how_many_pics": MessageLookupByLibrary.simpleMessage("Cuántas fotos"),
+        "ice_tag": MessageLookupByLibrary.simpleMessage("Hielo"),
+        "iceberg_tag": MessageLookupByLibrary.simpleMessage("Iceberg"),
+        "icicle_tag": MessageLookupByLibrary.simpleMessage("Carámbano"),
+        "icing_tag": MessageLookupByLibrary.simpleMessage("Formación de hielo"),
+        "icon_tag": MessageLookupByLibrary.simpleMessage("Icono"),
+        "infinite_tags":
+            MessageLookupByLibrary.simpleMessage("Etiquetas infinitas"),
+        "infrastructure_tag":
+            MessageLookupByLibrary.simpleMessage("Infraestructura"),
+        "insect_tag": MessageLookupByLibrary.simpleMessage("Insecto"),
+        "jacket_tag": MessageLookupByLibrary.simpleMessage("Chaqueta"),
+        "jeans_tag": MessageLookupByLibrary.simpleMessage("Pantalones"),
+        "jersey_tag": MessageLookupByLibrary.simpleMessage("Jersey"),
+        "jewellery_tag": MessageLookupByLibrary.simpleMessage("Joyería"),
+        "joker_tag": MessageLookupByLibrary.simpleMessage("Bufón"),
+        "juice_tag": MessageLookupByLibrary.simpleMessage("Jugo"),
+        "jungle_tag": MessageLookupByLibrary.simpleMessage("Selva"),
+        "junk_tag": MessageLookupByLibrary.simpleMessage("Basura"),
+        "kayak_tag": MessageLookupByLibrary.simpleMessage("Kayac"),
+        "keep_asking":
+            MessageLookupByLibrary.simpleMessage("Seguir preguntando"),
+        "keep_safe": MessageLookupByLibrary.simpleMessage(
+            "Tu foto ahora está segura en picPics. ¿Quieres eliminarla del carrete de tu cámara?"),
+        "kitchen_tag": MessageLookupByLibrary.simpleMessage("Cocina"),
+        "knitting_tag": MessageLookupByLibrary.simpleMessage("Tejido de punto"),
+        "lake_tag": MessageLookupByLibrary.simpleMessage("Lago"),
+        "lampshade_tag":
+            MessageLookupByLibrary.simpleMessage("Pantalla de lámpara"),
+        "language": MessageLookupByLibrary.simpleMessage("Idioma"),
+        "larva_tag": MessageLookupByLibrary.simpleMessage("Larva"),
+        "laugh_tag": MessageLookupByLibrary.simpleMessage("Risa"),
+        "leather_tag": MessageLookupByLibrary.simpleMessage("Cuero"),
+        "leggings_tag": MessageLookupByLibrary.simpleMessage("Polainas"),
+        "lego_tag": MessageLookupByLibrary.simpleMessage("Lego"),
+        "leisure_tag": MessageLookupByLibrary.simpleMessage("Ocio"),
+        "lighthouse_tag": MessageLookupByLibrary.simpleMessage("Faro"),
+        "lightning_tag": MessageLookupByLibrary.simpleMessage("Relámpago"),
+        "lipstick_tag": MessageLookupByLibrary.simpleMessage("Lápiz labial"),
+        "lock_with_pin": MessageLookupByLibrary.simpleMessage(
+            "Bloquea tus fotos privadas con un código PIN."),
+        "lock_your_photos":
+            MessageLookupByLibrary.simpleMessage("Bloquea tus fotos"),
+        "longboard_tag": MessageLookupByLibrary.simpleMessage("Longboard"),
+        "love_tag": MessageLookupByLibrary.simpleMessage("Amor"),
+        "loveseat_tag":
+            MessageLookupByLibrary.simpleMessage("Sofá de dos plazas"),
+        "lunch_tag": MessageLookupByLibrary.simpleMessage("Almuerzo"),
+        "marathon_tag": MessageLookupByLibrary.simpleMessage("Maratón"),
+        "meal_tag": MessageLookupByLibrary.simpleMessage("Comida"),
+        "menu_tag": MessageLookupByLibrary.simpleMessage("Menú"),
+        "metal_tag": MessageLookupByLibrary.simpleMessage("Metal"),
+        "militia_tag": MessageLookupByLibrary.simpleMessage("Milicia"),
+        "miniature_tag": MessageLookupByLibrary.simpleMessage("Miniatura"),
+        "mobile_phone_tag":
+            MessageLookupByLibrary.simpleMessage("Teléfono móvil"),
+        "model_tag": MessageLookupByLibrary.simpleMessage("Modelo"),
+        "money_tag": MessageLookupByLibrary.simpleMessage("Dinero"),
+        "monochrome_tag": MessageLookupByLibrary.simpleMessage("Monocromo"),
+        "month": MessageLookupByLibrary.simpleMessage("mes"),
+        "monument_tag": MessageLookupByLibrary.simpleMessage("Monumento"),
+        "moon_tag": MessageLookupByLibrary.simpleMessage("Luna"),
+        "mortarboard_tag": MessageLookupByLibrary.simpleMessage("Birrete"),
+        "mosque_tag": MessageLookupByLibrary.simpleMessage("Mezquita"),
+        "motorcycle_tag": MessageLookupByLibrary.simpleMessage("Motocicleta"),
+        "mountain_tag": MessageLookupByLibrary.simpleMessage("Montaña"),
+        "moustache_tag": MessageLookupByLibrary.simpleMessage("Bigote"),
+        "mouth_tag": MessageLookupByLibrary.simpleMessage("Boca"),
+        "mufti_tag": MessageLookupByLibrary.simpleMessage("Muftí"),
+        "muscle_tag": MessageLookupByLibrary.simpleMessage("Músculo"),
+        "museum_tag": MessageLookupByLibrary.simpleMessage("Museo"),
+        "musical_instrument_tag":
+            MessageLookupByLibrary.simpleMessage("Instrumento musical"),
+        "musical_tag": MessageLookupByLibrary.simpleMessage("Musical"),
+        "musician_tag": MessageLookupByLibrary.simpleMessage("Músico"),
+        "nail_tag": MessageLookupByLibrary.simpleMessage("Uña"),
+        "nebula_tag": MessageLookupByLibrary.simpleMessage("Nebulosa"),
+        "necklace_tag": MessageLookupByLibrary.simpleMessage("Collar"),
+        "neon_tag": MessageLookupByLibrary.simpleMessage("Neón"),
+        "net_tag": MessageLookupByLibrary.simpleMessage("Red"),
+        "new_secret_key":
+            MessageLookupByLibrary.simpleMessage("Nueva clave secreta"),
+        "news_tag": MessageLookupByLibrary.simpleMessage("Noticias"),
+        "newspaper_tag": MessageLookupByLibrary.simpleMessage("Periódico"),
+        "next": MessageLookupByLibrary.simpleMessage("Siguiente"),
+        "nightclub_tag": MessageLookupByLibrary.simpleMessage("Club nocturno"),
+        "no": MessageLookupByLibrary.simpleMessage("No"),
+        "no_ads": MessageLookupByLibrary.simpleMessage("Sin anuncios"),
+        "no_previous_purchase": MessageLookupByLibrary.simpleMessage(
+            "No hay ninguna compra previa"),
+        "no_tagged_photos": MessageLookupByLibrary.simpleMessage(
+            "Aún no tienes ninguna foto etiquetada"),
+        "no_tags_found": MessageLookupByLibrary.simpleMessage(
+            "No se encontró ninguna etiqueta"),
+        "no_valid_subscription": MessageLookupByLibrary.simpleMessage(
+            "No se pudo encontrar ninguna compra de suscripción válida."),
+        "notification_time":
+            MessageLookupByLibrary.simpleMessage("Hora de la notificación"),
+        "notifications": MessageLookupByLibrary.simpleMessage("Notificaciones"),
+        "ny_tag": MessageLookupByLibrary.simpleMessage("Nueva York"),
+        "odometer_tag":
+            MessageLookupByLibrary.simpleMessage("Cuentakilómetros"),
+        "ok": MessageLookupByLibrary.simpleMessage("Aceptar"),
+        "open_gallery": MessageLookupByLibrary.simpleMessage("Abrir galería"),
+        "organized_photos_description": MessageLookupByLibrary.simpleMessage(
+            "Fotos que ya se han etiquetado"),
+        "organized_photos_title":
+            MessageLookupByLibrary.simpleMessage("Fotos organizadas"),
+        "outerwear_tag": MessageLookupByLibrary.simpleMessage("Ropa de calle"),
+        "palace_tag": MessageLookupByLibrary.simpleMessage("Palacio"),
+        "paper_tag": MessageLookupByLibrary.simpleMessage("Papel"),
+        "park_tag": MessageLookupByLibrary.simpleMessage("Parque"),
+        "parties_tag": MessageLookupByLibrary.simpleMessage("Fiestas"),
+        "party_tag": MessageLookupByLibrary.simpleMessage("Fiesta"),
+        "passport_tag": MessageLookupByLibrary.simpleMessage("Pasaporte"),
+        "pasteles_tag": MessageLookupByLibrary.simpleMessage("Pasteles"),
+        "pattern_tag": MessageLookupByLibrary.simpleMessage("Patrón"),
+        "penguin_tag": MessageLookupByLibrary.simpleMessage("Pingüino"),
+        "person_tag": MessageLookupByLibrary.simpleMessage("Persona"),
+        "pet_tag": MessageLookupByLibrary.simpleMessage("Mascota"),
+        "petal_tag": MessageLookupByLibrary.simpleMessage("Pétalo"),
+        "pets_tag": MessageLookupByLibrary.simpleMessage("Mascotas"),
+        "pho_tag": MessageLookupByLibrary.simpleMessage("Pho"),
+        "photo_gallery_count": m1,
+        "photo_gallery_description":
+            MessageLookupByLibrary.simpleMessage("Fotos aún no organizadas"),
+        "photo_gallery_title":
+            MessageLookupByLibrary.simpleMessage("Galería de fotos"),
+        "photo_location":
+            MessageLookupByLibrary.simpleMessage("Ubicación de la foto"),
+        "photos_always_organized": MessageLookupByLibrary.simpleMessage(
+            "Ahora tus fotos siempre estarán organizadas"),
+        "piano_tag": MessageLookupByLibrary.simpleMessage("Piano"),
+        "picnic_tag": MessageLookupByLibrary.simpleMessage("Picnic"),
+        "picpics_photo_manager": MessageLookupByLibrary.simpleMessage(
+            "picPics: Administrador de fotos"),
+        "pie_tag": MessageLookupByLibrary.simpleMessage("Tarta"),
+        "pier_tag": MessageLookupByLibrary.simpleMessage("Muelle"),
+        "pillow_tag": MessageLookupByLibrary.simpleMessage("Almohada"),
+        "pitch_tag": MessageLookupByLibrary.simpleMessage("Tono"),
+        "pixie_bob_tag": MessageLookupByLibrary.simpleMessage("Pixie-bob"),
+        "pizza_tag": MessageLookupByLibrary.simpleMessage("Pizza"),
+        "placemat_tag":
+            MessageLookupByLibrary.simpleMessage("Mantel individual"),
+        "plant_tag": MessageLookupByLibrary.simpleMessage("Planta"),
+        "playground_tag":
+            MessageLookupByLibrary.simpleMessage("Patio de recreo"),
+        "plush_tag": MessageLookupByLibrary.simpleMessage("Felpa"),
+        "pocket_tag": MessageLookupByLibrary.simpleMessage("Bolsillo"),
+        "polo_tag": MessageLookupByLibrary.simpleMessage("Polo"),
+        "pomacentridae_tag":
+            MessageLookupByLibrary.simpleMessage("Pomacentridae"),
+        "pool_tag": MessageLookupByLibrary.simpleMessage("Piscina"),
+        "pop_music_tag": MessageLookupByLibrary.simpleMessage("Música pop"),
+        "porcelain_tag": MessageLookupByLibrary.simpleMessage("Porcelana"),
+        "poster_tag": MessageLookupByLibrary.simpleMessage("Póster"),
+        "prairie_tag": MessageLookupByLibrary.simpleMessage("Pradera"),
+        "premium_modal_description": m2,
+        "premium_modal_get_premium_description":
+            MessageLookupByLibrary.simpleMessage(
+                "TODAS LAS FUNCIONES SIN ANUNCIOS"),
+        "premium_modal_get_premium_title":
+            MessageLookupByLibrary.simpleMessage("Obtener cuenta Premium"),
+        "premium_modal_watch_ad": MessageLookupByLibrary.simpleMessage(
+            "Ver anuncio de vídeo para continuar"),
+        "presentation_tag":
+            MessageLookupByLibrary.simpleMessage("Presentación"),
+        "privacy_policy":
+            MessageLookupByLibrary.simpleMessage("Política de privacidad"),
+        "private_photos":
+            MessageLookupByLibrary.simpleMessage("Fotos privadas"),
+        "product_tag": MessageLookupByLibrary.simpleMessage("Producto"),
+        "prom_tag": MessageLookupByLibrary.simpleMessage("Paseo"),
+        "protect_with_encryption": MessageLookupByLibrary.simpleMessage(
+            "Protege tus fotos personales con cifrado al que sólo se puede acceder con una contraseña PIN."),
+        "quilting_tag": MessageLookupByLibrary.simpleMessage("Acolchado"),
+        "race_tag": MessageLookupByLibrary.simpleMessage("Carrera"),
+        "racing_tag": MessageLookupByLibrary.simpleMessage("Carreras"),
+        "rafting_tag": MessageLookupByLibrary.simpleMessage("Canotaje"),
+        "ragdoll_tag": MessageLookupByLibrary.simpleMessage("Muñeca de trapo"),
+        "rainbow_tag": MessageLookupByLibrary.simpleMessage("Arco iris"),
+        "ranch_tag": MessageLookupByLibrary.simpleMessage("Rancho"),
+        "rate_this_app":
+            MessageLookupByLibrary.simpleMessage("Califica esta aplicación"),
+        "receipt_tag": MessageLookupByLibrary.simpleMessage("Recibo"),
+        "recent_tags":
+            MessageLookupByLibrary.simpleMessage("Etiquetas recientes"),
+        "reef_tag": MessageLookupByLibrary.simpleMessage("Arrecife"),
+        "rein_tag": MessageLookupByLibrary.simpleMessage("Rienda"),
+        "require_secret_key":
+            MessageLookupByLibrary.simpleMessage("Requerir clave secreta"),
+        "restore_purchase":
+            MessageLookupByLibrary.simpleMessage("Restaurar compra"),
+        "rickshaw_tag": MessageLookupByLibrary.simpleMessage("Bicitaxi"),
+        "ring_tag": MessageLookupByLibrary.simpleMessage("Anillo"),
+        "river_tag": MessageLookupByLibrary.simpleMessage("Río"),
+        "road_tag": MessageLookupByLibrary.simpleMessage("La carretera "),
+        "rock_tag": MessageLookupByLibrary.simpleMessage("Rock"),
+        "rocket_tag": MessageLookupByLibrary.simpleMessage("Cohete"),
+        "rodeo_tag": MessageLookupByLibrary.simpleMessage("Rodeo"),
+        "roller_tag": MessageLookupByLibrary.simpleMessage("Rodillo"),
+        "roof_tag": MessageLookupByLibrary.simpleMessage("Techo"),
+        "room_tag": MessageLookupByLibrary.simpleMessage("Habitación"),
+        "rowing_tag": MessageLookupByLibrary.simpleMessage("Remo"),
+        "rugby_tag": MessageLookupByLibrary.simpleMessage("Rugby"),
+        "ruins_tag": MessageLookupByLibrary.simpleMessage("Restos"),
+        "running_tag": MessageLookupByLibrary.simpleMessage("Corriendo"),
+        "safari_tag": MessageLookupByLibrary.simpleMessage("Safari"),
+        "sailboat_tag": MessageLookupByLibrary.simpleMessage("Velero"),
+        "sand_tag": MessageLookupByLibrary.simpleMessage("Arena"),
+        "santa_claus_tag": MessageLookupByLibrary.simpleMessage("Papá Noel"),
+        "sari_tag": MessageLookupByLibrary.simpleMessage("Sari"),
+        "saucer_tag": MessageLookupByLibrary.simpleMessage("Platillo"),
+        "save": MessageLookupByLibrary.simpleMessage("guardar"),
+        "save_location":
+            MessageLookupByLibrary.simpleMessage("Guardar dirección"),
+        "scarf_tag": MessageLookupByLibrary.simpleMessage("Bufanda"),
+        "school_tag": MessageLookupByLibrary.simpleMessage("Colegio"),
+        "screenshot_tag":
+            MessageLookupByLibrary.simpleMessage("Captura de pantalla"),
+        "screenshots_tag":
+            MessageLookupByLibrary.simpleMessage("Capturas de pantalla"),
+        "scuba_diving_tag":
+            MessageLookupByLibrary.simpleMessage("Submarinismo"),
+        "seal_tag": MessageLookupByLibrary.simpleMessage("Sello"),
+        "search": MessageLookupByLibrary.simpleMessage("Buscar..."),
+        "search_all_tags_not_found": MessageLookupByLibrary.simpleMessage(
+            "No se encontraron imágenes con todas las etiquetas"),
+        "search_results":
+            MessageLookupByLibrary.simpleMessage("Resultados de la búsqueda"),
+        "secret_key_created": MessageLookupByLibrary.simpleMessage(
+            "¡Clave secreta creada correctamente!"),
+        "secret_photos": MessageLookupByLibrary.simpleMessage("Fotos secretas"),
+        "selfie_tag": MessageLookupByLibrary.simpleMessage("Selfie"),
+        "selfies_tag": MessageLookupByLibrary.simpleMessage("Selfies"),
+        "settings": MessageLookupByLibrary.simpleMessage("Ajustes"),
+        "share_with_friends":
+            MessageLookupByLibrary.simpleMessage("Compartir con amigos"),
+        "shelf_tag": MessageLookupByLibrary.simpleMessage("Estante"),
+        "shell_tag": MessageLookupByLibrary.simpleMessage("Cáscara"),
+        "shetland_sheepdog_tag":
+            MessageLookupByLibrary.simpleMessage("Perro pastor de Shetland"),
+        "shikoku_tag": MessageLookupByLibrary.simpleMessage("Shikoku"),
+        "shipwreck_tag": MessageLookupByLibrary.simpleMessage("Naufragio"),
+        "shoe_tag": MessageLookupByLibrary.simpleMessage("Zapato"),
+        "shorts_tag": MessageLookupByLibrary.simpleMessage("Pantalones cortos"),
+        "sign": MessageLookupByLibrary.simpleMessage("Firmar"),
+        "singer_tag": MessageLookupByLibrary.simpleMessage("Cantante"),
+        "sink_tag": MessageLookupByLibrary.simpleMessage("Lavabo"),
+        "sitting_tag": MessageLookupByLibrary.simpleMessage("Sentado"),
+        "skateboard_tag": MessageLookupByLibrary.simpleMessage("Patineta"),
+        "skateboarder_tag": MessageLookupByLibrary.simpleMessage("Skater"),
+        "skiff_tag": MessageLookupByLibrary.simpleMessage("Esquife"),
+        "skin_tag": MessageLookupByLibrary.simpleMessage("Piel"),
+        "sky_tag": MessageLookupByLibrary.simpleMessage("Cielo"),
+        "skyline_tag": MessageLookupByLibrary.simpleMessage("Horizonte"),
+        "skyscraper_tag": MessageLookupByLibrary.simpleMessage("Rascacielos"),
+        "sled_tag": MessageLookupByLibrary.simpleMessage("Trineo"),
+        "sledding_tag": MessageLookupByLibrary.simpleMessage("Trineo"),
+        "sleep_tag": MessageLookupByLibrary.simpleMessage("Dormir"),
+        "smile_tag": MessageLookupByLibrary.simpleMessage("Sonreír"),
+        "sneakers_tag": MessageLookupByLibrary.simpleMessage("Zapatillas"),
+        "snorkeling_tag": MessageLookupByLibrary.simpleMessage("Bucear"),
+        "snowboarding_tag":
+            MessageLookupByLibrary.simpleMessage("Snowboarding"),
+        "soccer_tag": MessageLookupByLibrary.simpleMessage("Fútbol"),
+        "softball_tag": MessageLookupByLibrary.simpleMessage("Sofbol"),
+        "soil_tag": MessageLookupByLibrary.simpleMessage("Suelo"),
+        "song_tag": MessageLookupByLibrary.simpleMessage("Canción"),
+        "space_tag": MessageLookupByLibrary.simpleMessage("Espacio"),
+        "sparkler_tag": MessageLookupByLibrary.simpleMessage("Bengala"),
+        "speedboat_tag": MessageLookupByLibrary.simpleMessage("Lancha rápida"),
+        "sphynx_tag": MessageLookupByLibrary.simpleMessage("Sphynx"),
+        "sports_tag": MessageLookupByLibrary.simpleMessage("Deportes"),
+        "stadium_tag": MessageLookupByLibrary.simpleMessage("Estadio"),
+        "stairs_tag": MessageLookupByLibrary.simpleMessage("Escalera"),
+        "standing_tag": MessageLookupByLibrary.simpleMessage("En pie"),
+        "star_tag": MessageLookupByLibrary.simpleMessage("Estrella"),
+        "start": MessageLookupByLibrary.simpleMessage("Iniciar"),
+        "start_tagging":
+            MessageLookupByLibrary.simpleMessage("Comienza a etiquetar"),
+        "statue_tag": MessageLookupByLibrary.simpleMessage("Estatua"),
+        "steaming_tag": MessageLookupByLibrary.simpleMessage("Humeante"),
+        "storm_tag": MessageLookupByLibrary.simpleMessage("Tormenta"),
+        "strap_tag": MessageLookupByLibrary.simpleMessage("Correa"),
+        "stuffed_toy_tag": MessageLookupByLibrary.simpleMessage("Peluche"),
+        "submarine_tag": MessageLookupByLibrary.simpleMessage("Submarino"),
+        "subwoofer_tag": MessageLookupByLibrary.simpleMessage("Subwoofer"),
+        "suggestions": MessageLookupByLibrary.simpleMessage("Sugerencias"),
+        "sunglasses_tag": MessageLookupByLibrary.simpleMessage("Gafas de sol"),
+        "sunset_tag": MessageLookupByLibrary.simpleMessage("Puesta de sol"),
+        "superhero_tag": MessageLookupByLibrary.simpleMessage("Superhéroe"),
+        "superman_tag": MessageLookupByLibrary.simpleMessage("Superhombre"),
+        "supervillain_tag":
+            MessageLookupByLibrary.simpleMessage("Supervillano"),
+        "supper_tag": MessageLookupByLibrary.simpleMessage("Cena"),
+        "surfboard_tag": MessageLookupByLibrary.simpleMessage("Tabla de surf"),
+        "surfing_tag": MessageLookupByLibrary.simpleMessage("Surf"),
+        "sushi_tag": MessageLookupByLibrary.simpleMessage("Sushi"),
+        "swamp_tag": MessageLookupByLibrary.simpleMessage("Pantano"),
+        "swimming_tag": MessageLookupByLibrary.simpleMessage("Nadando"),
+        "swimwear_tag": MessageLookupByLibrary.simpleMessage("Trajes de baño"),
+        "swing_tag": MessageLookupByLibrary.simpleMessage("Columpio"),
+        "tablecloth_tag": MessageLookupByLibrary.simpleMessage("Manteles"),
+        "tableware_tag": MessageLookupByLibrary.simpleMessage("Vajilla"),
+        "tag_multiple_photos_at_once": MessageLookupByLibrary.simpleMessage(
+            "Etiqueta varias fotos a la vez"),
+        "take_a_look": MessageLookupByLibrary.simpleMessage(
+            "¡Echa un vistazo a esta aplicación!"),
+        "take_a_look_description": m3,
+        "tattoo_tag": MessageLookupByLibrary.simpleMessage("Tatuaje"),
+        "team_tag": MessageLookupByLibrary.simpleMessage("Equipo"),
+        "television_tag": MessageLookupByLibrary.simpleMessage("Televisión"),
+        "temple_tag": MessageLookupByLibrary.simpleMessage("Templo"),
+        "terms_of_use": MessageLookupByLibrary.simpleMessage("Términos de uso"),
+        "textile_tag": MessageLookupByLibrary.simpleMessage("Textil"),
+        "thanksgiving_tag":
+            MessageLookupByLibrary.simpleMessage("Acción de gracias"),
+        "tie_tag": MessageLookupByLibrary.simpleMessage("es decir"),
+        "tights_tag": MessageLookupByLibrary.simpleMessage("Medias"),
+        "tile_tag": MessageLookupByLibrary.simpleMessage("Loseta"),
+        "time": MessageLookupByLibrary.simpleMessage("Hora"),
+        "tire_tag": MessageLookupByLibrary.simpleMessage("Neumático"),
+        "toe_tag": MessageLookupByLibrary.simpleMessage("Dedo del pie"),
+        "toggle_date": MessageLookupByLibrary.simpleMessage("Fecha"),
+        "toggle_days": MessageLookupByLibrary.simpleMessage("Días"),
+        "toggle_months": MessageLookupByLibrary.simpleMessage("Meses"),
+        "toggle_tags": MessageLookupByLibrary.simpleMessage("Etiquetas"),
+        "tower_tag": MessageLookupByLibrary.simpleMessage("Torre"),
+        "toy_tag": MessageLookupByLibrary.simpleMessage("Juguete"),
+        "track_tag": MessageLookupByLibrary.simpleMessage("Pista"),
+        "tractor_tag": MessageLookupByLibrary.simpleMessage("Tractor"),
+        "train_tag": MessageLookupByLibrary.simpleMessage("Entrenar"),
+        "travel_tag": MessageLookupByLibrary.simpleMessage("Viajes"),
+        "trunk_tag": MessageLookupByLibrary.simpleMessage("Maletero"),
+        "tsqr_tag": MessageLookupByLibrary.simpleMessage("Times Square"),
+        "tubing_tag": MessageLookupByLibrary.simpleMessage("Tubería"),
+        "turtle_tag": MessageLookupByLibrary.simpleMessage("Tortuga"),
+        "tutorial_daily_package": MessageLookupByLibrary.simpleMessage(
+            "Ofrecemos un paquete diario para que puedas organizar tu biblioteca gradualmente."),
+        "tutorial_however_you_want": MessageLookupByLibrary.simpleMessage(
+            "Organiza tus fotos añadiendo etiquetas, como \'familia\', \'mascotas\' o lo que quieras."),
+        "tutorial_just_swipe": MessageLookupByLibrary.simpleMessage(
+            "Después de añadir las etiquetas a tu foto, simplemente desliza para pasar a la siguiente."),
+        "tutorial_multiselect": MessageLookupByLibrary.simpleMessage(
+            "Puede \"tocar y mantener\" en sus fotos para etiquetar varias fotos a la vez."),
+        "tutorial_secret": MessageLookupByLibrary.simpleMessage(
+            "Oculte sus fotos privadas con una protección de código PIN, manteniéndolas seguras."),
+        "tuxedo_tag": MessageLookupByLibrary.simpleMessage("Smoking"),
+        "twig_tag": MessageLookupByLibrary.simpleMessage("Ramita"),
+        "umbrella_tag": MessageLookupByLibrary.simpleMessage("Paraguas"),
+        "underwater_tag": MessageLookupByLibrary.simpleMessage("Submarino"),
+        "unicycle_tag": MessageLookupByLibrary.simpleMessage("Monociclo"),
+        "unlimited_private_pics":
+            MessageLookupByLibrary.simpleMessage("Fotos privadas ilimitadas"),
+        "vacation_tag": MessageLookupByLibrary.simpleMessage("Vacaciones"),
+        "van_tag": MessageLookupByLibrary.simpleMessage("Camioneta"),
+        "vegetable_tag": MessageLookupByLibrary.simpleMessage("Vegetal"),
+        "vehicle_tag": MessageLookupByLibrary.simpleMessage("Vehículo"),
+        "veil_tag": MessageLookupByLibrary.simpleMessage("Velo"),
+        "view_hidden_photos": MessageLookupByLibrary.simpleMessage(
+            "Para ver tus fotos ocultas, retira el bloqueo en los ajustes de la aplicación. Puedes desbloquearlas con tu PIN."),
+        "volcano_tag": MessageLookupByLibrary.simpleMessage("Volcán"),
+        "wakeboarding_tag": MessageLookupByLibrary.simpleMessage("Wakeboard"),
+        "wall_tag": MessageLookupByLibrary.simpleMessage("Pared"),
+        "waterfall_tag": MessageLookupByLibrary.simpleMessage("Cascada"),
+        "waterfowl_tag": MessageLookupByLibrary.simpleMessage("Aves acuáticas"),
+        "waterskiing_tag":
+            MessageLookupByLibrary.simpleMessage("Esquí acuático"),
+        "web_page_tag": MessageLookupByLibrary.simpleMessage("Página web"),
+        "welcome": MessageLookupByLibrary.simpleMessage("¡Bienvenido!"),
+        "wetsuit_tag":
+            MessageLookupByLibrary.simpleMessage("Traje de neopreno"),
+        "wheel_tag": MessageLookupByLibrary.simpleMessage("Rueda"),
+        "wheelbarrow_tag": MessageLookupByLibrary.simpleMessage("Carretilla"),
+        "whiteboard_tag": MessageLookupByLibrary.simpleMessage("Pizarron"),
+        "windshield_tag": MessageLookupByLibrary.simpleMessage("Parabrisas"),
+        "windsurfing_tag": MessageLookupByLibrary.simpleMessage("Windsurfing"),
+        "wine_tag": MessageLookupByLibrary.simpleMessage("Vino"),
+        "wing_tag": MessageLookupByLibrary.simpleMessage("Ala"),
+        "wool_tag": MessageLookupByLibrary.simpleMessage("Lana"),
+        "work_tag": MessageLookupByLibrary.simpleMessage("Trabajo"),
+        "wreath_tag": MessageLookupByLibrary.simpleMessage("Guirnalda"),
+        "x_minutes": MessageLookupByLibrary.simpleMessage("20 minutos"),
+        "year": MessageLookupByLibrary.simpleMessage("año"),
+        "yes": MessageLookupByLibrary.simpleMessage("Sí"),
+        "you_are_premium":
+            MessageLookupByLibrary.simpleMessage("¡Eres Premium!"),
+        "your_secret_key":
+            MessageLookupByLibrary.simpleMessage("Tu clave secreta")
+      };
 }
