@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -19,148 +19,226 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'ms';
 
-  static m0(email) => "Kunci akses telah dihantar ke ${email}";
+  static String m0(email) => "Kunci akses telah dihantar ke ${email}";
 
-  static m1(howMany) => "${Intl.plural(howMany, zero: 'Tiada gambar yang dipilih', one: '1 gambar dipilih', other: '${howMany} gambar dipilih')}";
+  static String m1(howMany) =>
+      "${Intl.plural(howMany, zero: 'Tiada gambar yang dipilih', one: '1 gambar dipilih', other: '${howMany} gambar dipilih')}";
 
-  static m2(number) => "Anda telah melengkapkan ${number} gambar harian percuma anda, adakah anda ingin teruskan?";
+  static String m2(number) =>
+      "Anda telah melengkapkan ${number} gambar harian percuma anda, adakah anda ingin teruskan?";
 
-  static m3(url) => "Untuk menguruskan semua gambar anda, pergi ke ${url}";
+  static String m3(url) =>
+      "Untuk menguruskan semua gambar anda, pergi ke ${url}";
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "access_code" : MessageLookupByLibrary.simpleMessage("Kod akses"),
-    "access_code_sent" : m0,
-    "add_multiple_tags" : MessageLookupByLibrary.simpleMessage("Tambah beberapa tag"),
-    "add_tag" : MessageLookupByLibrary.simpleMessage("Tambah tag"),
-    "add_tags" : MessageLookupByLibrary.simpleMessage("Tambah tag"),
-    "all_at_once" : MessageLookupByLibrary.simpleMessage("Susun banyak gambar sekaligus"),
-    "all_photos_were_tagged" : MessageLookupByLibrary.simpleMessage("There are no more photos to organize."),
-    "all_search_tags" : MessageLookupByLibrary.simpleMessage("Semua Tag Carian"),
-    "always" : MessageLookupByLibrary.simpleMessage("Selalu"),
-    "ask_photo_library_permission" : MessageLookupByLibrary.simpleMessage("Kami memerlukan akses ke pustaka foto anda supaya anda dapat mula mengatur foto anda dengan picPics. Jangan bimbang data anda tidak akan meninggalkan peranti anda!"),
-    "auto_renewable_first_part" : MessageLookupByLibrary.simpleMessage("Langganan adalah "),
-    "auto_renewable_second_part" : MessageLookupByLibrary.simpleMessage("boleh diperbaharui secara automatik."),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Batalkan"),
-    "cancel_anytime" : MessageLookupByLibrary.simpleMessage("Batalkan bila-bila"),
-    "close" : MessageLookupByLibrary.simpleMessage("Tutup"),
-    "confirm_email" : MessageLookupByLibrary.simpleMessage("Sahkan e-mel pendaftaran anda untuk menerima kod akses anda"),
-    "confirm_secret_key" : MessageLookupByLibrary.simpleMessage("Sahkan kunci rahsia"),
-    "continue_string" : MessageLookupByLibrary.simpleMessage("Teruskan"),
-    "country" : MessageLookupByLibrary.simpleMessage("negara"),
-    "daily_challenge" : MessageLookupByLibrary.simpleMessage("Cabaran harian"),
-    "daily_challenge_permission_description" : MessageLookupByLibrary.simpleMessage("Supaya kami dapat menghantar cabaran harian anda, kami memerlukan kebenaran untuk menghantar pemberitahuan, jadi, anda perlu memberi kebenaran pemberitahuan dalam pilihan peranti mudah alih anda"),
-    "daily_goal" : MessageLookupByLibrary.simpleMessage("Matlamat harian"),
-    "daily_notification_description" : MessageLookupByLibrary.simpleMessage("It\'s time to complete your picPics daily challenge!"),
-    "daily_notification_title" : MessageLookupByLibrary.simpleMessage("Daily challenge"),
-    "delete" : MessageLookupByLibrary.simpleMessage("Padamkan"),
-    "device_has_no_pics" : MessageLookupByLibrary.simpleMessage("Peranti ini tiada gambar di galeri, jadi tiada gambar yang dapat ditag."),
-    "disable_secret" : MessageLookupByLibrary.simpleMessage("Adakah anda ingin menyembunyikan gambar ini?"),
-    "dont_ask_again" : MessageLookupByLibrary.simpleMessage("Jangan tanya lagi"),
-    "edit_tag" : MessageLookupByLibrary.simpleMessage("Sunting tag"),
-    "email" : MessageLookupByLibrary.simpleMessage("E-mel"),
-    "enable_faceid" : MessageLookupByLibrary.simpleMessage("Dayakan Face ID"),
-    "enable_fingerprint" : MessageLookupByLibrary.simpleMessage("Dayakan Cap Jari"),
-    "enable_irisscanner" : MessageLookupByLibrary.simpleMessage("Dayakan Pengimbas Iris"),
-    "enable_touchid" : MessageLookupByLibrary.simpleMessage("Dayakan Touch ID"),
-    "export_all_gallery" : MessageLookupByLibrary.simpleMessage("Eksport semua galeri"),
-    "export_library" : MessageLookupByLibrary.simpleMessage("Eksport Perpustakaan"),
-    "family_tag" : MessageLookupByLibrary.simpleMessage("Keluarga"),
-    "feedback_bug_report" : MessageLookupByLibrary.simpleMessage("Maklum balas & laporan pepijat"),
-    "find_easily" : MessageLookupByLibrary.simpleMessage("Cari gambar anda dengan mudah"),
-    "foods_tag" : MessageLookupByLibrary.simpleMessage("Makanan"),
-    "forgot_secret_key" : MessageLookupByLibrary.simpleMessage("Terlupa kunci rahsia?"),
-    "full_screen" : MessageLookupByLibrary.simpleMessage("Skrin Penuh"),
-    "gallery_access_permission" : MessageLookupByLibrary.simpleMessage("Kebenaran akses"),
-    "gallery_access_permission_description" : MessageLookupByLibrary.simpleMessage("Untuk mula menguruskan gambar anda, kami memerlukan kebenaran untuk mengaksesnya"),
-    "gallery_access_reason" : MessageLookupByLibrary.simpleMessage("Untuk menguruskan gambar anda, kami memerlukan akses ke galeri gambar anda"),
-    "get_premium_description" : MessageLookupByLibrary.simpleMessage("UNTUK MEMILIKI SEMUA CIRI-CIRI INI"),
-    "get_premium_now" : MessageLookupByLibrary.simpleMessage("Dapatkan premium sekarang!"),
-    "get_premium_title" : MessageLookupByLibrary.simpleMessage("DAPATKAN PREMIUM"),
-    "home_tag" : MessageLookupByLibrary.simpleMessage("Rumah"),
-    "how_many_pics" : MessageLookupByLibrary.simpleMessage("Berapa banyak gambar"),
-    "infinite_tags" : MessageLookupByLibrary.simpleMessage("Tag tanpa had"),
-    "keep_asking" : MessageLookupByLibrary.simpleMessage("Teruskan bertanya"),
-    "keep_safe" : MessageLookupByLibrary.simpleMessage("Gambar anda kini selamat di picPics. Adakah anda ingin memadamkannya dari gulungan kamera anda?"),
-    "language" : MessageLookupByLibrary.simpleMessage("Bahasa"),
-    "lock_with_pin" : MessageLookupByLibrary.simpleMessage("Kunci gambar peribadi anda dengan kod PIN."),
-    "lock_your_photos" : MessageLookupByLibrary.simpleMessage("Kunci gambar anda"),
-    "month" : MessageLookupByLibrary.simpleMessage("bulan"),
-    "new_secret_key" : MessageLookupByLibrary.simpleMessage("Kunci rahsia baru"),
-    "next" : MessageLookupByLibrary.simpleMessage("Seterusnya"),
-    "no" : MessageLookupByLibrary.simpleMessage("Tidak"),
-    "no_ads" : MessageLookupByLibrary.simpleMessage("Tiada Iklan"),
-    "no_previous_purchase" : MessageLookupByLibrary.simpleMessage("Tiada Pembelian Sebelumnya"),
-    "no_tagged_photos" : MessageLookupByLibrary.simpleMessage("Anda belum ada gambar yang ditag"),
-    "no_tags_found" : MessageLookupByLibrary.simpleMessage("Tiada tag dijumpai"),
-    "no_valid_subscription" : MessageLookupByLibrary.simpleMessage("Tidak dapat mencari pembelian langganan yang sah."),
-    "notification_time" : MessageLookupByLibrary.simpleMessage("Masa pemberitahuan"),
-    "notifications" : MessageLookupByLibrary.simpleMessage("Pemberitahuan"),
-    "ny_tag" : MessageLookupByLibrary.simpleMessage("NY"),
-    "ok" : MessageLookupByLibrary.simpleMessage("Okey"),
-    "open_gallery" : MessageLookupByLibrary.simpleMessage("Buka galeri"),
-    "organized_photos_description" : MessageLookupByLibrary.simpleMessage("Gambar yang sudah ditag"),
-    "organized_photos_title" : MessageLookupByLibrary.simpleMessage("Gambar Terurus"),
-    "parties_tag" : MessageLookupByLibrary.simpleMessage("Pesta"),
-    "pets_tag" : MessageLookupByLibrary.simpleMessage("Haiwan Peliharaan"),
-    "photo_gallery_count" : m1,
-    "photo_gallery_description" : MessageLookupByLibrary.simpleMessage("Gambar belum diuruskan"),
-    "photo_gallery_title" : MessageLookupByLibrary.simpleMessage("Galeri Gambar"),
-    "photo_location" : MessageLookupByLibrary.simpleMessage("Lokasi gambar"),
-    "photos_always_organized" : MessageLookupByLibrary.simpleMessage("Kini gambar anda akan sentiasa terurus"),
-    "picpics_photo_manager" : MessageLookupByLibrary.simpleMessage("picPics - Pengurus Gambar"),
-    "premium_modal_description" : m2,
-    "premium_modal_get_premium_description" : MessageLookupByLibrary.simpleMessage("SEMUA CIRI TANPA IKLAN"),
-    "premium_modal_get_premium_title" : MessageLookupByLibrary.simpleMessage("Dapatkan Akaun Premium"),
-    "premium_modal_watch_ad" : MessageLookupByLibrary.simpleMessage("Tontoni iklan video untuk teruskan"),
-    "privacy_policy" : MessageLookupByLibrary.simpleMessage("Dasar Privasi"),
-    "private_photos" : MessageLookupByLibrary.simpleMessage("Foto Peribadi"),
-    "protect_with_encryption" : MessageLookupByLibrary.simpleMessage("Lindungi gambar peribadi anda dengan penyulitan yang hanya dapat diakses dengan kata laluan PIN."),
-    "rate_this_app" : MessageLookupByLibrary.simpleMessage("Nilaikan aplikasi ini"),
-    "recent_tags" : MessageLookupByLibrary.simpleMessage("Tag Terkini"),
-    "require_secret_key" : MessageLookupByLibrary.simpleMessage("Memerlukan kunci rahsia"),
-    "restore_purchase" : MessageLookupByLibrary.simpleMessage("Pulihkan pembelian"),
-    "save" : MessageLookupByLibrary.simpleMessage("simpan"),
-    "save_location" : MessageLookupByLibrary.simpleMessage("Simpan lokasi"),
-    "screenshots_tag" : MessageLookupByLibrary.simpleMessage("Tangkapan Skrin"),
-    "search" : MessageLookupByLibrary.simpleMessage("Cari..."),
-    "search_all_tags_not_found" : MessageLookupByLibrary.simpleMessage("Tiada gambar dijumpai dengan semua tag padanya"),
-    "search_results" : MessageLookupByLibrary.simpleMessage("Keputusan carian"),
-    "secret_key_created" : MessageLookupByLibrary.simpleMessage("Kunci rahsia berjaya dicipta!"),
-    "secret_photos" : MessageLookupByLibrary.simpleMessage("Gambar Rahsia"),
-    "selfies_tag" : MessageLookupByLibrary.simpleMessage("Swafoto"),
-    "settings" : MessageLookupByLibrary.simpleMessage("Tetapan"),
-    "share_with_friends" : MessageLookupByLibrary.simpleMessage("Kongsi bersama rakan"),
-    "sign" : MessageLookupByLibrary.simpleMessage("Tanda"),
-    "sports_tag" : MessageLookupByLibrary.simpleMessage("Sukan"),
-    "start" : MessageLookupByLibrary.simpleMessage("Mulakan"),
-    "start_tagging" : MessageLookupByLibrary.simpleMessage("Mula tag"),
-    "suggestions" : MessageLookupByLibrary.simpleMessage("Cadangan"),
-    "tag_multiple_photos_at_once" : MessageLookupByLibrary.simpleMessage("ag banyak gambar sekaligus"),
-    "take_a_look" : MessageLookupByLibrary.simpleMessage("Lihatlah aplikasi ini!"),
-    "take_a_look_description" : m3,
-    "terms_of_use" : MessageLookupByLibrary.simpleMessage("Syarat Penggunaan"),
-    "time" : MessageLookupByLibrary.simpleMessage("Masa"),
-    "toggle_date" : MessageLookupByLibrary.simpleMessage("Tarikh"),
-    "toggle_days" : MessageLookupByLibrary.simpleMessage("Hari"),
-    "toggle_months" : MessageLookupByLibrary.simpleMessage("Bulan"),
-    "toggle_tags" : MessageLookupByLibrary.simpleMessage("Tag"),
-    "travel_tag" : MessageLookupByLibrary.simpleMessage("Pengembaraan"),
-    "tsqr_tag" : MessageLookupByLibrary.simpleMessage("Times Square"),
-    "tutorial_daily_package" : MessageLookupByLibrary.simpleMessage("Kami membawa pakej harian untuk anda bagi menguruskan perpustakaan anda secara beransur-ansur."),
-    "tutorial_however_you_want" : MessageLookupByLibrary.simpleMessage("Susun gambar anda dengan menambahkan tag, seperti \'keluarga\', \'binatang peliharaan\', atau apa sahaja yang anda mahukan."),
-    "tutorial_just_swipe" : MessageLookupByLibrary.simpleMessage("Setelah menambahkan tag pada gambar anda, leret sahaja untuk pergi ke gambar seterusnya."),
-    "tutorial_multiselect" : MessageLookupByLibrary.simpleMessage("Anda boleh \"ketuk & tahan\" pada foto anda untuk menandai banyak foto sekaligus."),
-    "tutorial_secret" : MessageLookupByLibrary.simpleMessage("Sembunyikan foto peribadi anda dengan perlindungan kod pin, jaga keselamatannya."),
-    "unlimited_private_pics" : MessageLookupByLibrary.simpleMessage("Foto peribadi tanpa had"),
-    "vacation_tag" : MessageLookupByLibrary.simpleMessage("Percutian"),
-    "view_hidden_photos" : MessageLookupByLibrary.simpleMessage("Untuk melihat gambar tersembunyi anda, lepaskan kunci dalam tetapan aplikasi. Anda boleh membuka kunci dengan PIN anda."),
-    "welcome" : MessageLookupByLibrary.simpleMessage("Selamat Datang!"),
-    "work_tag" : MessageLookupByLibrary.simpleMessage("Kerja"),
-    "x_minutes" : MessageLookupByLibrary.simpleMessage("20 min"),
-    "year" : MessageLookupByLibrary.simpleMessage("tahun"),
-    "yes" : MessageLookupByLibrary.simpleMessage("Ya"),
-    "you_are_premium" : MessageLookupByLibrary.simpleMessage("Anda premium!"),
-    "your_secret_key" : MessageLookupByLibrary.simpleMessage("Kunci rahsia anda")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "access_code": MessageLookupByLibrary.simpleMessage("Kod akses"),
+        "access_code_sent": m0,
+        "add_multiple_tags":
+            MessageLookupByLibrary.simpleMessage("Tambah beberapa tag"),
+        "add_tag": MessageLookupByLibrary.simpleMessage("Tambah tag"),
+        "add_tags": MessageLookupByLibrary.simpleMessage("Tambah tag"),
+        "all_at_once": MessageLookupByLibrary.simpleMessage(
+            "Susun banyak gambar sekaligus"),
+        "all_photos_were_tagged": MessageLookupByLibrary.simpleMessage(
+            "There are no more photos to organize."),
+        "all_search_tags":
+            MessageLookupByLibrary.simpleMessage("Semua Tag Carian"),
+        "always": MessageLookupByLibrary.simpleMessage("Selalu"),
+        "ask_photo_library_permission": MessageLookupByLibrary.simpleMessage(
+            "Kami memerlukan akses ke pustaka foto anda supaya anda dapat mula mengatur foto anda dengan picPics. Jangan bimbang data anda tidak akan meninggalkan peranti anda!"),
+        "auto_renewable_first_part":
+            MessageLookupByLibrary.simpleMessage("Langganan adalah "),
+        "auto_renewable_second_part": MessageLookupByLibrary.simpleMessage(
+            "boleh diperbaharui secara automatik."),
+        "cancel": MessageLookupByLibrary.simpleMessage("Batalkan"),
+        "cancel_anytime":
+            MessageLookupByLibrary.simpleMessage("Batalkan bila-bila"),
+        "close": MessageLookupByLibrary.simpleMessage("Tutup"),
+        "confirm_email": MessageLookupByLibrary.simpleMessage(
+            "Sahkan e-mel pendaftaran anda untuk menerima kod akses anda"),
+        "confirm_secret_key":
+            MessageLookupByLibrary.simpleMessage("Sahkan kunci rahsia"),
+        "continue_string": MessageLookupByLibrary.simpleMessage("Teruskan"),
+        "country": MessageLookupByLibrary.simpleMessage("negara"),
+        "daily_challenge":
+            MessageLookupByLibrary.simpleMessage("Cabaran harian"),
+        "daily_challenge_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Supaya kami dapat menghantar cabaran harian anda, kami memerlukan kebenaran untuk menghantar pemberitahuan, jadi, anda perlu memberi kebenaran pemberitahuan dalam pilihan peranti mudah alih anda"),
+        "daily_goal": MessageLookupByLibrary.simpleMessage("Matlamat harian"),
+        "daily_notification_description": MessageLookupByLibrary.simpleMessage(
+            "It\'s time to complete your picPics daily challenge!"),
+        "daily_notification_title":
+            MessageLookupByLibrary.simpleMessage("Daily challenge"),
+        "delete": MessageLookupByLibrary.simpleMessage("Padamkan"),
+        "device_has_no_pics": MessageLookupByLibrary.simpleMessage(
+            "Peranti ini tiada gambar di galeri, jadi tiada gambar yang dapat ditag."),
+        "disable_secret": MessageLookupByLibrary.simpleMessage(
+            "Adakah anda ingin menyembunyikan gambar ini?"),
+        "dont_ask_again":
+            MessageLookupByLibrary.simpleMessage("Jangan tanya lagi"),
+        "edit_tag": MessageLookupByLibrary.simpleMessage("Sunting tag"),
+        "email": MessageLookupByLibrary.simpleMessage("E-mel"),
+        "enable_faceid":
+            MessageLookupByLibrary.simpleMessage("Dayakan Face ID"),
+        "enable_fingerprint":
+            MessageLookupByLibrary.simpleMessage("Dayakan Cap Jari"),
+        "enable_irisscanner":
+            MessageLookupByLibrary.simpleMessage("Dayakan Pengimbas Iris"),
+        "enable_touchid":
+            MessageLookupByLibrary.simpleMessage("Dayakan Touch ID"),
+        "export_all_gallery":
+            MessageLookupByLibrary.simpleMessage("Eksport semua galeri"),
+        "export_library":
+            MessageLookupByLibrary.simpleMessage("Eksport Perpustakaan"),
+        "family_tag": MessageLookupByLibrary.simpleMessage("Keluarga"),
+        "feedback_bug_report": MessageLookupByLibrary.simpleMessage(
+            "Maklum balas & laporan pepijat"),
+        "find_easily": MessageLookupByLibrary.simpleMessage(
+            "Cari gambar anda dengan mudah"),
+        "foods_tag": MessageLookupByLibrary.simpleMessage("Makanan"),
+        "forgot_secret_key":
+            MessageLookupByLibrary.simpleMessage("Terlupa kunci rahsia?"),
+        "full_screen": MessageLookupByLibrary.simpleMessage("Skrin Penuh"),
+        "gallery_access_permission":
+            MessageLookupByLibrary.simpleMessage("Kebenaran akses"),
+        "gallery_access_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Untuk mula menguruskan gambar anda, kami memerlukan kebenaran untuk mengaksesnya"),
+        "gallery_access_reason": MessageLookupByLibrary.simpleMessage(
+            "Untuk menguruskan gambar anda, kami memerlukan akses ke galeri gambar anda"),
+        "get_premium_description": MessageLookupByLibrary.simpleMessage(
+            "UNTUK MEMILIKI SEMUA CIRI-CIRI INI"),
+        "get_premium_now":
+            MessageLookupByLibrary.simpleMessage("Dapatkan premium sekarang!"),
+        "get_premium_title":
+            MessageLookupByLibrary.simpleMessage("DAPATKAN PREMIUM"),
+        "home_tag": MessageLookupByLibrary.simpleMessage("Rumah"),
+        "how_many_pics":
+            MessageLookupByLibrary.simpleMessage("Berapa banyak gambar"),
+        "infinite_tags": MessageLookupByLibrary.simpleMessage("Tag tanpa had"),
+        "keep_asking":
+            MessageLookupByLibrary.simpleMessage("Teruskan bertanya"),
+        "keep_safe": MessageLookupByLibrary.simpleMessage(
+            "Gambar anda kini selamat di picPics. Adakah anda ingin memadamkannya dari gulungan kamera anda?"),
+        "language": MessageLookupByLibrary.simpleMessage("Bahasa"),
+        "lock_with_pin": MessageLookupByLibrary.simpleMessage(
+            "Kunci gambar peribadi anda dengan kod PIN."),
+        "lock_your_photos":
+            MessageLookupByLibrary.simpleMessage("Kunci gambar anda"),
+        "month": MessageLookupByLibrary.simpleMessage("bulan"),
+        "new_secret_key":
+            MessageLookupByLibrary.simpleMessage("Kunci rahsia baru"),
+        "next": MessageLookupByLibrary.simpleMessage("Seterusnya"),
+        "no": MessageLookupByLibrary.simpleMessage("Tidak"),
+        "no_ads": MessageLookupByLibrary.simpleMessage("Tiada Iklan"),
+        "no_previous_purchase":
+            MessageLookupByLibrary.simpleMessage("Tiada Pembelian Sebelumnya"),
+        "no_tagged_photos": MessageLookupByLibrary.simpleMessage(
+            "Anda belum ada gambar yang ditag"),
+        "no_tags_found":
+            MessageLookupByLibrary.simpleMessage("Tiada tag dijumpai"),
+        "no_valid_subscription": MessageLookupByLibrary.simpleMessage(
+            "Tidak dapat mencari pembelian langganan yang sah."),
+        "notification_time":
+            MessageLookupByLibrary.simpleMessage("Masa pemberitahuan"),
+        "notifications": MessageLookupByLibrary.simpleMessage("Pemberitahuan"),
+        "ny_tag": MessageLookupByLibrary.simpleMessage("NY"),
+        "ok": MessageLookupByLibrary.simpleMessage("Okey"),
+        "open_gallery": MessageLookupByLibrary.simpleMessage("Buka galeri"),
+        "organized_photos_description":
+            MessageLookupByLibrary.simpleMessage("Gambar yang sudah ditag"),
+        "organized_photos_title":
+            MessageLookupByLibrary.simpleMessage("Gambar Terurus"),
+        "parties_tag": MessageLookupByLibrary.simpleMessage("Pesta"),
+        "pets_tag": MessageLookupByLibrary.simpleMessage("Haiwan Peliharaan"),
+        "photo_gallery_count": m1,
+        "photo_gallery_description":
+            MessageLookupByLibrary.simpleMessage("Gambar belum diuruskan"),
+        "photo_gallery_title":
+            MessageLookupByLibrary.simpleMessage("Galeri Gambar"),
+        "photo_location": MessageLookupByLibrary.simpleMessage("Lokasi gambar"),
+        "photos_always_organized": MessageLookupByLibrary.simpleMessage(
+            "Kini gambar anda akan sentiasa terurus"),
+        "picpics_photo_manager":
+            MessageLookupByLibrary.simpleMessage("picPics - Pengurus Gambar"),
+        "premium_modal_description": m2,
+        "premium_modal_get_premium_description":
+            MessageLookupByLibrary.simpleMessage("SEMUA CIRI TANPA IKLAN"),
+        "premium_modal_get_premium_title":
+            MessageLookupByLibrary.simpleMessage("Dapatkan Akaun Premium"),
+        "premium_modal_watch_ad": MessageLookupByLibrary.simpleMessage(
+            "Tontoni iklan video untuk teruskan"),
+        "privacy_policy": MessageLookupByLibrary.simpleMessage("Dasar Privasi"),
+        "private_photos": MessageLookupByLibrary.simpleMessage("Foto Peribadi"),
+        "protect_with_encryption": MessageLookupByLibrary.simpleMessage(
+            "Lindungi gambar peribadi anda dengan penyulitan yang hanya dapat diakses dengan kata laluan PIN."),
+        "rate_this_app":
+            MessageLookupByLibrary.simpleMessage("Nilaikan aplikasi ini"),
+        "recent_tags": MessageLookupByLibrary.simpleMessage("Tag Terkini"),
+        "require_secret_key":
+            MessageLookupByLibrary.simpleMessage("Memerlukan kunci rahsia"),
+        "restore_purchase":
+            MessageLookupByLibrary.simpleMessage("Pulihkan pembelian"),
+        "save": MessageLookupByLibrary.simpleMessage("simpan"),
+        "save_location": MessageLookupByLibrary.simpleMessage("Simpan lokasi"),
+        "screenshots_tag":
+            MessageLookupByLibrary.simpleMessage("Tangkapan Skrin"),
+        "search": MessageLookupByLibrary.simpleMessage("Cari..."),
+        "search_all_tags_not_found": MessageLookupByLibrary.simpleMessage(
+            "Tiada gambar dijumpai dengan semua tag padanya"),
+        "search_results":
+            MessageLookupByLibrary.simpleMessage("Keputusan carian"),
+        "secret_key_created": MessageLookupByLibrary.simpleMessage(
+            "Kunci rahsia berjaya dicipta!"),
+        "secret_photos": MessageLookupByLibrary.simpleMessage("Gambar Rahsia"),
+        "selfies_tag": MessageLookupByLibrary.simpleMessage("Swafoto"),
+        "settings": MessageLookupByLibrary.simpleMessage("Tetapan"),
+        "share_with_friends":
+            MessageLookupByLibrary.simpleMessage("Kongsi bersama rakan"),
+        "sign": MessageLookupByLibrary.simpleMessage("Tanda"),
+        "sports_tag": MessageLookupByLibrary.simpleMessage("Sukan"),
+        "start": MessageLookupByLibrary.simpleMessage("Mulakan"),
+        "start_tagging": MessageLookupByLibrary.simpleMessage("Mula tag"),
+        "suggestions": MessageLookupByLibrary.simpleMessage("Cadangan"),
+        "tag_multiple_photos_at_once":
+            MessageLookupByLibrary.simpleMessage("ag banyak gambar sekaligus"),
+        "take_a_look":
+            MessageLookupByLibrary.simpleMessage("Lihatlah aplikasi ini!"),
+        "take_a_look_description": m3,
+        "terms_of_use":
+            MessageLookupByLibrary.simpleMessage("Syarat Penggunaan"),
+        "time": MessageLookupByLibrary.simpleMessage("Masa"),
+        "toggle_date": MessageLookupByLibrary.simpleMessage("Tarikh"),
+        "toggle_days": MessageLookupByLibrary.simpleMessage("Hari"),
+        "toggle_months": MessageLookupByLibrary.simpleMessage("Bulan"),
+        "toggle_tags": MessageLookupByLibrary.simpleMessage("Tag"),
+        "travel_tag": MessageLookupByLibrary.simpleMessage("Pengembaraan"),
+        "tsqr_tag": MessageLookupByLibrary.simpleMessage("Times Square"),
+        "tutorial_daily_package": MessageLookupByLibrary.simpleMessage(
+            "Kami membawa pakej harian untuk anda bagi menguruskan perpustakaan anda secara beransur-ansur."),
+        "tutorial_however_you_want": MessageLookupByLibrary.simpleMessage(
+            "Susun gambar anda dengan menambahkan tag, seperti \'keluarga\', \'binatang peliharaan\', atau apa sahaja yang anda mahukan."),
+        "tutorial_just_swipe": MessageLookupByLibrary.simpleMessage(
+            "Setelah menambahkan tag pada gambar anda, leret sahaja untuk pergi ke gambar seterusnya."),
+        "tutorial_multiselect": MessageLookupByLibrary.simpleMessage(
+            "Anda boleh \"ketuk & tahan\" pada foto anda untuk menandai banyak foto sekaligus."),
+        "tutorial_secret": MessageLookupByLibrary.simpleMessage(
+            "Sembunyikan foto peribadi anda dengan perlindungan kod pin, jaga keselamatannya."),
+        "unlimited_private_pics":
+            MessageLookupByLibrary.simpleMessage("Foto peribadi tanpa had"),
+        "vacation_tag": MessageLookupByLibrary.simpleMessage("Percutian"),
+        "view_hidden_photos": MessageLookupByLibrary.simpleMessage(
+            "Untuk melihat gambar tersembunyi anda, lepaskan kunci dalam tetapan aplikasi. Anda boleh membuka kunci dengan PIN anda."),
+        "welcome": MessageLookupByLibrary.simpleMessage("Selamat Datang!"),
+        "work_tag": MessageLookupByLibrary.simpleMessage("Kerja"),
+        "x_minutes": MessageLookupByLibrary.simpleMessage("20 min"),
+        "year": MessageLookupByLibrary.simpleMessage("tahun"),
+        "yes": MessageLookupByLibrary.simpleMessage("Ya"),
+        "you_are_premium":
+            MessageLookupByLibrary.simpleMessage("Anda premium!"),
+        "your_secret_key":
+            MessageLookupByLibrary.simpleMessage("Kunci rahsia anda")
+      };
 }
