@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -19,569 +19,669 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'pt';
 
-  static m0(email) => "Um código de acesso foi enviado para ${email}";
+  static String m0(email) => "Um código de acesso foi enviado para ${email}";
 
-  static m1(howMany) => "${Intl.plural(howMany, zero: 'Nenhuma foto selecionada', one: '1 foto selecionada', other: '${howMany} fotos selecionadas')}";
+  static String m1(howMany) =>
+      "${Intl.plural(howMany, zero: 'Nenhuma foto selecionada', one: '1 foto selecionada', other: '${howMany} fotos selecionadas')}";
 
-  static m2(number) => "Você completo seus ${number} pics diários gratuítos, deseja continuar?";
+  static String m2(number) =>
+      "Você completo seus ${number} pics diários gratuítos, deseja continuar?";
 
-  static m3(url) => "Para organizar suas fotos, veja neste link: ${url}";
+  static String m3(url) => "Para organizar suas fotos, veja neste link: ${url}";
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "access_code" : MessageLookupByLibrary.simpleMessage("Código de Acesso"),
-    "access_code_sent" : m0,
-    "add_multiple_tags" : MessageLookupByLibrary.simpleMessage("Adicione múltiplas tags"),
-    "add_tag" : MessageLookupByLibrary.simpleMessage("Add tag"),
-    "add_tags" : MessageLookupByLibrary.simpleMessage("Add tags"),
-    "aerospace_engineering_tag" : MessageLookupByLibrary.simpleMessage("Engenharia aeroespacial"),
-    "aircraft_tag" : MessageLookupByLibrary.simpleMessage("Aeronave"),
-    "airliner_tag" : MessageLookupByLibrary.simpleMessage("Avião comercial"),
-    "airplane_tag" : MessageLookupByLibrary.simpleMessage("Avião"),
-    "alcohol_tag" : MessageLookupByLibrary.simpleMessage("Álcool"),
-    "all_at_once" : MessageLookupByLibrary.simpleMessage("Organize multiple photos at once"),
-    "all_photos_were_tagged" : MessageLookupByLibrary.simpleMessage("Não há mais nenhuma foto para organizar."),
-    "all_search_tags" : MessageLookupByLibrary.simpleMessage("Todas as tags da pesquisa"),
-    "always" : MessageLookupByLibrary.simpleMessage("Always"),
-    "aquarium_tag" : MessageLookupByLibrary.simpleMessage("Aquário"),
-    "archery_tag" : MessageLookupByLibrary.simpleMessage("Arco e flecha"),
-    "armrest_tag" : MessageLookupByLibrary.simpleMessage("Apoio de braço"),
-    "army_tag" : MessageLookupByLibrary.simpleMessage("Exército"),
-    "ask_photo_library_permission" : MessageLookupByLibrary.simpleMessage("Precisamos de acesso à sua biblioteca de fotos para que você possa começar a organizar suas fotos com picPics. Não se preocupe, seus dados nunca sairão do seu dispositivo!"),
-    "asphalt_tag" : MessageLookupByLibrary.simpleMessage("Asfalto"),
-    "aurora_tag" : MessageLookupByLibrary.simpleMessage("Aurora"),
-    "auto_renewable_first_part" : MessageLookupByLibrary.simpleMessage("A assinatura é "),
-    "auto_renewable_second_part" : MessageLookupByLibrary.simpleMessage("auto-renovável."),
-    "aviation_tag" : MessageLookupByLibrary.simpleMessage("Aviação"),
-    "baby_tag" : MessageLookupByLibrary.simpleMessage("Bebê"),
-    "backpacking_tag" : MessageLookupByLibrary.simpleMessage("Mochilão"),
-    "badminton_tag" : MessageLookupByLibrary.simpleMessage("Badminton"),
-    "bag_tag" : MessageLookupByLibrary.simpleMessage("Bolsa"),
-    "balance_tag" : MessageLookupByLibrary.simpleMessage("Equilíbrio"),
-    "balloon_tag" : MessageLookupByLibrary.simpleMessage("Balão"),
-    "ballroom_tag" : MessageLookupByLibrary.simpleMessage("Salão de baile"),
-    "bangle_tag" : MessageLookupByLibrary.simpleMessage("Pulseira"),
-    "bangs_tag" : MessageLookupByLibrary.simpleMessage("Franja"),
-    "bar_tag" : MessageLookupByLibrary.simpleMessage("Bar"),
-    "barn_tag" : MessageLookupByLibrary.simpleMessage("Celeiro"),
-    "basset_hound_tag" : MessageLookupByLibrary.simpleMessage("Basset hound"),
-    "bathing_tag" : MessageLookupByLibrary.simpleMessage("Banho"),
-    "bathroom_tag" : MessageLookupByLibrary.simpleMessage("Banheiro"),
-    "beach_tag" : MessageLookupByLibrary.simpleMessage("Praia"),
-    "beanie_tag" : MessageLookupByLibrary.simpleMessage("Gorro"),
-    "bear_tag" : MessageLookupByLibrary.simpleMessage("Urso"),
-    "beard_tag" : MessageLookupByLibrary.simpleMessage("Barba"),
-    "bedroom_tag" : MessageLookupByLibrary.simpleMessage("Quarto"),
-    "bench_tag" : MessageLookupByLibrary.simpleMessage("Banco"),
-    "bento_tag" : MessageLookupByLibrary.simpleMessage("Marmita"),
-    "bicycle_tag" : MessageLookupByLibrary.simpleMessage("Bicicleta"),
-    "bird_tag" : MessageLookupByLibrary.simpleMessage("Pássaro"),
-    "blazer_tag" : MessageLookupByLibrary.simpleMessage("Blazer"),
-    "boat_tag" : MessageLookupByLibrary.simpleMessage("Barco"),
-    "bonfire_tag" : MessageLookupByLibrary.simpleMessage("Fogueira"),
-    "boxer_tag" : MessageLookupByLibrary.simpleMessage("Boxer"),
-    "bracelet_tag" : MessageLookupByLibrary.simpleMessage("Pulseira"),
-    "branch_tag" : MessageLookupByLibrary.simpleMessage("Ramo"),
-    "bread_tag" : MessageLookupByLibrary.simpleMessage("Pão"),
-    "brick_tag" : MessageLookupByLibrary.simpleMessage("Tijolo"),
-    "bride_tag" : MessageLookupByLibrary.simpleMessage("Noiva"),
-    "bridge_tag" : MessageLookupByLibrary.simpleMessage("Ponte"),
-    "bridle_tag" : MessageLookupByLibrary.simpleMessage("Refrear"),
-    "brig_tag" : MessageLookupByLibrary.simpleMessage("Brigue"),
-    "building_tag" : MessageLookupByLibrary.simpleMessage("Edifício"),
-    "bull_tag" : MessageLookupByLibrary.simpleMessage("Touro"),
-    "bullfighting_tag" : MessageLookupByLibrary.simpleMessage("Touradas"),
-    "bumper_tag" : MessageLookupByLibrary.simpleMessage("Parachoque"),
-    "bunk_bed_tag" : MessageLookupByLibrary.simpleMessage("Beliche"),
-    "bus_tag" : MessageLookupByLibrary.simpleMessage("Ônibus"),
-    "butterfly_tag" : MessageLookupByLibrary.simpleMessage("Borboleta"),
-    "cabinetry_tag" : MessageLookupByLibrary.simpleMessage("Gabinete"),
-    "cage_tag" : MessageLookupByLibrary.simpleMessage("Cela"),
-    "cairn_terrier_tag" : MessageLookupByLibrary.simpleMessage("Cairn terrier"),
-    "cake_tag" : MessageLookupByLibrary.simpleMessage("Bolo"),
-    "camping_tag" : MessageLookupByLibrary.simpleMessage("Acampamento"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Cancelar"),
-    "cancel_anytime" : MessageLookupByLibrary.simpleMessage("Cancele a qualquer momento"),
-    "canoe_tag" : MessageLookupByLibrary.simpleMessage("Canoa"),
-    "canyon_tag" : MessageLookupByLibrary.simpleMessage("Canyon"),
-    "cap_tag" : MessageLookupByLibrary.simpleMessage("Boné"),
-    "cappuccino_tag" : MessageLookupByLibrary.simpleMessage("Cappuccino"),
-    "car_tag" : MessageLookupByLibrary.simpleMessage("Carro"),
-    "carnival_tag" : MessageLookupByLibrary.simpleMessage("Carnaval"),
-    "casino_tag" : MessageLookupByLibrary.simpleMessage("Casino"),
-    "castle_tag" : MessageLookupByLibrary.simpleMessage("Castelo"),
-    "cat_tag" : MessageLookupByLibrary.simpleMessage("Gato"),
-    "cathedral_tag" : MessageLookupByLibrary.simpleMessage("Catedral"),
-    "cattle_tag" : MessageLookupByLibrary.simpleMessage("Gado"),
-    "cavalier_tag" : MessageLookupByLibrary.simpleMessage("Cavaleiro"),
-    "cave_tag" : MessageLookupByLibrary.simpleMessage("Caverna"),
-    "caving_tag" : MessageLookupByLibrary.simpleMessage("Caving"),
-    "centrepiece_tag" : MessageLookupByLibrary.simpleMessage("Peça central"),
-    "chain_tag" : MessageLookupByLibrary.simpleMessage("Cadeia"),
-    "chair_tag" : MessageLookupByLibrary.simpleMessage("Cadeira"),
-    "cheeseburger_tag" : MessageLookupByLibrary.simpleMessage("Cheeseburger"),
-    "christmas_tag" : MessageLookupByLibrary.simpleMessage("Natal"),
-    "church_tag" : MessageLookupByLibrary.simpleMessage("Igreja"),
-    "circus_tag" : MessageLookupByLibrary.simpleMessage("Circo"),
-    "class_tag" : MessageLookupByLibrary.simpleMessage("Classe"),
-    "cliff_tag" : MessageLookupByLibrary.simpleMessage("Penhasco"),
-    "clipper_tag" : MessageLookupByLibrary.simpleMessage("Cortador"),
-    "clock_tag" : MessageLookupByLibrary.simpleMessage("Relógio"),
-    "close" : MessageLookupByLibrary.simpleMessage("Fechar"),
-    "clown_tag" : MessageLookupByLibrary.simpleMessage("Palhaço"),
-    "coffee_tag" : MessageLookupByLibrary.simpleMessage("Café"),
-    "cola_tag" : MessageLookupByLibrary.simpleMessage("Cola"),
-    "comet_tag" : MessageLookupByLibrary.simpleMessage("Cometa"),
-    "comics_tag" : MessageLookupByLibrary.simpleMessage("Quadrinhos"),
-    "community_tag" : MessageLookupByLibrary.simpleMessage("Comunidade"),
-    "computer_tag" : MessageLookupByLibrary.simpleMessage("Computador"),
-    "concert_tag" : MessageLookupByLibrary.simpleMessage("Concerto"),
-    "confirm_email" : MessageLookupByLibrary.simpleMessage("Confirm your registration email to receive your access code"),
-    "confirm_secret_key" : MessageLookupByLibrary.simpleMessage("Confirme a chave"),
-    "construction_tag" : MessageLookupByLibrary.simpleMessage("Construção"),
-    "continue_string" : MessageLookupByLibrary.simpleMessage("Continuar"),
-    "cookie_tag" : MessageLookupByLibrary.simpleMessage("Biscoito"),
-    "cookware_and_bakeware_tag" : MessageLookupByLibrary.simpleMessage("Panelas e assadeiras"),
-    "cool_tag" : MessageLookupByLibrary.simpleMessage("Legal"),
-    "cotton_tag" : MessageLookupByLibrary.simpleMessage("Algodão"),
-    "couch_tag" : MessageLookupByLibrary.simpleMessage("Sofá"),
-    "countertop_tag" : MessageLookupByLibrary.simpleMessage("Bancada"),
-    "country" : MessageLookupByLibrary.simpleMessage("país"),
-    "couscous_tag" : MessageLookupByLibrary.simpleMessage("Cuscuz"),
-    "crew_tag" : MessageLookupByLibrary.simpleMessage("Tripulação"),
-    "crochet_tag" : MessageLookupByLibrary.simpleMessage("Crochê"),
-    "crocodile_tag" : MessageLookupByLibrary.simpleMessage("Crocodilo"),
-    "crowd_tag" : MessageLookupByLibrary.simpleMessage("Multidão"),
-    "cuisine_tag" : MessageLookupByLibrary.simpleMessage("Cozinha"),
-    "cup_tag" : MessageLookupByLibrary.simpleMessage("Copo"),
-    "curling_tag" : MessageLookupByLibrary.simpleMessage("Curling"),
-    "curtain_tag" : MessageLookupByLibrary.simpleMessage("Cortina"),
-    "cushion_tag" : MessageLookupByLibrary.simpleMessage("Almofada"),
-    "cutlery_tag" : MessageLookupByLibrary.simpleMessage("Talheres"),
-    "cycling_tag" : MessageLookupByLibrary.simpleMessage("Ciclismo"),
-    "daily_challenge" : MessageLookupByLibrary.simpleMessage("Desafio diário"),
-    "daily_challenge_permission_description" : MessageLookupByLibrary.simpleMessage("Para podermos enviar seus desafios diários, precisamos de autorização para enviar notificações; portanto, é necessário que você autorize as notificações nas opções do seu telefone celular."),
-    "daily_goal" : MessageLookupByLibrary.simpleMessage("Objetivo diário"),
-    "daily_notification_description" : MessageLookupByLibrary.simpleMessage("Está na hora de completar o seu desáfio. Entre no picPics para completá-lo!"),
-    "daily_notification_title" : MessageLookupByLibrary.simpleMessage("Desafio diário"),
-    "dalmatian_tag" : MessageLookupByLibrary.simpleMessage("Dálmata"),
-    "dam_tag" : MessageLookupByLibrary.simpleMessage("Barragem"),
-    "dance_tag" : MessageLookupByLibrary.simpleMessage("Dança"),
-    "deejay_tag" : MessageLookupByLibrary.simpleMessage("DJ"),
-    "delete" : MessageLookupByLibrary.simpleMessage("Deletar"),
-    "denim_tag" : MessageLookupByLibrary.simpleMessage("Jeans"),
-    "desert_tag" : MessageLookupByLibrary.simpleMessage("Deserto"),
-    "desk_tag" : MessageLookupByLibrary.simpleMessage("Mesa"),
-    "device_has_no_pics" : MessageLookupByLibrary.simpleMessage("Esse dispositivo não tem nenhuma foto na galeria, assim não há nenhuma foto que possa ser taggeada."),
-    "dinosaur_tag" : MessageLookupByLibrary.simpleMessage("Dinosauro"),
-    "disable_secret" : MessageLookupByLibrary.simpleMessage("Deseja desocultar esta foto?"),
-    "dog_tag" : MessageLookupByLibrary.simpleMessage("Cachorro"),
-    "doily_tag" : MessageLookupByLibrary.simpleMessage("Doily"),
-    "dont_ask_again" : MessageLookupByLibrary.simpleMessage("Não pergunte mais"),
-    "dragon_tag" : MessageLookupByLibrary.simpleMessage("Dragão"),
-    "drawer_tag" : MessageLookupByLibrary.simpleMessage("Gaveta"),
-    "dress_tag" : MessageLookupByLibrary.simpleMessage("Vestido"),
-    "duck_tag" : MessageLookupByLibrary.simpleMessage("Pato"),
-    "dude_tag" : MessageLookupByLibrary.simpleMessage("Cara"),
-    "dune_tag" : MessageLookupByLibrary.simpleMessage("Duna"),
-    "ear_tag" : MessageLookupByLibrary.simpleMessage("Orelha"),
-    "eating_tag" : MessageLookupByLibrary.simpleMessage("Comendo"),
-    "edit_tag" : MessageLookupByLibrary.simpleMessage("Editar tag"),
-    "email" : MessageLookupByLibrary.simpleMessage("Email"),
-    "enable_faceid" : MessageLookupByLibrary.simpleMessage("Habilitar Face ID"),
-    "enable_fingerprint" : MessageLookupByLibrary.simpleMessage("Habilitar Impressão Digital"),
-    "enable_irisscanner" : MessageLookupByLibrary.simpleMessage("Habilitar Scanner de Iris"),
-    "enable_touchid" : MessageLookupByLibrary.simpleMessage("Habilitar Touch ID"),
-    "event_tag" : MessageLookupByLibrary.simpleMessage("Evento"),
-    "export_all_gallery" : MessageLookupByLibrary.simpleMessage("Exporte toda galeria"),
-    "export_library" : MessageLookupByLibrary.simpleMessage("Exportar biblioteca"),
-    "eyelash_tag" : MessageLookupByLibrary.simpleMessage("Cílios"),
-    "factory_tag" : MessageLookupByLibrary.simpleMessage("Fábrica"),
-    "family_tag" : MessageLookupByLibrary.simpleMessage("Família"),
-    "farm_tag" : MessageLookupByLibrary.simpleMessage("Fazenda"),
-    "fast_food_tag" : MessageLookupByLibrary.simpleMessage("Fast food"),
-    "feedback_bug_report" : MessageLookupByLibrary.simpleMessage("Feedback & Bugs"),
-    "ferris_wheel_tag" : MessageLookupByLibrary.simpleMessage("Roda gigante"),
-    "fiction_tag" : MessageLookupByLibrary.simpleMessage("Ficção"),
-    "field_tag" : MessageLookupByLibrary.simpleMessage("Campo"),
-    "find_easily" : MessageLookupByLibrary.simpleMessage("Easily find your photos"),
-    "fire_tag" : MessageLookupByLibrary.simpleMessage("Fogo"),
-    "fireworks_tag" : MessageLookupByLibrary.simpleMessage("Fogos de artifício"),
-    "fishing_tag" : MessageLookupByLibrary.simpleMessage("Pesca"),
-    "flag_tag" : MessageLookupByLibrary.simpleMessage("Sinalizar"),
-    "flap_tag" : MessageLookupByLibrary.simpleMessage("Flap"),
-    "flesh_tag" : MessageLookupByLibrary.simpleMessage("Corpo"),
-    "flora_tag" : MessageLookupByLibrary.simpleMessage("Flora"),
-    "flower_tag" : MessageLookupByLibrary.simpleMessage("Flor"),
-    "flowerpot_tag" : MessageLookupByLibrary.simpleMessage("Vaso de flores"),
-    "fog_tag" : MessageLookupByLibrary.simpleMessage("Névoa"),
-    "food_tag" : MessageLookupByLibrary.simpleMessage("Comida"),
-    "foods_tag" : MessageLookupByLibrary.simpleMessage("Comidas"),
-    "foot_tag" : MessageLookupByLibrary.simpleMessage("Pé"),
-    "forest_tag" : MessageLookupByLibrary.simpleMessage("Floresta"),
-    "forgot_secret_key" : MessageLookupByLibrary.simpleMessage("Forgot secret key?"),
-    "frigate_tag" : MessageLookupByLibrary.simpleMessage("Fragata"),
-    "fruit_tag" : MessageLookupByLibrary.simpleMessage("Frutas"),
-    "full_screen" : MessageLookupByLibrary.simpleMessage("Tela cheia"),
-    "fun_tag" : MessageLookupByLibrary.simpleMessage("Diversão"),
-    "fur_tag" : MessageLookupByLibrary.simpleMessage("Casaco de pele"),
-    "gallery_access_permission" : MessageLookupByLibrary.simpleMessage("Permissão de acesso"),
-    "gallery_access_permission_description" : MessageLookupByLibrary.simpleMessage("Para começarmos a organizar suas fotos, precisamos de autorização para acessá-las\""),
-    "gallery_access_reason" : MessageLookupByLibrary.simpleMessage("Para organizar suas fotos precisamos de acesso a sua galeria"),
-    "garden_tag" : MessageLookupByLibrary.simpleMessage("Jardim"),
-    "gelato_tag" : MessageLookupByLibrary.simpleMessage("Sorvete"),
-    "gerbil_tag" : MessageLookupByLibrary.simpleMessage("Gerbil"),
-    "get_premium_description" : MessageLookupByLibrary.simpleMessage("PARA TER TODOS OS RECURSOS"),
-    "get_premium_now" : MessageLookupByLibrary.simpleMessage("Seja premium agora!"),
-    "get_premium_title" : MessageLookupByLibrary.simpleMessage("SEJA PREMIUM"),
-    "glacier_tag" : MessageLookupByLibrary.simpleMessage("Geleira"),
-    "glasses_tag" : MessageLookupByLibrary.simpleMessage("Óculos"),
-    "glitter_tag" : MessageLookupByLibrary.simpleMessage("Glitter"),
-    "goggles_tag" : MessageLookupByLibrary.simpleMessage("Óculos"),
-    "gown_tag" : MessageLookupByLibrary.simpleMessage("Vestido"),
-    "graduation_tag" : MessageLookupByLibrary.simpleMessage("Graduação"),
-    "grandparent_tag" : MessageLookupByLibrary.simpleMessage("Avós"),
-    "groom_tag" : MessageLookupByLibrary.simpleMessage("Noivo"),
-    "gymnastics_tag" : MessageLookupByLibrary.simpleMessage("Ginástica"),
-    "hair_tag" : MessageLookupByLibrary.simpleMessage("Cabelo"),
-    "hand_tag" : MessageLookupByLibrary.simpleMessage("Mão"),
-    "handbag_tag" : MessageLookupByLibrary.simpleMessage("Bolsa"),
-    "handrail_tag" : MessageLookupByLibrary.simpleMessage("Corrimão"),
-    "hanukkah_tag" : MessageLookupByLibrary.simpleMessage("Hanukkah"),
-    "hat_tag" : MessageLookupByLibrary.simpleMessage("Chapéu"),
-    "heart_tag" : MessageLookupByLibrary.simpleMessage("Coração"),
-    "helicopter_tag" : MessageLookupByLibrary.simpleMessage("Helicóptero"),
-    "helmet_tag" : MessageLookupByLibrary.simpleMessage("Capacete"),
-    "herd_tag" : MessageLookupByLibrary.simpleMessage("Rebanho"),
-    "himalayan_tag" : MessageLookupByLibrary.simpleMessage("Himalaia"),
-    "home_tag" : MessageLookupByLibrary.simpleMessage("Casa"),
-    "horn_tag" : MessageLookupByLibrary.simpleMessage("Chifre"),
-    "horse_tag" : MessageLookupByLibrary.simpleMessage("Cavalo"),
-    "hot_dog_tag" : MessageLookupByLibrary.simpleMessage("Cachorro-quente"),
-    "how_many_pics" : MessageLookupByLibrary.simpleMessage("Quantas fotos?"),
-    "ice_tag" : MessageLookupByLibrary.simpleMessage("Gelo"),
-    "iceberg_tag" : MessageLookupByLibrary.simpleMessage("Iceberg"),
-    "icicle_tag" : MessageLookupByLibrary.simpleMessage("Icicle"),
-    "icing_tag" : MessageLookupByLibrary.simpleMessage("Glacê"),
-    "icon_tag" : MessageLookupByLibrary.simpleMessage("Ícone"),
-    "infinite_tags" : MessageLookupByLibrary.simpleMessage("Tags a vontade"),
-    "infrastructure_tag" : MessageLookupByLibrary.simpleMessage("Infraestrutura"),
-    "insect_tag" : MessageLookupByLibrary.simpleMessage("Inseto"),
-    "jacket_tag" : MessageLookupByLibrary.simpleMessage("Jaqueta"),
-    "jeans_tag" : MessageLookupByLibrary.simpleMessage("Jeans"),
-    "jersey_tag" : MessageLookupByLibrary.simpleMessage("Camisola"),
-    "jewellery_tag" : MessageLookupByLibrary.simpleMessage("Joalharia"),
-    "joker_tag" : MessageLookupByLibrary.simpleMessage("Palhaço"),
-    "juice_tag" : MessageLookupByLibrary.simpleMessage("Juice"),
-    "jungle_tag" : MessageLookupByLibrary.simpleMessage("Selva"),
-    "junk_tag" : MessageLookupByLibrary.simpleMessage("Lixo"),
-    "kayak_tag" : MessageLookupByLibrary.simpleMessage("Caiaque"),
-    "keep_asking" : MessageLookupByLibrary.simpleMessage("Pergunte sempre"),
-    "keep_safe" : MessageLookupByLibrary.simpleMessage("Sua foto agora está segura no picPics. Deseja excluí-la do rolo da câmera?"),
-    "kitchen_tag" : MessageLookupByLibrary.simpleMessage("Cozinha"),
-    "knitting_tag" : MessageLookupByLibrary.simpleMessage("Tricô"),
-    "lake_tag" : MessageLookupByLibrary.simpleMessage("Lago"),
-    "lampshade_tag" : MessageLookupByLibrary.simpleMessage("Abajur"),
-    "language" : MessageLookupByLibrary.simpleMessage("Lingua"),
-    "larva_tag" : MessageLookupByLibrary.simpleMessage("Larva"),
-    "laugh_tag" : MessageLookupByLibrary.simpleMessage("Risada"),
-    "leather_tag" : MessageLookupByLibrary.simpleMessage("Couro"),
-    "leggings_tag" : MessageLookupByLibrary.simpleMessage("Leggings"),
-    "lego_tag" : MessageLookupByLibrary.simpleMessage("Lego"),
-    "leisure_tag" : MessageLookupByLibrary.simpleMessage("Lazer"),
-    "lighthouse_tag" : MessageLookupByLibrary.simpleMessage("Farol"),
-    "lightning_tag" : MessageLookupByLibrary.simpleMessage("Relâmpago"),
-    "lipstick_tag" : MessageLookupByLibrary.simpleMessage("Batom"),
-    "lock_with_pin" : MessageLookupByLibrary.simpleMessage("Lock down your private photos with a PIN passcode."),
-    "lock_your_photos" : MessageLookupByLibrary.simpleMessage("Lock your photos"),
-    "longboard_tag" : MessageLookupByLibrary.simpleMessage("Longboard"),
-    "love_tag" : MessageLookupByLibrary.simpleMessage("Amor"),
-    "loveseat_tag" : MessageLookupByLibrary.simpleMessage("Loveseat"),
-    "lunch_tag" : MessageLookupByLibrary.simpleMessage("Almoço"),
-    "marathon_tag" : MessageLookupByLibrary.simpleMessage("Maratona"),
-    "meal_tag" : MessageLookupByLibrary.simpleMessage("Refeição"),
-    "menu_tag" : MessageLookupByLibrary.simpleMessage("Menu"),
-    "metal_tag" : MessageLookupByLibrary.simpleMessage("Metal"),
-    "militia_tag" : MessageLookupByLibrary.simpleMessage("Milícia"),
-    "miniature_tag" : MessageLookupByLibrary.simpleMessage("Miniatura"),
-    "mobile_phone_tag" : MessageLookupByLibrary.simpleMessage("Telefone celular"),
-    "model_tag" : MessageLookupByLibrary.simpleMessage("Modelo"),
-    "money_tag" : MessageLookupByLibrary.simpleMessage("Dinheiro"),
-    "monochrome_tag" : MessageLookupByLibrary.simpleMessage("Monocromático"),
-    "month" : MessageLookupByLibrary.simpleMessage("mês"),
-    "monument_tag" : MessageLookupByLibrary.simpleMessage("Monumento"),
-    "moon_tag" : MessageLookupByLibrary.simpleMessage("Lua"),
-    "mortarboard_tag" : MessageLookupByLibrary.simpleMessage("Mortarboard"),
-    "mosque_tag" : MessageLookupByLibrary.simpleMessage("Mesquita"),
-    "motorcycle_tag" : MessageLookupByLibrary.simpleMessage("motocicleta"),
-    "mountain_tag" : MessageLookupByLibrary.simpleMessage("Montanha"),
-    "moustache_tag" : MessageLookupByLibrary.simpleMessage("Bigode"),
-    "mouth_tag" : MessageLookupByLibrary.simpleMessage("Boca"),
-    "mufti_tag" : MessageLookupByLibrary.simpleMessage("À paisana"),
-    "muscle_tag" : MessageLookupByLibrary.simpleMessage("Músculo"),
-    "museum_tag" : MessageLookupByLibrary.simpleMessage("Museu"),
-    "musical_instrument_tag" : MessageLookupByLibrary.simpleMessage("Instrumento musical"),
-    "musical_tag" : MessageLookupByLibrary.simpleMessage("Musical"),
-    "musician_tag" : MessageLookupByLibrary.simpleMessage("Músico"),
-    "nail_tag" : MessageLookupByLibrary.simpleMessage("Unha"),
-    "nebula_tag" : MessageLookupByLibrary.simpleMessage("Nebulosa"),
-    "necklace_tag" : MessageLookupByLibrary.simpleMessage("Colar"),
-    "neon_tag" : MessageLookupByLibrary.simpleMessage("Neon"),
-    "net_tag" : MessageLookupByLibrary.simpleMessage("Rede"),
-    "new_secret_key" : MessageLookupByLibrary.simpleMessage("Nova chave secreta"),
-    "news_tag" : MessageLookupByLibrary.simpleMessage("Notícias"),
-    "newspaper_tag" : MessageLookupByLibrary.simpleMessage("Jornal"),
-    "next" : MessageLookupByLibrary.simpleMessage("Próximo"),
-    "nightclub_tag" : MessageLookupByLibrary.simpleMessage("Boate"),
-    "no" : MessageLookupByLibrary.simpleMessage("Não"),
-    "no_ads" : MessageLookupByLibrary.simpleMessage("Sem propagandas"),
-    "no_previous_purchase" : MessageLookupByLibrary.simpleMessage("Nenhuma compra encontrada"),
-    "no_tagged_photos" : MessageLookupByLibrary.simpleMessage("Você não tem nenhuma foto taggeada ainda"),
-    "no_tags_found" : MessageLookupByLibrary.simpleMessage("Nenhuma tag encontrada"),
-    "no_valid_subscription" : MessageLookupByLibrary.simpleMessage("Não foi possível encontrar uma compra de assinatura válida."),
-    "notification_time" : MessageLookupByLibrary.simpleMessage("Horário da notificação"),
-    "notifications" : MessageLookupByLibrary.simpleMessage("Notificações"),
-    "ny_tag" : MessageLookupByLibrary.simpleMessage("NY"),
-    "odometer_tag" : MessageLookupByLibrary.simpleMessage("Odômetro"),
-    "ok" : MessageLookupByLibrary.simpleMessage("OK"),
-    "open_gallery" : MessageLookupByLibrary.simpleMessage("Abrir galeria"),
-    "organized_photos_description" : MessageLookupByLibrary.simpleMessage("Fotos que já foram taggeadas"),
-    "organized_photos_title" : MessageLookupByLibrary.simpleMessage("Fotos organizadas"),
-    "outerwear_tag" : MessageLookupByLibrary.simpleMessage("Agasalhos"),
-    "palace_tag" : MessageLookupByLibrary.simpleMessage("Palácio"),
-    "paper_tag" : MessageLookupByLibrary.simpleMessage("Papel"),
-    "park_tag" : MessageLookupByLibrary.simpleMessage("Parque"),
-    "parties_tag" : MessageLookupByLibrary.simpleMessage("Festas"),
-    "party_tag" : MessageLookupByLibrary.simpleMessage("Festa"),
-    "passport_tag" : MessageLookupByLibrary.simpleMessage("Passaporte"),
-    "pasteles_tag" : MessageLookupByLibrary.simpleMessage("Pasteles"),
-    "pattern_tag" : MessageLookupByLibrary.simpleMessage("Padrôes"),
-    "penguin_tag" : MessageLookupByLibrary.simpleMessage("Pinguim"),
-    "person_tag" : MessageLookupByLibrary.simpleMessage("Pessoa"),
-    "pet_tag" : MessageLookupByLibrary.simpleMessage("Pet"),
-    "petal_tag" : MessageLookupByLibrary.simpleMessage("Pétala"),
-    "pets_tag" : MessageLookupByLibrary.simpleMessage("Pets"),
-    "pho_tag" : MessageLookupByLibrary.simpleMessage("Pho"),
-    "photo_gallery_count" : m1,
-    "photo_gallery_description" : MessageLookupByLibrary.simpleMessage("Fotos não organizadas"),
-    "photo_gallery_title" : MessageLookupByLibrary.simpleMessage("Galeria de fotos"),
-    "photo_location" : MessageLookupByLibrary.simpleMessage("Local da foto"),
-    "photos_always_organized" : MessageLookupByLibrary.simpleMessage("Agora suas fotos estarão sempre organizadas"),
-    "piano_tag" : MessageLookupByLibrary.simpleMessage("Piano"),
-    "picnic_tag" : MessageLookupByLibrary.simpleMessage("Piquenique"),
-    "picpics_photo_manager" : MessageLookupByLibrary.simpleMessage("picPics - Photo Manager"),
-    "pie_tag" : MessageLookupByLibrary.simpleMessage("Torta"),
-    "pier_tag" : MessageLookupByLibrary.simpleMessage("Pier"),
-    "pillow_tag" : MessageLookupByLibrary.simpleMessage("Travesseiro"),
-    "pitch_tag" : MessageLookupByLibrary.simpleMessage("Arremesso"),
-    "pixie_bob_tag" : MessageLookupByLibrary.simpleMessage("Pixie-bob"),
-    "pizza_tag" : MessageLookupByLibrary.simpleMessage("Pizza"),
-    "placemat_tag" : MessageLookupByLibrary.simpleMessage("Descanso de mesa"),
-    "plant_tag" : MessageLookupByLibrary.simpleMessage("Planta"),
-    "playground_tag" : MessageLookupByLibrary.simpleMessage("Recreio"),
-    "plush_tag" : MessageLookupByLibrary.simpleMessage("Pelúcia"),
-    "pocket_tag" : MessageLookupByLibrary.simpleMessage("Bolso"),
-    "polo_tag" : MessageLookupByLibrary.simpleMessage("Pólo"),
-    "pomacentridae_tag" : MessageLookupByLibrary.simpleMessage("Pomacentridae"),
-    "pool_tag" : MessageLookupByLibrary.simpleMessage("Piscina"),
-    "pop_music_tag" : MessageLookupByLibrary.simpleMessage("Música Pop"),
-    "porcelain_tag" : MessageLookupByLibrary.simpleMessage("Porcelana"),
-    "poster_tag" : MessageLookupByLibrary.simpleMessage("Poster"),
-    "prairie_tag" : MessageLookupByLibrary.simpleMessage("Pradaria"),
-    "premium_modal_description" : m2,
-    "premium_modal_get_premium_description" : MessageLookupByLibrary.simpleMessage("Todos os recursos e sem anuncios"),
-    "premium_modal_get_premium_title" : MessageLookupByLibrary.simpleMessage("Obter conta premium"),
-    "premium_modal_watch_ad" : MessageLookupByLibrary.simpleMessage("Assista ao anúncio para continuar"),
-    "presentation_tag" : MessageLookupByLibrary.simpleMessage("Apresentação"),
-    "privacy_policy" : MessageLookupByLibrary.simpleMessage("Política de privacidade"),
-    "private_photos" : MessageLookupByLibrary.simpleMessage("Fotos privadas"),
-    "product_tag" : MessageLookupByLibrary.simpleMessage("Produto"),
-    "prom_tag" : MessageLookupByLibrary.simpleMessage("Baile"),
-    "protect_with_encryption" : MessageLookupByLibrary.simpleMessage("Protect your personal photos with encryption only accessible with a PIN password."),
-    "quilting_tag" : MessageLookupByLibrary.simpleMessage("Quilting"),
-    "race_tag" : MessageLookupByLibrary.simpleMessage("Corrida"),
-    "racing_tag" : MessageLookupByLibrary.simpleMessage("Corrida"),
-    "rafting_tag" : MessageLookupByLibrary.simpleMessage("Rafting"),
-    "ragdoll_tag" : MessageLookupByLibrary.simpleMessage("Ragdoll"),
-    "rainbow_tag" : MessageLookupByLibrary.simpleMessage("Arco Iris"),
-    "ranch_tag" : MessageLookupByLibrary.simpleMessage("Rancho"),
-    "rate_this_app" : MessageLookupByLibrary.simpleMessage("Avalie esse app"),
-    "receipt_tag" : MessageLookupByLibrary.simpleMessage("Recibo"),
-    "recent_tags" : MessageLookupByLibrary.simpleMessage("Tags recentes"),
-    "reef_tag" : MessageLookupByLibrary.simpleMessage("Recife"),
-    "rein_tag" : MessageLookupByLibrary.simpleMessage("Rédea"),
-    "require_secret_key" : MessageLookupByLibrary.simpleMessage("Require secret key"),
-    "restore_purchase" : MessageLookupByLibrary.simpleMessage("Restaurar compra"),
-    "rickshaw_tag" : MessageLookupByLibrary.simpleMessage("Riquixá"),
-    "ring_tag" : MessageLookupByLibrary.simpleMessage("Toque"),
-    "river_tag" : MessageLookupByLibrary.simpleMessage("Rio"),
-    "road_tag" : MessageLookupByLibrary.simpleMessage("Caminho"),
-    "rock_tag" : MessageLookupByLibrary.simpleMessage("Pedra"),
-    "rocket_tag" : MessageLookupByLibrary.simpleMessage("Foguete"),
-    "rodeo_tag" : MessageLookupByLibrary.simpleMessage("Rodeio"),
-    "roller_tag" : MessageLookupByLibrary.simpleMessage("Roller"),
-    "roof_tag" : MessageLookupByLibrary.simpleMessage("Cobertura"),
-    "room_tag" : MessageLookupByLibrary.simpleMessage("Sala"),
-    "rowing_tag" : MessageLookupByLibrary.simpleMessage("Remo"),
-    "rugby_tag" : MessageLookupByLibrary.simpleMessage("Rugby"),
-    "ruins_tag" : MessageLookupByLibrary.simpleMessage("Ruínas"),
-    "running_tag" : MessageLookupByLibrary.simpleMessage("Corrida"),
-    "safari_tag" : MessageLookupByLibrary.simpleMessage("Safari"),
-    "sailboat_tag" : MessageLookupByLibrary.simpleMessage("Veleiro"),
-    "sand_tag" : MessageLookupByLibrary.simpleMessage("Areia"),
-    "santa_claus_tag" : MessageLookupByLibrary.simpleMessage("Papai Noel"),
-    "sari_tag" : MessageLookupByLibrary.simpleMessage("Sari"),
-    "saucer_tag" : MessageLookupByLibrary.simpleMessage("Pires"),
-    "save" : MessageLookupByLibrary.simpleMessage("economize"),
-    "save_location" : MessageLookupByLibrary.simpleMessage("Confirmar localização"),
-    "scarf_tag" : MessageLookupByLibrary.simpleMessage("Cachecol"),
-    "school_tag" : MessageLookupByLibrary.simpleMessage("Escola"),
-    "screenshot_tag" : MessageLookupByLibrary.simpleMessage("Screenshot"),
-    "screenshots_tag" : MessageLookupByLibrary.simpleMessage("Screenshots"),
-    "scuba_diving_tag" : MessageLookupByLibrary.simpleMessage("Mergulho"),
-    "seal_tag" : MessageLookupByLibrary.simpleMessage("Selo"),
-    "search" : MessageLookupByLibrary.simpleMessage("Pesquisar..."),
-    "search_all_tags_not_found" : MessageLookupByLibrary.simpleMessage("Nenhuma foto foi encontrada com todas as tags."),
-    "search_results" : MessageLookupByLibrary.simpleMessage("Resultados da busca"),
-    "secret_key_created" : MessageLookupByLibrary.simpleMessage("Chave secreta criada com sucesso!"),
-    "secret_photos" : MessageLookupByLibrary.simpleMessage("Fotos Secretas"),
-    "selfie_tag" : MessageLookupByLibrary.simpleMessage("Selfie"),
-    "selfies_tag" : MessageLookupByLibrary.simpleMessage("Selfies"),
-    "settings" : MessageLookupByLibrary.simpleMessage("Opções"),
-    "share_with_friends" : MessageLookupByLibrary.simpleMessage("Compartilhar"),
-    "shelf_tag" : MessageLookupByLibrary.simpleMessage("Prateleira"),
-    "shell_tag" : MessageLookupByLibrary.simpleMessage("Concha"),
-    "shetland_sheepdog_tag" : MessageLookupByLibrary.simpleMessage("Pastor de shetland"),
-    "shikoku_tag" : MessageLookupByLibrary.simpleMessage("Shikoku"),
-    "shipwreck_tag" : MessageLookupByLibrary.simpleMessage("Naufrágio"),
-    "shoe_tag" : MessageLookupByLibrary.simpleMessage("Sapato"),
-    "shorts_tag" : MessageLookupByLibrary.simpleMessage("Shorts"),
-    "sign" : MessageLookupByLibrary.simpleMessage("Por"),
-    "singer_tag" : MessageLookupByLibrary.simpleMessage("Cantor"),
-    "sink_tag" : MessageLookupByLibrary.simpleMessage("Pia"),
-    "sitting_tag" : MessageLookupByLibrary.simpleMessage("Sentado"),
-    "skateboard_tag" : MessageLookupByLibrary.simpleMessage("Skateboard"),
-    "skateboarder_tag" : MessageLookupByLibrary.simpleMessage("Skatista"),
-    "skiff_tag" : MessageLookupByLibrary.simpleMessage("Esquife"),
-    "skin_tag" : MessageLookupByLibrary.simpleMessage("Pele"),
-    "sky_tag" : MessageLookupByLibrary.simpleMessage("Céu"),
-    "skyline_tag" : MessageLookupByLibrary.simpleMessage("Horizonte"),
-    "skyscraper_tag" : MessageLookupByLibrary.simpleMessage("Arranha-céu"),
-    "sled_tag" : MessageLookupByLibrary.simpleMessage("Trenó"),
-    "sledding_tag" : MessageLookupByLibrary.simpleMessage("Trenó"),
-    "sleep_tag" : MessageLookupByLibrary.simpleMessage("Dormir"),
-    "smile_tag" : MessageLookupByLibrary.simpleMessage("Sorriso"),
-    "sneakers_tag" : MessageLookupByLibrary.simpleMessage("Tênis"),
-    "snorkeling_tag" : MessageLookupByLibrary.simpleMessage("Snorkeling"),
-    "snowboarding_tag" : MessageLookupByLibrary.simpleMessage("Snowboarding"),
-    "soccer_tag" : MessageLookupByLibrary.simpleMessage("Futebol"),
-    "softball_tag" : MessageLookupByLibrary.simpleMessage("Softball"),
-    "soil_tag" : MessageLookupByLibrary.simpleMessage("Solo"),
-    "song_tag" : MessageLookupByLibrary.simpleMessage("Música"),
-    "space_tag" : MessageLookupByLibrary.simpleMessage("Espaço"),
-    "sparkler_tag" : MessageLookupByLibrary.simpleMessage("Sparkler"),
-    "speedboat_tag" : MessageLookupByLibrary.simpleMessage("Lancha"),
-    "sphynx_tag" : MessageLookupByLibrary.simpleMessage("Sphynx"),
-    "sports_tag" : MessageLookupByLibrary.simpleMessage("Sports"),
-    "stadium_tag" : MessageLookupByLibrary.simpleMessage("Estádio"),
-    "stairs_tag" : MessageLookupByLibrary.simpleMessage("Escadas"),
-    "standing_tag" : MessageLookupByLibrary.simpleMessage("De pé"),
-    "star_tag" : MessageLookupByLibrary.simpleMessage("Estrela"),
-    "start" : MessageLookupByLibrary.simpleMessage("Começar"),
-    "start_tagging" : MessageLookupByLibrary.simpleMessage("Começe a taggear"),
-    "statue_tag" : MessageLookupByLibrary.simpleMessage("Estátua"),
-    "steaming_tag" : MessageLookupByLibrary.simpleMessage("Fumegante"),
-    "storm_tag" : MessageLookupByLibrary.simpleMessage("Tempestade"),
-    "strap_tag" : MessageLookupByLibrary.simpleMessage("Alça"),
-    "stuffed_toy_tag" : MessageLookupByLibrary.simpleMessage("Brinquedo de pelúcia"),
-    "submarine_tag" : MessageLookupByLibrary.simpleMessage("Submarino"),
-    "subwoofer_tag" : MessageLookupByLibrary.simpleMessage("Subwoofer"),
-    "suggestions" : MessageLookupByLibrary.simpleMessage("Sugestões"),
-    "sunglasses_tag" : MessageLookupByLibrary.simpleMessage("Óculos de sol"),
-    "sunset_tag" : MessageLookupByLibrary.simpleMessage("Pôr do sol"),
-    "superhero_tag" : MessageLookupByLibrary.simpleMessage("Super heroi"),
-    "superman_tag" : MessageLookupByLibrary.simpleMessage("Superman"),
-    "supervillain_tag" : MessageLookupByLibrary.simpleMessage("Super-vilão"),
-    "supper_tag" : MessageLookupByLibrary.simpleMessage("Ceia"),
-    "surfboard_tag" : MessageLookupByLibrary.simpleMessage("Prancha de surf"),
-    "surfing_tag" : MessageLookupByLibrary.simpleMessage("Surfing"),
-    "sushi_tag" : MessageLookupByLibrary.simpleMessage("Sushi"),
-    "swamp_tag" : MessageLookupByLibrary.simpleMessage("Pântano"),
-    "swimming_tag" : MessageLookupByLibrary.simpleMessage("Natação"),
-    "swimwear_tag" : MessageLookupByLibrary.simpleMessage("Roupa de banho"),
-    "swing_tag" : MessageLookupByLibrary.simpleMessage("Balanço"),
-    "tablecloth_tag" : MessageLookupByLibrary.simpleMessage("Toalha de mesa"),
-    "tableware_tag" : MessageLookupByLibrary.simpleMessage("Talheres"),
-    "tag_multiple_photos_at_once" : MessageLookupByLibrary.simpleMessage("Marque várias fotos de uma vez só"),
-    "take_a_look" : MessageLookupByLibrary.simpleMessage("Dê uma olhada nesse aplicativo!"),
-    "take_a_look_description" : m3,
-    "tattoo_tag" : MessageLookupByLibrary.simpleMessage("Tattoo"),
-    "team_tag" : MessageLookupByLibrary.simpleMessage("Time"),
-    "television_tag" : MessageLookupByLibrary.simpleMessage("Televisão"),
-    "temple_tag" : MessageLookupByLibrary.simpleMessage("Templo"),
-    "terms_of_use" : MessageLookupByLibrary.simpleMessage("Termos de uso"),
-    "textile_tag" : MessageLookupByLibrary.simpleMessage("Têxtil"),
-    "thanksgiving_tag" : MessageLookupByLibrary.simpleMessage("Ação de graças"),
-    "tie_tag" : MessageLookupByLibrary.simpleMessage("Empate"),
-    "tights_tag" : MessageLookupByLibrary.simpleMessage("Calças"),
-    "tile_tag" : MessageLookupByLibrary.simpleMessage("Bloco"),
-    "time" : MessageLookupByLibrary.simpleMessage("Horário"),
-    "tire_tag" : MessageLookupByLibrary.simpleMessage("Pneu"),
-    "toe_tag" : MessageLookupByLibrary.simpleMessage("Dedo do pé"),
-    "toggle_date" : MessageLookupByLibrary.simpleMessage("Data"),
-    "toggle_days" : MessageLookupByLibrary.simpleMessage("Dias"),
-    "toggle_months" : MessageLookupByLibrary.simpleMessage("Meses"),
-    "toggle_tags" : MessageLookupByLibrary.simpleMessage("Tags"),
-    "tower_tag" : MessageLookupByLibrary.simpleMessage("Torre"),
-    "toy_tag" : MessageLookupByLibrary.simpleMessage("Brinquedo"),
-    "track_tag" : MessageLookupByLibrary.simpleMessage("Caminho"),
-    "tractor_tag" : MessageLookupByLibrary.simpleMessage("Trator"),
-    "train_tag" : MessageLookupByLibrary.simpleMessage("Trem"),
-    "travel_tag" : MessageLookupByLibrary.simpleMessage("Viagens"),
-    "trunk_tag" : MessageLookupByLibrary.simpleMessage("Tronco"),
-    "tsqr_tag" : MessageLookupByLibrary.simpleMessage("Times Square"),
-    "tubing_tag" : MessageLookupByLibrary.simpleMessage("Tubulação"),
-    "turtle_tag" : MessageLookupByLibrary.simpleMessage("Tartaruga"),
-    "tutorial_daily_package" : MessageLookupByLibrary.simpleMessage("Trazemos diariamente um pacote para você organizar aos poucos sua biblioteca."),
-    "tutorial_however_you_want" : MessageLookupByLibrary.simpleMessage("Organize suas fotos adicionando tags, como \"família\", \"pets\", ou o que você quiser."),
-    "tutorial_just_swipe" : MessageLookupByLibrary.simpleMessage("Depois de adicionar as tags na sua foto, basta fazer um swipe para ir para a próxima."),
-    "tutorial_multiselect" : MessageLookupByLibrary.simpleMessage("Você pode \"tocar e segurar\" em suas fotos para marcar várias fotos de uma vez."),
-    "tutorial_secret" : MessageLookupByLibrary.simpleMessage("Oculte suas fotos privadas com proteção por senha, mantendo-as seguras."),
-    "tuxedo_tag" : MessageLookupByLibrary.simpleMessage("Smoking"),
-    "twig_tag" : MessageLookupByLibrary.simpleMessage("Galho"),
-    "umbrella_tag" : MessageLookupByLibrary.simpleMessage("Guarda-chuva"),
-    "underwater_tag" : MessageLookupByLibrary.simpleMessage("Subaquático"),
-    "unicycle_tag" : MessageLookupByLibrary.simpleMessage("Monociclo"),
-    "unlimited_private_pics" : MessageLookupByLibrary.simpleMessage("Fotos privadas ilimitadas"),
-    "vacation_tag" : MessageLookupByLibrary.simpleMessage("Férias"),
-    "van_tag" : MessageLookupByLibrary.simpleMessage("Van"),
-    "vegetable_tag" : MessageLookupByLibrary.simpleMessage("Vegetal"),
-    "vehicle_tag" : MessageLookupByLibrary.simpleMessage("Veículo"),
-    "veil_tag" : MessageLookupByLibrary.simpleMessage("Véu"),
-    "view_hidden_photos" : MessageLookupByLibrary.simpleMessage("Para ver suas fotos ocultas, libere o bloqueio nas configurações do aplicativo. Você pode desbloqueá-las com seu PIN."),
-    "volcano_tag" : MessageLookupByLibrary.simpleMessage("Vulcão"),
-    "wakeboarding_tag" : MessageLookupByLibrary.simpleMessage("Wakeboarding"),
-    "wall_tag" : MessageLookupByLibrary.simpleMessage("Parede"),
-    "waterfall_tag" : MessageLookupByLibrary.simpleMessage("Cachoeira"),
-    "waterfowl_tag" : MessageLookupByLibrary.simpleMessage("Aves aquáticas"),
-    "waterskiing_tag" : MessageLookupByLibrary.simpleMessage("Waterskiing"),
-    "web_page_tag" : MessageLookupByLibrary.simpleMessage("Site"),
-    "welcome" : MessageLookupByLibrary.simpleMessage("Bem-vindo!"),
-    "wetsuit_tag" : MessageLookupByLibrary.simpleMessage("Roupa de mergulho"),
-    "wheel_tag" : MessageLookupByLibrary.simpleMessage("Roda"),
-    "wheelbarrow_tag" : MessageLookupByLibrary.simpleMessage("Carrinho de mão"),
-    "whiteboard_tag" : MessageLookupByLibrary.simpleMessage("Quadro branco"),
-    "windshield_tag" : MessageLookupByLibrary.simpleMessage("Parabrisa"),
-    "windsurfing_tag" : MessageLookupByLibrary.simpleMessage("Windsurf"),
-    "wine_tag" : MessageLookupByLibrary.simpleMessage("Vinho"),
-    "wing_tag" : MessageLookupByLibrary.simpleMessage("Asa"),
-    "wool_tag" : MessageLookupByLibrary.simpleMessage("Lã"),
-    "work_tag" : MessageLookupByLibrary.simpleMessage("Trabalho"),
-    "wreath_tag" : MessageLookupByLibrary.simpleMessage("Coroa"),
-    "x_minutes" : MessageLookupByLibrary.simpleMessage("20 min"),
-    "year" : MessageLookupByLibrary.simpleMessage("ano"),
-    "yes" : MessageLookupByLibrary.simpleMessage("Sim"),
-    "you_are_premium" : MessageLookupByLibrary.simpleMessage("Você é premium!"),
-    "your_secret_key" : MessageLookupByLibrary.simpleMessage("Sua chave secreta")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "access_code": MessageLookupByLibrary.simpleMessage("Código de Acesso"),
+        "access_code_sent": m0,
+        "add_multiple_tags":
+            MessageLookupByLibrary.simpleMessage("Adicione múltiplas tags"),
+        "add_tag": MessageLookupByLibrary.simpleMessage("Add tag"),
+        "add_tags": MessageLookupByLibrary.simpleMessage("Add tags"),
+        "aerospace_engineering_tag":
+            MessageLookupByLibrary.simpleMessage("Engenharia aeroespacial"),
+        "aircraft_tag": MessageLookupByLibrary.simpleMessage("Aeronave"),
+        "airliner_tag": MessageLookupByLibrary.simpleMessage("Avião comercial"),
+        "airplane_tag": MessageLookupByLibrary.simpleMessage("Avião"),
+        "alcohol_tag": MessageLookupByLibrary.simpleMessage("Álcool"),
+        "all_at_once": MessageLookupByLibrary.simpleMessage(
+            "Organize multiple photos at once"),
+        "all_photos_were_tagged": MessageLookupByLibrary.simpleMessage(
+            "Não há mais nenhuma foto para organizar."),
+        "all_search_tags":
+            MessageLookupByLibrary.simpleMessage("Todas as tags da pesquisa"),
+        "always": MessageLookupByLibrary.simpleMessage("Always"),
+        "aquarium_tag": MessageLookupByLibrary.simpleMessage("Aquário"),
+        "archery_tag": MessageLookupByLibrary.simpleMessage("Arco e flecha"),
+        "armrest_tag": MessageLookupByLibrary.simpleMessage("Apoio de braço"),
+        "army_tag": MessageLookupByLibrary.simpleMessage("Exército"),
+        "ask_photo_library_permission": MessageLookupByLibrary.simpleMessage(
+            "Precisamos de acesso à sua biblioteca de fotos para que você possa começar a organizar suas fotos com picPics. Não se preocupe, seus dados nunca sairão do seu dispositivo!"),
+        "asphalt_tag": MessageLookupByLibrary.simpleMessage("Asfalto"),
+        "aurora_tag": MessageLookupByLibrary.simpleMessage("Aurora"),
+        "auto_renewable_first_part":
+            MessageLookupByLibrary.simpleMessage("A assinatura é "),
+        "auto_renewable_second_part":
+            MessageLookupByLibrary.simpleMessage("auto-renovável."),
+        "aviation_tag": MessageLookupByLibrary.simpleMessage("Aviação"),
+        "baby_tag": MessageLookupByLibrary.simpleMessage("Bebê"),
+        "backpacking_tag": MessageLookupByLibrary.simpleMessage("Mochilão"),
+        "badminton_tag": MessageLookupByLibrary.simpleMessage("Badminton"),
+        "bag_tag": MessageLookupByLibrary.simpleMessage("Bolsa"),
+        "balance_tag": MessageLookupByLibrary.simpleMessage("Equilíbrio"),
+        "balloon_tag": MessageLookupByLibrary.simpleMessage("Balão"),
+        "ballroom_tag": MessageLookupByLibrary.simpleMessage("Salão de baile"),
+        "bangle_tag": MessageLookupByLibrary.simpleMessage("Pulseira"),
+        "bangs_tag": MessageLookupByLibrary.simpleMessage("Franja"),
+        "bar_tag": MessageLookupByLibrary.simpleMessage("Bar"),
+        "barn_tag": MessageLookupByLibrary.simpleMessage("Celeiro"),
+        "basset_hound_tag":
+            MessageLookupByLibrary.simpleMessage("Basset hound"),
+        "bathing_tag": MessageLookupByLibrary.simpleMessage("Banho"),
+        "bathroom_tag": MessageLookupByLibrary.simpleMessage("Banheiro"),
+        "beach_tag": MessageLookupByLibrary.simpleMessage("Praia"),
+        "beanie_tag": MessageLookupByLibrary.simpleMessage("Gorro"),
+        "bear_tag": MessageLookupByLibrary.simpleMessage("Urso"),
+        "beard_tag": MessageLookupByLibrary.simpleMessage("Barba"),
+        "bedroom_tag": MessageLookupByLibrary.simpleMessage("Quarto"),
+        "bench_tag": MessageLookupByLibrary.simpleMessage("Banco"),
+        "bento_tag": MessageLookupByLibrary.simpleMessage("Marmita"),
+        "bicycle_tag": MessageLookupByLibrary.simpleMessage("Bicicleta"),
+        "bird_tag": MessageLookupByLibrary.simpleMessage("Pássaro"),
+        "blazer_tag": MessageLookupByLibrary.simpleMessage("Blazer"),
+        "boat_tag": MessageLookupByLibrary.simpleMessage("Barco"),
+        "bonfire_tag": MessageLookupByLibrary.simpleMessage("Fogueira"),
+        "boxer_tag": MessageLookupByLibrary.simpleMessage("Boxer"),
+        "bracelet_tag": MessageLookupByLibrary.simpleMessage("Pulseira"),
+        "branch_tag": MessageLookupByLibrary.simpleMessage("Ramo"),
+        "bread_tag": MessageLookupByLibrary.simpleMessage("Pão"),
+        "brick_tag": MessageLookupByLibrary.simpleMessage("Tijolo"),
+        "bride_tag": MessageLookupByLibrary.simpleMessage("Noiva"),
+        "bridge_tag": MessageLookupByLibrary.simpleMessage("Ponte"),
+        "bridle_tag": MessageLookupByLibrary.simpleMessage("Refrear"),
+        "brig_tag": MessageLookupByLibrary.simpleMessage("Brigue"),
+        "building_tag": MessageLookupByLibrary.simpleMessage("Edifício"),
+        "bull_tag": MessageLookupByLibrary.simpleMessage("Touro"),
+        "bullfighting_tag": MessageLookupByLibrary.simpleMessage("Touradas"),
+        "bumper_tag": MessageLookupByLibrary.simpleMessage("Parachoque"),
+        "bunk_bed_tag": MessageLookupByLibrary.simpleMessage("Beliche"),
+        "bus_tag": MessageLookupByLibrary.simpleMessage("Ônibus"),
+        "butterfly_tag": MessageLookupByLibrary.simpleMessage("Borboleta"),
+        "cabinetry_tag": MessageLookupByLibrary.simpleMessage("Gabinete"),
+        "cage_tag": MessageLookupByLibrary.simpleMessage("Cela"),
+        "cairn_terrier_tag":
+            MessageLookupByLibrary.simpleMessage("Cairn terrier"),
+        "cake_tag": MessageLookupByLibrary.simpleMessage("Bolo"),
+        "camping_tag": MessageLookupByLibrary.simpleMessage("Acampamento"),
+        "cancel": MessageLookupByLibrary.simpleMessage("Cancelar"),
+        "cancel_anytime":
+            MessageLookupByLibrary.simpleMessage("Cancele a qualquer momento"),
+        "canoe_tag": MessageLookupByLibrary.simpleMessage("Canoa"),
+        "canyon_tag": MessageLookupByLibrary.simpleMessage("Canyon"),
+        "cap_tag": MessageLookupByLibrary.simpleMessage("Boné"),
+        "cappuccino_tag": MessageLookupByLibrary.simpleMessage("Cappuccino"),
+        "car_tag": MessageLookupByLibrary.simpleMessage("Carro"),
+        "carnival_tag": MessageLookupByLibrary.simpleMessage("Carnaval"),
+        "casino_tag": MessageLookupByLibrary.simpleMessage("Casino"),
+        "castle_tag": MessageLookupByLibrary.simpleMessage("Castelo"),
+        "cat_tag": MessageLookupByLibrary.simpleMessage("Gato"),
+        "cathedral_tag": MessageLookupByLibrary.simpleMessage("Catedral"),
+        "cattle_tag": MessageLookupByLibrary.simpleMessage("Gado"),
+        "cavalier_tag": MessageLookupByLibrary.simpleMessage("Cavaleiro"),
+        "cave_tag": MessageLookupByLibrary.simpleMessage("Caverna"),
+        "caving_tag": MessageLookupByLibrary.simpleMessage("Caving"),
+        "centrepiece_tag": MessageLookupByLibrary.simpleMessage("Peça central"),
+        "chain_tag": MessageLookupByLibrary.simpleMessage("Cadeia"),
+        "chair_tag": MessageLookupByLibrary.simpleMessage("Cadeira"),
+        "cheeseburger_tag":
+            MessageLookupByLibrary.simpleMessage("Cheeseburger"),
+        "christmas_tag": MessageLookupByLibrary.simpleMessage("Natal"),
+        "church_tag": MessageLookupByLibrary.simpleMessage("Igreja"),
+        "circus_tag": MessageLookupByLibrary.simpleMessage("Circo"),
+        "class_tag": MessageLookupByLibrary.simpleMessage("Classe"),
+        "cliff_tag": MessageLookupByLibrary.simpleMessage("Penhasco"),
+        "clipper_tag": MessageLookupByLibrary.simpleMessage("Cortador"),
+        "clock_tag": MessageLookupByLibrary.simpleMessage("Relógio"),
+        "close": MessageLookupByLibrary.simpleMessage("Fechar"),
+        "clown_tag": MessageLookupByLibrary.simpleMessage("Palhaço"),
+        "coffee_tag": MessageLookupByLibrary.simpleMessage("Café"),
+        "cola_tag": MessageLookupByLibrary.simpleMessage("Cola"),
+        "comet_tag": MessageLookupByLibrary.simpleMessage("Cometa"),
+        "comics_tag": MessageLookupByLibrary.simpleMessage("Quadrinhos"),
+        "community_tag": MessageLookupByLibrary.simpleMessage("Comunidade"),
+        "computer_tag": MessageLookupByLibrary.simpleMessage("Computador"),
+        "concert_tag": MessageLookupByLibrary.simpleMessage("Concerto"),
+        "confirm_email": MessageLookupByLibrary.simpleMessage(
+            "Confirm your registration email to receive your access code"),
+        "confirm_secret_key":
+            MessageLookupByLibrary.simpleMessage("Confirme a chave"),
+        "construction_tag": MessageLookupByLibrary.simpleMessage("Construção"),
+        "continue_string": MessageLookupByLibrary.simpleMessage("Continuar"),
+        "cookie_tag": MessageLookupByLibrary.simpleMessage("Biscoito"),
+        "cookware_and_bakeware_tag":
+            MessageLookupByLibrary.simpleMessage("Panelas e assadeiras"),
+        "cool_tag": MessageLookupByLibrary.simpleMessage("Legal"),
+        "cotton_tag": MessageLookupByLibrary.simpleMessage("Algodão"),
+        "couch_tag": MessageLookupByLibrary.simpleMessage("Sofá"),
+        "countertop_tag": MessageLookupByLibrary.simpleMessage("Bancada"),
+        "country": MessageLookupByLibrary.simpleMessage("país"),
+        "couscous_tag": MessageLookupByLibrary.simpleMessage("Cuscuz"),
+        "crew_tag": MessageLookupByLibrary.simpleMessage("Tripulação"),
+        "crochet_tag": MessageLookupByLibrary.simpleMessage("Crochê"),
+        "crocodile_tag": MessageLookupByLibrary.simpleMessage("Crocodilo"),
+        "crowd_tag": MessageLookupByLibrary.simpleMessage("Multidão"),
+        "cuisine_tag": MessageLookupByLibrary.simpleMessage("Cozinha"),
+        "cup_tag": MessageLookupByLibrary.simpleMessage("Copo"),
+        "curling_tag": MessageLookupByLibrary.simpleMessage("Curling"),
+        "curtain_tag": MessageLookupByLibrary.simpleMessage("Cortina"),
+        "cushion_tag": MessageLookupByLibrary.simpleMessage("Almofada"),
+        "cutlery_tag": MessageLookupByLibrary.simpleMessage("Talheres"),
+        "cycling_tag": MessageLookupByLibrary.simpleMessage("Ciclismo"),
+        "daily_challenge":
+            MessageLookupByLibrary.simpleMessage("Desafio diário"),
+        "daily_challenge_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Para podermos enviar seus desafios diários, precisamos de autorização para enviar notificações; portanto, é necessário que você autorize as notificações nas opções do seu telefone celular."),
+        "daily_goal": MessageLookupByLibrary.simpleMessage("Objetivo diário"),
+        "daily_notification_description": MessageLookupByLibrary.simpleMessage(
+            "Está na hora de completar o seu desáfio. Entre no picPics para completá-lo!"),
+        "daily_notification_title":
+            MessageLookupByLibrary.simpleMessage("Desafio diário"),
+        "dalmatian_tag": MessageLookupByLibrary.simpleMessage("Dálmata"),
+        "dam_tag": MessageLookupByLibrary.simpleMessage("Barragem"),
+        "dance_tag": MessageLookupByLibrary.simpleMessage("Dança"),
+        "deejay_tag": MessageLookupByLibrary.simpleMessage("DJ"),
+        "delete": MessageLookupByLibrary.simpleMessage("Deletar"),
+        "denim_tag": MessageLookupByLibrary.simpleMessage("Jeans"),
+        "desert_tag": MessageLookupByLibrary.simpleMessage("Deserto"),
+        "desk_tag": MessageLookupByLibrary.simpleMessage("Mesa"),
+        "device_has_no_pics": MessageLookupByLibrary.simpleMessage(
+            "Esse dispositivo não tem nenhuma foto na galeria, assim não há nenhuma foto que possa ser taggeada."),
+        "dinosaur_tag": MessageLookupByLibrary.simpleMessage("Dinosauro"),
+        "disable_secret": MessageLookupByLibrary.simpleMessage(
+            "Deseja desocultar esta foto?"),
+        "dog_tag": MessageLookupByLibrary.simpleMessage("Cachorro"),
+        "doily_tag": MessageLookupByLibrary.simpleMessage("Doily"),
+        "dont_ask_again":
+            MessageLookupByLibrary.simpleMessage("Não pergunte mais"),
+        "dragon_tag": MessageLookupByLibrary.simpleMessage("Dragão"),
+        "drawer_tag": MessageLookupByLibrary.simpleMessage("Gaveta"),
+        "dress_tag": MessageLookupByLibrary.simpleMessage("Vestido"),
+        "duck_tag": MessageLookupByLibrary.simpleMessage("Pato"),
+        "dude_tag": MessageLookupByLibrary.simpleMessage("Cara"),
+        "dune_tag": MessageLookupByLibrary.simpleMessage("Duna"),
+        "ear_tag": MessageLookupByLibrary.simpleMessage("Orelha"),
+        "eating_tag": MessageLookupByLibrary.simpleMessage("Comendo"),
+        "edit_tag": MessageLookupByLibrary.simpleMessage("Editar tag"),
+        "email": MessageLookupByLibrary.simpleMessage("Email"),
+        "enable_faceid":
+            MessageLookupByLibrary.simpleMessage("Habilitar Face ID"),
+        "enable_fingerprint":
+            MessageLookupByLibrary.simpleMessage("Habilitar Impressão Digital"),
+        "enable_irisscanner":
+            MessageLookupByLibrary.simpleMessage("Habilitar Scanner de Iris"),
+        "enable_touchid":
+            MessageLookupByLibrary.simpleMessage("Habilitar Touch ID"),
+        "event_tag": MessageLookupByLibrary.simpleMessage("Evento"),
+        "export_all_gallery":
+            MessageLookupByLibrary.simpleMessage("Exporte toda galeria"),
+        "export_library":
+            MessageLookupByLibrary.simpleMessage("Exportar biblioteca"),
+        "eyelash_tag": MessageLookupByLibrary.simpleMessage("Cílios"),
+        "factory_tag": MessageLookupByLibrary.simpleMessage("Fábrica"),
+        "family_tag": MessageLookupByLibrary.simpleMessage("Família"),
+        "farm_tag": MessageLookupByLibrary.simpleMessage("Fazenda"),
+        "fast_food_tag": MessageLookupByLibrary.simpleMessage("Fast food"),
+        "feedback_bug_report":
+            MessageLookupByLibrary.simpleMessage("Feedback & Bugs"),
+        "ferris_wheel_tag":
+            MessageLookupByLibrary.simpleMessage("Roda gigante"),
+        "fiction_tag": MessageLookupByLibrary.simpleMessage("Ficção"),
+        "field_tag": MessageLookupByLibrary.simpleMessage("Campo"),
+        "find_easily":
+            MessageLookupByLibrary.simpleMessage("Easily find your photos"),
+        "fire_tag": MessageLookupByLibrary.simpleMessage("Fogo"),
+        "fireworks_tag":
+            MessageLookupByLibrary.simpleMessage("Fogos de artifício"),
+        "fishing_tag": MessageLookupByLibrary.simpleMessage("Pesca"),
+        "flag_tag": MessageLookupByLibrary.simpleMessage("Sinalizar"),
+        "flap_tag": MessageLookupByLibrary.simpleMessage("Flap"),
+        "flesh_tag": MessageLookupByLibrary.simpleMessage("Corpo"),
+        "flora_tag": MessageLookupByLibrary.simpleMessage("Flora"),
+        "flower_tag": MessageLookupByLibrary.simpleMessage("Flor"),
+        "flowerpot_tag": MessageLookupByLibrary.simpleMessage("Vaso de flores"),
+        "fog_tag": MessageLookupByLibrary.simpleMessage("Névoa"),
+        "food_tag": MessageLookupByLibrary.simpleMessage("Comida"),
+        "foods_tag": MessageLookupByLibrary.simpleMessage("Comidas"),
+        "foot_tag": MessageLookupByLibrary.simpleMessage("Pé"),
+        "forest_tag": MessageLookupByLibrary.simpleMessage("Floresta"),
+        "forgot_secret_key":
+            MessageLookupByLibrary.simpleMessage("Forgot secret key?"),
+        "frigate_tag": MessageLookupByLibrary.simpleMessage("Fragata"),
+        "fruit_tag": MessageLookupByLibrary.simpleMessage("Frutas"),
+        "full_screen": MessageLookupByLibrary.simpleMessage("Tela cheia"),
+        "fun_tag": MessageLookupByLibrary.simpleMessage("Diversão"),
+        "fur_tag": MessageLookupByLibrary.simpleMessage("Casaco de pele"),
+        "gallery_access_permission":
+            MessageLookupByLibrary.simpleMessage("Permissão de acesso"),
+        "gallery_access_permission_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Para começarmos a organizar suas fotos, precisamos de autorização para acessá-las\""),
+        "gallery_access_reason": MessageLookupByLibrary.simpleMessage(
+            "Para organizar suas fotos precisamos de acesso a sua galeria"),
+        "garden_tag": MessageLookupByLibrary.simpleMessage("Jardim"),
+        "gelato_tag": MessageLookupByLibrary.simpleMessage("Sorvete"),
+        "gerbil_tag": MessageLookupByLibrary.simpleMessage("Gerbil"),
+        "get_premium_description":
+            MessageLookupByLibrary.simpleMessage("PARA TER TODOS OS RECURSOS"),
+        "get_premium_now":
+            MessageLookupByLibrary.simpleMessage("Seja premium agora!"),
+        "get_premium_title":
+            MessageLookupByLibrary.simpleMessage("SEJA PREMIUM"),
+        "glacier_tag": MessageLookupByLibrary.simpleMessage("Geleira"),
+        "glasses_tag": MessageLookupByLibrary.simpleMessage("Óculos"),
+        "glitter_tag": MessageLookupByLibrary.simpleMessage("Glitter"),
+        "goggles_tag": MessageLookupByLibrary.simpleMessage("Óculos"),
+        "gown_tag": MessageLookupByLibrary.simpleMessage("Vestido"),
+        "graduation_tag": MessageLookupByLibrary.simpleMessage("Graduação"),
+        "grandparent_tag": MessageLookupByLibrary.simpleMessage("Avós"),
+        "groom_tag": MessageLookupByLibrary.simpleMessage("Noivo"),
+        "gymnastics_tag": MessageLookupByLibrary.simpleMessage("Ginástica"),
+        "hair_tag": MessageLookupByLibrary.simpleMessage("Cabelo"),
+        "hand_tag": MessageLookupByLibrary.simpleMessage("Mão"),
+        "handbag_tag": MessageLookupByLibrary.simpleMessage("Bolsa"),
+        "handrail_tag": MessageLookupByLibrary.simpleMessage("Corrimão"),
+        "hanukkah_tag": MessageLookupByLibrary.simpleMessage("Hanukkah"),
+        "hat_tag": MessageLookupByLibrary.simpleMessage("Chapéu"),
+        "heart_tag": MessageLookupByLibrary.simpleMessage("Coração"),
+        "helicopter_tag": MessageLookupByLibrary.simpleMessage("Helicóptero"),
+        "helmet_tag": MessageLookupByLibrary.simpleMessage("Capacete"),
+        "herd_tag": MessageLookupByLibrary.simpleMessage("Rebanho"),
+        "himalayan_tag": MessageLookupByLibrary.simpleMessage("Himalaia"),
+        "home_tag": MessageLookupByLibrary.simpleMessage("Casa"),
+        "horn_tag": MessageLookupByLibrary.simpleMessage("Chifre"),
+        "horse_tag": MessageLookupByLibrary.simpleMessage("Cavalo"),
+        "hot_dog_tag": MessageLookupByLibrary.simpleMessage("Cachorro-quente"),
+        "how_many_pics": MessageLookupByLibrary.simpleMessage("Quantas fotos?"),
+        "ice_tag": MessageLookupByLibrary.simpleMessage("Gelo"),
+        "iceberg_tag": MessageLookupByLibrary.simpleMessage("Iceberg"),
+        "icicle_tag": MessageLookupByLibrary.simpleMessage("Icicle"),
+        "icing_tag": MessageLookupByLibrary.simpleMessage("Glacê"),
+        "icon_tag": MessageLookupByLibrary.simpleMessage("Ícone"),
+        "infinite_tags": MessageLookupByLibrary.simpleMessage("Tags a vontade"),
+        "infrastructure_tag":
+            MessageLookupByLibrary.simpleMessage("Infraestrutura"),
+        "insect_tag": MessageLookupByLibrary.simpleMessage("Inseto"),
+        "jacket_tag": MessageLookupByLibrary.simpleMessage("Jaqueta"),
+        "jeans_tag": MessageLookupByLibrary.simpleMessage("Jeans"),
+        "jersey_tag": MessageLookupByLibrary.simpleMessage("Camisola"),
+        "jewellery_tag": MessageLookupByLibrary.simpleMessage("Joalharia"),
+        "joker_tag": MessageLookupByLibrary.simpleMessage("Palhaço"),
+        "juice_tag": MessageLookupByLibrary.simpleMessage("Juice"),
+        "jungle_tag": MessageLookupByLibrary.simpleMessage("Selva"),
+        "junk_tag": MessageLookupByLibrary.simpleMessage("Lixo"),
+        "kayak_tag": MessageLookupByLibrary.simpleMessage("Caiaque"),
+        "keep_asking": MessageLookupByLibrary.simpleMessage("Pergunte sempre"),
+        "keep_safe": MessageLookupByLibrary.simpleMessage(
+            "Sua foto agora está segura no picPics. Deseja excluí-la do rolo da câmera?"),
+        "kitchen_tag": MessageLookupByLibrary.simpleMessage("Cozinha"),
+        "knitting_tag": MessageLookupByLibrary.simpleMessage("Tricô"),
+        "lake_tag": MessageLookupByLibrary.simpleMessage("Lago"),
+        "lampshade_tag": MessageLookupByLibrary.simpleMessage("Abajur"),
+        "language": MessageLookupByLibrary.simpleMessage("Lingua"),
+        "larva_tag": MessageLookupByLibrary.simpleMessage("Larva"),
+        "laugh_tag": MessageLookupByLibrary.simpleMessage("Risada"),
+        "leather_tag": MessageLookupByLibrary.simpleMessage("Couro"),
+        "leggings_tag": MessageLookupByLibrary.simpleMessage("Leggings"),
+        "lego_tag": MessageLookupByLibrary.simpleMessage("Lego"),
+        "leisure_tag": MessageLookupByLibrary.simpleMessage("Lazer"),
+        "lighthouse_tag": MessageLookupByLibrary.simpleMessage("Farol"),
+        "lightning_tag": MessageLookupByLibrary.simpleMessage("Relâmpago"),
+        "lipstick_tag": MessageLookupByLibrary.simpleMessage("Batom"),
+        "lock_with_pin": MessageLookupByLibrary.simpleMessage(
+            "Lock down your private photos with a PIN passcode."),
+        "lock_your_photos":
+            MessageLookupByLibrary.simpleMessage("Lock your photos"),
+        "longboard_tag": MessageLookupByLibrary.simpleMessage("Longboard"),
+        "love_tag": MessageLookupByLibrary.simpleMessage("Amor"),
+        "loveseat_tag": MessageLookupByLibrary.simpleMessage("Loveseat"),
+        "lunch_tag": MessageLookupByLibrary.simpleMessage("Almoço"),
+        "marathon_tag": MessageLookupByLibrary.simpleMessage("Maratona"),
+        "meal_tag": MessageLookupByLibrary.simpleMessage("Refeição"),
+        "menu_tag": MessageLookupByLibrary.simpleMessage("Menu"),
+        "metal_tag": MessageLookupByLibrary.simpleMessage("Metal"),
+        "militia_tag": MessageLookupByLibrary.simpleMessage("Milícia"),
+        "miniature_tag": MessageLookupByLibrary.simpleMessage("Miniatura"),
+        "mobile_phone_tag":
+            MessageLookupByLibrary.simpleMessage("Telefone celular"),
+        "model_tag": MessageLookupByLibrary.simpleMessage("Modelo"),
+        "money_tag": MessageLookupByLibrary.simpleMessage("Dinheiro"),
+        "monochrome_tag": MessageLookupByLibrary.simpleMessage("Monocromático"),
+        "month": MessageLookupByLibrary.simpleMessage("mês"),
+        "monument_tag": MessageLookupByLibrary.simpleMessage("Monumento"),
+        "moon_tag": MessageLookupByLibrary.simpleMessage("Lua"),
+        "mortarboard_tag": MessageLookupByLibrary.simpleMessage("Mortarboard"),
+        "mosque_tag": MessageLookupByLibrary.simpleMessage("Mesquita"),
+        "motorcycle_tag": MessageLookupByLibrary.simpleMessage("motocicleta"),
+        "mountain_tag": MessageLookupByLibrary.simpleMessage("Montanha"),
+        "moustache_tag": MessageLookupByLibrary.simpleMessage("Bigode"),
+        "mouth_tag": MessageLookupByLibrary.simpleMessage("Boca"),
+        "mufti_tag": MessageLookupByLibrary.simpleMessage("À paisana"),
+        "muscle_tag": MessageLookupByLibrary.simpleMessage("Músculo"),
+        "museum_tag": MessageLookupByLibrary.simpleMessage("Museu"),
+        "musical_instrument_tag":
+            MessageLookupByLibrary.simpleMessage("Instrumento musical"),
+        "musical_tag": MessageLookupByLibrary.simpleMessage("Musical"),
+        "musician_tag": MessageLookupByLibrary.simpleMessage("Músico"),
+        "nail_tag": MessageLookupByLibrary.simpleMessage("Unha"),
+        "nebula_tag": MessageLookupByLibrary.simpleMessage("Nebulosa"),
+        "necklace_tag": MessageLookupByLibrary.simpleMessage("Colar"),
+        "neon_tag": MessageLookupByLibrary.simpleMessage("Neon"),
+        "net_tag": MessageLookupByLibrary.simpleMessage("Rede"),
+        "new_secret_key":
+            MessageLookupByLibrary.simpleMessage("Nova chave secreta"),
+        "news_tag": MessageLookupByLibrary.simpleMessage("Notícias"),
+        "newspaper_tag": MessageLookupByLibrary.simpleMessage("Jornal"),
+        "next": MessageLookupByLibrary.simpleMessage("Próximo"),
+        "nightclub_tag": MessageLookupByLibrary.simpleMessage("Boate"),
+        "no": MessageLookupByLibrary.simpleMessage("Não"),
+        "no_ads": MessageLookupByLibrary.simpleMessage("Sem propagandas"),
+        "no_previous_purchase":
+            MessageLookupByLibrary.simpleMessage("Nenhuma compra encontrada"),
+        "no_tagged_photos": MessageLookupByLibrary.simpleMessage(
+            "Você não tem nenhuma foto taggeada ainda"),
+        "no_tags_found":
+            MessageLookupByLibrary.simpleMessage("Nenhuma tag encontrada"),
+        "no_valid_subscription": MessageLookupByLibrary.simpleMessage(
+            "Não foi possível encontrar uma compra de assinatura válida."),
+        "notification_time":
+            MessageLookupByLibrary.simpleMessage("Horário da notificação"),
+        "notifications": MessageLookupByLibrary.simpleMessage("Notificações"),
+        "ny_tag": MessageLookupByLibrary.simpleMessage("NY"),
+        "odometer_tag": MessageLookupByLibrary.simpleMessage("Odômetro"),
+        "ok": MessageLookupByLibrary.simpleMessage("OK"),
+        "open_gallery": MessageLookupByLibrary.simpleMessage("Abrir galeria"),
+        "organized_photos_description": MessageLookupByLibrary.simpleMessage(
+            "Fotos que já foram taggeadas"),
+        "organized_photos_title":
+            MessageLookupByLibrary.simpleMessage("Fotos organizadas"),
+        "outerwear_tag": MessageLookupByLibrary.simpleMessage("Agasalhos"),
+        "palace_tag": MessageLookupByLibrary.simpleMessage("Palácio"),
+        "paper_tag": MessageLookupByLibrary.simpleMessage("Papel"),
+        "park_tag": MessageLookupByLibrary.simpleMessage("Parque"),
+        "parties_tag": MessageLookupByLibrary.simpleMessage("Festas"),
+        "party_tag": MessageLookupByLibrary.simpleMessage("Festa"),
+        "passport_tag": MessageLookupByLibrary.simpleMessage("Passaporte"),
+        "pasteles_tag": MessageLookupByLibrary.simpleMessage("Pasteles"),
+        "pattern_tag": MessageLookupByLibrary.simpleMessage("Padrôes"),
+        "penguin_tag": MessageLookupByLibrary.simpleMessage("Pinguim"),
+        "person_tag": MessageLookupByLibrary.simpleMessage("Pessoa"),
+        "pet_tag": MessageLookupByLibrary.simpleMessage("Pet"),
+        "petal_tag": MessageLookupByLibrary.simpleMessage("Pétala"),
+        "pets_tag": MessageLookupByLibrary.simpleMessage("Pets"),
+        "pho_tag": MessageLookupByLibrary.simpleMessage("Pho"),
+        "photo_gallery_count": m1,
+        "photo_gallery_description":
+            MessageLookupByLibrary.simpleMessage("Fotos não organizadas"),
+        "photo_gallery_title":
+            MessageLookupByLibrary.simpleMessage("Galeria de fotos"),
+        "photo_location": MessageLookupByLibrary.simpleMessage("Local da foto"),
+        "photos_always_organized": MessageLookupByLibrary.simpleMessage(
+            "Agora suas fotos estarão sempre organizadas"),
+        "piano_tag": MessageLookupByLibrary.simpleMessage("Piano"),
+        "picnic_tag": MessageLookupByLibrary.simpleMessage("Piquenique"),
+        "picpics_photo_manager":
+            MessageLookupByLibrary.simpleMessage("picPics - Photo Manager"),
+        "pie_tag": MessageLookupByLibrary.simpleMessage("Torta"),
+        "pier_tag": MessageLookupByLibrary.simpleMessage("Pier"),
+        "pillow_tag": MessageLookupByLibrary.simpleMessage("Travesseiro"),
+        "pitch_tag": MessageLookupByLibrary.simpleMessage("Arremesso"),
+        "pixie_bob_tag": MessageLookupByLibrary.simpleMessage("Pixie-bob"),
+        "pizza_tag": MessageLookupByLibrary.simpleMessage("Pizza"),
+        "placemat_tag":
+            MessageLookupByLibrary.simpleMessage("Descanso de mesa"),
+        "plant_tag": MessageLookupByLibrary.simpleMessage("Planta"),
+        "playground_tag": MessageLookupByLibrary.simpleMessage("Recreio"),
+        "plush_tag": MessageLookupByLibrary.simpleMessage("Pelúcia"),
+        "pocket_tag": MessageLookupByLibrary.simpleMessage("Bolso"),
+        "polo_tag": MessageLookupByLibrary.simpleMessage("Pólo"),
+        "pomacentridae_tag":
+            MessageLookupByLibrary.simpleMessage("Pomacentridae"),
+        "pool_tag": MessageLookupByLibrary.simpleMessage("Piscina"),
+        "pop_music_tag": MessageLookupByLibrary.simpleMessage("Música Pop"),
+        "porcelain_tag": MessageLookupByLibrary.simpleMessage("Porcelana"),
+        "poster_tag": MessageLookupByLibrary.simpleMessage("Poster"),
+        "prairie_tag": MessageLookupByLibrary.simpleMessage("Pradaria"),
+        "premium_modal_description": m2,
+        "premium_modal_get_premium_description":
+            MessageLookupByLibrary.simpleMessage(
+                "Todos os recursos e sem anuncios"),
+        "premium_modal_get_premium_title":
+            MessageLookupByLibrary.simpleMessage("Obter conta premium"),
+        "premium_modal_watch_ad": MessageLookupByLibrary.simpleMessage(
+            "Assista ao anúncio para continuar"),
+        "presentation_tag":
+            MessageLookupByLibrary.simpleMessage("Apresentação"),
+        "privacy_policy":
+            MessageLookupByLibrary.simpleMessage("Política de privacidade"),
+        "private_photos":
+            MessageLookupByLibrary.simpleMessage("Fotos privadas"),
+        "product_tag": MessageLookupByLibrary.simpleMessage("Produto"),
+        "prom_tag": MessageLookupByLibrary.simpleMessage("Baile"),
+        "protect_with_encryption": MessageLookupByLibrary.simpleMessage(
+            "Protect your personal photos with encryption only accessible with a PIN password."),
+        "quilting_tag": MessageLookupByLibrary.simpleMessage("Quilting"),
+        "race_tag": MessageLookupByLibrary.simpleMessage("Corrida"),
+        "racing_tag": MessageLookupByLibrary.simpleMessage("Corrida"),
+        "rafting_tag": MessageLookupByLibrary.simpleMessage("Rafting"),
+        "ragdoll_tag": MessageLookupByLibrary.simpleMessage("Ragdoll"),
+        "rainbow_tag": MessageLookupByLibrary.simpleMessage("Arco Iris"),
+        "ranch_tag": MessageLookupByLibrary.simpleMessage("Rancho"),
+        "rate_this_app":
+            MessageLookupByLibrary.simpleMessage("Avalie esse app"),
+        "receipt_tag": MessageLookupByLibrary.simpleMessage("Recibo"),
+        "recent_tags": MessageLookupByLibrary.simpleMessage("Tags recentes"),
+        "reef_tag": MessageLookupByLibrary.simpleMessage("Recife"),
+        "rein_tag": MessageLookupByLibrary.simpleMessage("Rédea"),
+        "require_secret_key":
+            MessageLookupByLibrary.simpleMessage("Require secret key"),
+        "restore_purchase":
+            MessageLookupByLibrary.simpleMessage("Restaurar compra"),
+        "rickshaw_tag": MessageLookupByLibrary.simpleMessage("Riquixá"),
+        "ring_tag": MessageLookupByLibrary.simpleMessage("Toque"),
+        "river_tag": MessageLookupByLibrary.simpleMessage("Rio"),
+        "road_tag": MessageLookupByLibrary.simpleMessage("Caminho"),
+        "rock_tag": MessageLookupByLibrary.simpleMessage("Pedra"),
+        "rocket_tag": MessageLookupByLibrary.simpleMessage("Foguete"),
+        "rodeo_tag": MessageLookupByLibrary.simpleMessage("Rodeio"),
+        "roller_tag": MessageLookupByLibrary.simpleMessage("Roller"),
+        "roof_tag": MessageLookupByLibrary.simpleMessage("Cobertura"),
+        "room_tag": MessageLookupByLibrary.simpleMessage("Sala"),
+        "rowing_tag": MessageLookupByLibrary.simpleMessage("Remo"),
+        "rugby_tag": MessageLookupByLibrary.simpleMessage("Rugby"),
+        "ruins_tag": MessageLookupByLibrary.simpleMessage("Ruínas"),
+        "running_tag": MessageLookupByLibrary.simpleMessage("Corrida"),
+        "safari_tag": MessageLookupByLibrary.simpleMessage("Safari"),
+        "sailboat_tag": MessageLookupByLibrary.simpleMessage("Veleiro"),
+        "sand_tag": MessageLookupByLibrary.simpleMessage("Areia"),
+        "santa_claus_tag": MessageLookupByLibrary.simpleMessage("Papai Noel"),
+        "sari_tag": MessageLookupByLibrary.simpleMessage("Sari"),
+        "saucer_tag": MessageLookupByLibrary.simpleMessage("Pires"),
+        "save": MessageLookupByLibrary.simpleMessage("economize"),
+        "save_location":
+            MessageLookupByLibrary.simpleMessage("Confirmar localização"),
+        "scarf_tag": MessageLookupByLibrary.simpleMessage("Cachecol"),
+        "school_tag": MessageLookupByLibrary.simpleMessage("Escola"),
+        "screenshot_tag": MessageLookupByLibrary.simpleMessage("Screenshot"),
+        "screenshots_tag": MessageLookupByLibrary.simpleMessage("Screenshots"),
+        "scuba_diving_tag": MessageLookupByLibrary.simpleMessage("Mergulho"),
+        "seal_tag": MessageLookupByLibrary.simpleMessage("Selo"),
+        "search": MessageLookupByLibrary.simpleMessage("Pesquisar..."),
+        "search_all_tags_not_found": MessageLookupByLibrary.simpleMessage(
+            "Nenhuma foto foi encontrada com todas as tags."),
+        "search_results":
+            MessageLookupByLibrary.simpleMessage("Resultados da busca"),
+        "secret_key_created": MessageLookupByLibrary.simpleMessage(
+            "Chave secreta criada com sucesso!"),
+        "secret_photos": MessageLookupByLibrary.simpleMessage("Fotos Secretas"),
+        "selfie_tag": MessageLookupByLibrary.simpleMessage("Selfie"),
+        "selfies_tag": MessageLookupByLibrary.simpleMessage("Selfies"),
+        "settings": MessageLookupByLibrary.simpleMessage("Opções"),
+        "share_with_friends":
+            MessageLookupByLibrary.simpleMessage("Compartilhar"),
+        "shelf_tag": MessageLookupByLibrary.simpleMessage("Prateleira"),
+        "shell_tag": MessageLookupByLibrary.simpleMessage("Concha"),
+        "shetland_sheepdog_tag":
+            MessageLookupByLibrary.simpleMessage("Pastor de shetland"),
+        "shikoku_tag": MessageLookupByLibrary.simpleMessage("Shikoku"),
+        "shipwreck_tag": MessageLookupByLibrary.simpleMessage("Naufrágio"),
+        "shoe_tag": MessageLookupByLibrary.simpleMessage("Sapato"),
+        "shorts_tag": MessageLookupByLibrary.simpleMessage("Shorts"),
+        "sign": MessageLookupByLibrary.simpleMessage("Por"),
+        "singer_tag": MessageLookupByLibrary.simpleMessage("Cantor"),
+        "sink_tag": MessageLookupByLibrary.simpleMessage("Pia"),
+        "sitting_tag": MessageLookupByLibrary.simpleMessage("Sentado"),
+        "skateboard_tag": MessageLookupByLibrary.simpleMessage("Skateboard"),
+        "skateboarder_tag": MessageLookupByLibrary.simpleMessage("Skatista"),
+        "skiff_tag": MessageLookupByLibrary.simpleMessage("Esquife"),
+        "skin_tag": MessageLookupByLibrary.simpleMessage("Pele"),
+        "sky_tag": MessageLookupByLibrary.simpleMessage("Céu"),
+        "skyline_tag": MessageLookupByLibrary.simpleMessage("Horizonte"),
+        "skyscraper_tag": MessageLookupByLibrary.simpleMessage("Arranha-céu"),
+        "sled_tag": MessageLookupByLibrary.simpleMessage("Trenó"),
+        "sledding_tag": MessageLookupByLibrary.simpleMessage("Trenó"),
+        "sleep_tag": MessageLookupByLibrary.simpleMessage("Dormir"),
+        "smile_tag": MessageLookupByLibrary.simpleMessage("Sorriso"),
+        "sneakers_tag": MessageLookupByLibrary.simpleMessage("Tênis"),
+        "snorkeling_tag": MessageLookupByLibrary.simpleMessage("Snorkeling"),
+        "snowboarding_tag":
+            MessageLookupByLibrary.simpleMessage("Snowboarding"),
+        "soccer_tag": MessageLookupByLibrary.simpleMessage("Futebol"),
+        "softball_tag": MessageLookupByLibrary.simpleMessage("Softball"),
+        "soil_tag": MessageLookupByLibrary.simpleMessage("Solo"),
+        "song_tag": MessageLookupByLibrary.simpleMessage("Música"),
+        "space_tag": MessageLookupByLibrary.simpleMessage("Espaço"),
+        "sparkler_tag": MessageLookupByLibrary.simpleMessage("Sparkler"),
+        "speedboat_tag": MessageLookupByLibrary.simpleMessage("Lancha"),
+        "sphynx_tag": MessageLookupByLibrary.simpleMessage("Sphynx"),
+        "sports_tag": MessageLookupByLibrary.simpleMessage("Sports"),
+        "stadium_tag": MessageLookupByLibrary.simpleMessage("Estádio"),
+        "stairs_tag": MessageLookupByLibrary.simpleMessage("Escadas"),
+        "standing_tag": MessageLookupByLibrary.simpleMessage("De pé"),
+        "star_tag": MessageLookupByLibrary.simpleMessage("Estrela"),
+        "start": MessageLookupByLibrary.simpleMessage("Começar"),
+        "start_tagging":
+            MessageLookupByLibrary.simpleMessage("Começe a taggear"),
+        "statue_tag": MessageLookupByLibrary.simpleMessage("Estátua"),
+        "steaming_tag": MessageLookupByLibrary.simpleMessage("Fumegante"),
+        "storm_tag": MessageLookupByLibrary.simpleMessage("Tempestade"),
+        "strap_tag": MessageLookupByLibrary.simpleMessage("Alça"),
+        "stuffed_toy_tag":
+            MessageLookupByLibrary.simpleMessage("Brinquedo de pelúcia"),
+        "submarine_tag": MessageLookupByLibrary.simpleMessage("Submarino"),
+        "subwoofer_tag": MessageLookupByLibrary.simpleMessage("Subwoofer"),
+        "suggestions": MessageLookupByLibrary.simpleMessage("Sugestões"),
+        "sunglasses_tag": MessageLookupByLibrary.simpleMessage("Óculos de sol"),
+        "sunset_tag": MessageLookupByLibrary.simpleMessage("Pôr do sol"),
+        "superhero_tag": MessageLookupByLibrary.simpleMessage("Super heroi"),
+        "superman_tag": MessageLookupByLibrary.simpleMessage("Superman"),
+        "supervillain_tag": MessageLookupByLibrary.simpleMessage("Super-vilão"),
+        "supper_tag": MessageLookupByLibrary.simpleMessage("Ceia"),
+        "surfboard_tag":
+            MessageLookupByLibrary.simpleMessage("Prancha de surf"),
+        "surfing_tag": MessageLookupByLibrary.simpleMessage("Surfing"),
+        "sushi_tag": MessageLookupByLibrary.simpleMessage("Sushi"),
+        "swamp_tag": MessageLookupByLibrary.simpleMessage("Pântano"),
+        "swimming_tag": MessageLookupByLibrary.simpleMessage("Natação"),
+        "swimwear_tag": MessageLookupByLibrary.simpleMessage("Roupa de banho"),
+        "swing_tag": MessageLookupByLibrary.simpleMessage("Balanço"),
+        "tablecloth_tag":
+            MessageLookupByLibrary.simpleMessage("Toalha de mesa"),
+        "tableware_tag": MessageLookupByLibrary.simpleMessage("Talheres"),
+        "tag_multiple_photos_at_once": MessageLookupByLibrary.simpleMessage(
+            "Marque várias fotos de uma vez só"),
+        "take_a_look": MessageLookupByLibrary.simpleMessage(
+            "Dê uma olhada nesse aplicativo!"),
+        "take_a_look_description": m3,
+        "tattoo_tag": MessageLookupByLibrary.simpleMessage("Tattoo"),
+        "team_tag": MessageLookupByLibrary.simpleMessage("Time"),
+        "television_tag": MessageLookupByLibrary.simpleMessage("Televisão"),
+        "temple_tag": MessageLookupByLibrary.simpleMessage("Templo"),
+        "terms_of_use": MessageLookupByLibrary.simpleMessage("Termos de uso"),
+        "textile_tag": MessageLookupByLibrary.simpleMessage("Têxtil"),
+        "thanksgiving_tag":
+            MessageLookupByLibrary.simpleMessage("Ação de graças"),
+        "tie_tag": MessageLookupByLibrary.simpleMessage("Empate"),
+        "tights_tag": MessageLookupByLibrary.simpleMessage("Calças"),
+        "tile_tag": MessageLookupByLibrary.simpleMessage("Bloco"),
+        "time": MessageLookupByLibrary.simpleMessage("Horário"),
+        "tire_tag": MessageLookupByLibrary.simpleMessage("Pneu"),
+        "toe_tag": MessageLookupByLibrary.simpleMessage("Dedo do pé"),
+        "toggle_date": MessageLookupByLibrary.simpleMessage("Data"),
+        "toggle_days": MessageLookupByLibrary.simpleMessage("Dias"),
+        "toggle_months": MessageLookupByLibrary.simpleMessage("Meses"),
+        "toggle_tags": MessageLookupByLibrary.simpleMessage("Tags"),
+        "tower_tag": MessageLookupByLibrary.simpleMessage("Torre"),
+        "toy_tag": MessageLookupByLibrary.simpleMessage("Brinquedo"),
+        "track_tag": MessageLookupByLibrary.simpleMessage("Caminho"),
+        "tractor_tag": MessageLookupByLibrary.simpleMessage("Trator"),
+        "train_tag": MessageLookupByLibrary.simpleMessage("Trem"),
+        "travel_tag": MessageLookupByLibrary.simpleMessage("Viagens"),
+        "trunk_tag": MessageLookupByLibrary.simpleMessage("Tronco"),
+        "tsqr_tag": MessageLookupByLibrary.simpleMessage("Times Square"),
+        "tubing_tag": MessageLookupByLibrary.simpleMessage("Tubulação"),
+        "turtle_tag": MessageLookupByLibrary.simpleMessage("Tartaruga"),
+        "tutorial_daily_package": MessageLookupByLibrary.simpleMessage(
+            "Trazemos diariamente um pacote para você organizar aos poucos sua biblioteca."),
+        "tutorial_however_you_want": MessageLookupByLibrary.simpleMessage(
+            "Organize suas fotos adicionando tags, como \"família\", \"pets\", ou o que você quiser."),
+        "tutorial_just_swipe": MessageLookupByLibrary.simpleMessage(
+            "Depois de adicionar as tags na sua foto, basta fazer um swipe para ir para a próxima."),
+        "tutorial_multiselect": MessageLookupByLibrary.simpleMessage(
+            "Você pode \"tocar e segurar\" em suas fotos para marcar várias fotos de uma vez."),
+        "tutorial_secret": MessageLookupByLibrary.simpleMessage(
+            "Oculte suas fotos privadas com proteção por senha, mantendo-as seguras."),
+        "tuxedo_tag": MessageLookupByLibrary.simpleMessage("Smoking"),
+        "twig_tag": MessageLookupByLibrary.simpleMessage("Galho"),
+        "umbrella_tag": MessageLookupByLibrary.simpleMessage("Guarda-chuva"),
+        "underwater_tag": MessageLookupByLibrary.simpleMessage("Subaquático"),
+        "unicycle_tag": MessageLookupByLibrary.simpleMessage("Monociclo"),
+        "unlimited_private_pics":
+            MessageLookupByLibrary.simpleMessage("Fotos privadas ilimitadas"),
+        "vacation_tag": MessageLookupByLibrary.simpleMessage("Férias"),
+        "van_tag": MessageLookupByLibrary.simpleMessage("Van"),
+        "vegetable_tag": MessageLookupByLibrary.simpleMessage("Vegetal"),
+        "vehicle_tag": MessageLookupByLibrary.simpleMessage("Veículo"),
+        "veil_tag": MessageLookupByLibrary.simpleMessage("Véu"),
+        "view_hidden_photos": MessageLookupByLibrary.simpleMessage(
+            "Para ver suas fotos ocultas, libere o bloqueio nas configurações do aplicativo. Você pode desbloqueá-las com seu PIN."),
+        "volcano_tag": MessageLookupByLibrary.simpleMessage("Vulcão"),
+        "wakeboarding_tag":
+            MessageLookupByLibrary.simpleMessage("Wakeboarding"),
+        "wall_tag": MessageLookupByLibrary.simpleMessage("Parede"),
+        "waterfall_tag": MessageLookupByLibrary.simpleMessage("Cachoeira"),
+        "waterfowl_tag": MessageLookupByLibrary.simpleMessage("Aves aquáticas"),
+        "waterskiing_tag": MessageLookupByLibrary.simpleMessage("Waterskiing"),
+        "web_page_tag": MessageLookupByLibrary.simpleMessage("Site"),
+        "welcome": MessageLookupByLibrary.simpleMessage("Bem-vindo!"),
+        "wetsuit_tag":
+            MessageLookupByLibrary.simpleMessage("Roupa de mergulho"),
+        "wheel_tag": MessageLookupByLibrary.simpleMessage("Roda"),
+        "wheelbarrow_tag":
+            MessageLookupByLibrary.simpleMessage("Carrinho de mão"),
+        "whiteboard_tag": MessageLookupByLibrary.simpleMessage("Quadro branco"),
+        "windshield_tag": MessageLookupByLibrary.simpleMessage("Parabrisa"),
+        "windsurfing_tag": MessageLookupByLibrary.simpleMessage("Windsurf"),
+        "wine_tag": MessageLookupByLibrary.simpleMessage("Vinho"),
+        "wing_tag": MessageLookupByLibrary.simpleMessage("Asa"),
+        "wool_tag": MessageLookupByLibrary.simpleMessage("Lã"),
+        "work_tag": MessageLookupByLibrary.simpleMessage("Trabalho"),
+        "wreath_tag": MessageLookupByLibrary.simpleMessage("Coroa"),
+        "x_minutes": MessageLookupByLibrary.simpleMessage("20 min"),
+        "year": MessageLookupByLibrary.simpleMessage("ano"),
+        "yes": MessageLookupByLibrary.simpleMessage("Sim"),
+        "you_are_premium":
+            MessageLookupByLibrary.simpleMessage("Você é premium!"),
+        "your_secret_key":
+            MessageLookupByLibrary.simpleMessage("Sua chave secreta")
+      };
 }
